% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{add_file_header}
\alias{add_file_header}
\title{Add header comments to the input files}
\usage{
add_file_header(filelist, con)
}
\arguments{
\item{filelist}{An object created by one of the r4ss::SS_read* functions.}

\item{con}{File to write to (passed to \code{con} input to \code{writeLines()})}
}
\description{
Lines starting with #C at the top of the file are carried over to the
*.ss_new files by Stock Synthesis
This function modifies any existing header to add or replace lines
written by r4ss that state the write time of the file.
}
\author{
Yukio Takeuchi, Ian G. Taylor
}
