% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{add_legend}
\alias{add_legend}
\title{Add legend to plots}
\usage{
add_legend(legendlabels, cumulative = FALSE)
}
\arguments{
\item{legendlabels}{Optional vector of labels to include in legend.}

\item{cumulative}{default for legend location set to topleft}
}
\description{
ss3diags function to add legend to plots
}
