% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{calc_var_adjust}
\alias{calc_var_adjust}
\title{Calculate variance adjustments for discard or mean body weight data}
\usage{
calc_var_adjust(data, type = c("CV", "sd"))
}
\arguments{
\item{data}{Either the "discard" or "mnwgt" elements of the list
returned by \code{\link[=SS_output]{SS_output()}}. Other data types might work here but haven't
been tested.}

\item{type}{Either "CV" or "sd" specifying the type of control file variance
adjustment, where the SS3 options are \verb{2=add_to_discard_stddev`` and }3=add_to_bodywt_CV\verb{, so if }data\verb{is discard data, type should be "CV" and if}data` is mean body weight, type should be "sd".}
}
\value{
A table of input and estimated uncertainty values in units of both
CV and sd including the following:
\itemize{
\item \code{fleet} is the fleet number
\item \code{mean_out} is the mean of the expected values
\item \code{mean_in} is the mean of the observed values
\item \code{CV_in} is the mean input CV
\item \code{sd_in} is the mean input SD values (which may include variance
adjustments already)
\item \code{sd_out} is the SD of the observed relative to the expected values,
calculated as described above
\item \code{CV_out} is the CV of the observed relative to the expected, calculated
as described above
\item \code{added} is the value that could be added to any existing value in the
"Input variance adjustments factors" section of the control file.
\item \code{type} is the data type code used in "Input variance adjustments factors"
}
}
\description{
Function developed for U.S. west coast Sablefish assessment in 2019 to tune
discard data or mean body weight data which are common inputs for
U.S. west coast groundfish assessments but as of 2023 have not often
had any data weighting method applied to them.
}
\details{
The calculation is based on sd_out = sqrt(mean(Obs - Exp)^2)).
Added sd is calculated as sd_out - sd_in where sd_in is the mean of the
input standard deviations (possibly including existing variance adjustments).
When a CV adjustment is required, the sd_out is converted to CV_out by
dividing by the mean of the expected values and with the added CV calculated
as CV_out - CV_in.
}
\author{
Kelli F. Johnson
}
