% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputlist.R
\name{check_inputlist}
\alias{check_inputlist}
\title{Check input argument \code{inputlist}}
\usage{
check_inputlist(inputlist)
}
\arguments{
\item{inputlist}{List created by the \code{SS_read()} function with
elements "dat", "ctl", "start", "fore", and (optionally) "wtatage" and "par".}
}
\value{
Either TRUE if the input list is valid, or FALSE if not, with
a warning about which elements are missing.
}
\description{
Check the elements of the \code{inputlist} list used as an argument in
\code{SS_write()} function.
}
\seealso{
\code{\link[=SS_write]{SS_write()}}
}
\author{
Kelli F. Johnson, Ian G. Taylor
}
