% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check input argument \code{model}}
\usage{
check_model(model, mydir = getwd())
}
\arguments{
\item{model}{Name of the Stock Synthesis model file (which has the .exe for
on Windows) in \code{mydir} without the extension (if any), e.g.,
\code{"ss"} or \code{"ss_win"}.}

\item{mydir}{The directory where \code{model} is located.}
}
\value{
A cleaned \code{model} name based on the input argument.
}
\description{
Check that the executable name provided in \code{model},
an input argument to numerous \code{r4ss} functions,
does not contain the extension and is available.
}
\author{
Kelli F. Johnson
}
