% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{check_replist}
\alias{check_replist}
\title{Check if the replist input is something that was created by SS_output()}
\usage{
check_replist(replist)
}
\arguments{
\item{replist}{An object to test}
}
\description{
Check if the replist input is something that was created by SS_output()
}
