% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_recdev_plot.R
\name{cohort_contributions}
\alias{cohort_contributions}
\title{Plot relative cohort contributions to spawning output in ending year + 1}
\usage{
cohort_contributions(replist, min_contribution = 0.01)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{min_contribution}{Minimum relative contribution to include in the plot
(this impacts the range of years/cohorts shown).}
}
\value{
ggplot object
}
\description{
Plot shows recruitment deviations and relative cohort contributions
(the product of numbers, maturity, and fecundity) to the spawning output in
the final year + 1 of the model. Figure suggested by Andre Punt at a 2025
PFMC Groundfish Subcommittee meeting.
}
\examples{
\dontrun{
}
}
\author{
Ian G. Taylor
}
