% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_selectivity.R
\name{empirical_selectivity}
\alias{empirical_selectivity}
\alias{empirical_selectivity.default}
\alias{empirical_selectivity.SS_empirical_selectivity}
\alias{empirical_selectivity.SS_output}
\title{Empirical selectivity diagnostic}
\usage{
empirical_selectivity(object, ...)

\method{empirical_selectivity}{default}(object, thr = 1e-05, ...)

\method{empirical_selectivity}{SS_empirical_selectivity}(object, fleet = NULL, sex = 1, by = "length")

\method{empirical_selectivity}{SS_output}(object, fleet = NULL, sex = 1, by = "length")
}
\arguments{
\item{object}{Typically, the result of reading Stock Synthesis model outputs
with \code{SS_output}.}

\item{...}{Additional arguments for specific methods, see Details.}

\item{fleet}{Integer, fleet number to compute the empirical selectivity.}

\item{sex}{If the fleet targets a specific sex, defaults to 1.}

\item{by}{Either 'length' or 'age', defaults to 'length'.}
}
\description{
Function to compute the empirical selectivity diagnostic for SS
outputs and methods to extract them for a particular fleet and time.
}
\details{
This function is called on run time by \code{SS_output}, so the
empirical selectivities are already computed (for length and age) in
your 'rep' object. An additional call to the function can be used to extract
the empirical selectivities for a particular fleet or sex.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: General constructor

\item \code{SS_empirical_selectivity}: Extract from empirical_selectivity objects

\item \code{SS_output}: Extracts from SS_output object
}}

\examples{
\dontrun{

rep = BigEye_2020
es = empirical_selectivity(rep, fleet=2, by="length")
}
}
