% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{get_areacols}
\alias{get_areacols}
\title{Get default vector of colors for each area}
\usage{
get_areacols(areacols, nareas)
}
\arguments{
\item{areacols}{Optional vector of colors for each area if model has
multiple areas. NULL value will be replaced by a default set of areas.}

\item{nareas}{number of areas}
}
\description{
this was previously contained within SS_plots() and 4 of the SSplotXXX()
functions.
}
\author{
Ian G. Taylor
}
