% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.r
\name{get_comments}
\alias{get_comments}
\title{Collect comments lines starting from "#C"
in datfile, ctlfile, starter.ss, forecast.ss etc}
\usage{
get_comments(dat, defaultComments = NULL)
}
\arguments{
\item{dat}{vector of strings usually outputs of readLines(*)
* is filename of datfile, ctlfile etc}

\item{defaultComments}{vector of strings
default : NULL, to read whole comments
If this function finds lines containg one of elements
of defaultComments, those lines will be ignored
e.g. c("^#C file created using the SS_writectl function in the R package r4ss",
"^#C file write time:")
is given, comments generated by SS_writectl_3.30 will be ignored.}
}
\description{
This function is used internally by SS_readdat_3.30,
SS_readctl_3.30.
This will identify 1st numeric data in dat (vector of string)
Then this function collects lines starting "#C" from lines
above 1st numeric data.
}
\seealso{
\code{\link{SS_readdat}}, \code{\link{SS_readdat_3.30}},
\code{\link{SS_readctl}}, \code{\link{SS_readctl_3.30}}
}
\author{
Yukio Takeuchi
}
