% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{get_dat_new_name}
\alias{get_dat_new_name}
\title{Get the name of the data .ss_new file in a directory}
\usage{
get_dat_new_name(dir)
}
\arguments{
\item{dir}{A file path to the directory of interest.
When omitted, functions typically default to the current working
directory or the model directory stored in \code{replist}.}
}
\value{
A string with the name of the data .ss_new file. If not found, will
be NA. Both of strings are searched for using \code{dir(pattern = )} and
if both exist, then \code{data_echo.ss_new} is returned. If the \code{dir} input
points to github, then \code{dir()} doesn't work and \code{data_echo.ss_new} is
always returned.
}
\description{
In previous versions of Stock Synthesis,
the file new data file was named \code{data.ss_new}.
\verb{_echo} was added to the name when the file was parsed into three separate
files.
}
\seealso{
\link{get_par_name}
}
