% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss3_exe.r
\name{get_ss3_exe_no_version}
\alias{get_ss3_exe_no_version}
\title{Download latest Stock Synthesis (SS3) executable.}
\usage{
get_ss3_exe_no_version(dir = NULL)
}
\arguments{
\item{dir}{The directory that you would like the executable downloaded to.}
}
\value{
A string of the full file path to the downloaded executable
}
\description{
The \code{get_ss3_exe_no_version()} function gets the latest release version from
github using the {gh} package and uses that tag to download the executable
for the appropriate operating system to the directory \code{dir} (if dir = NULL,
then the executable is downloaded to the working directory).
}
\details{
Downloads the latest release SS3 executable according to operating system.
}
\examples{
\dontrun{
get_ss3_exe_no_version()
}
}
\author{
Elizabeth F. Gugliotti
}
