% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_comps.R
\name{get_tuning_table}
\alias{get_tuning_table}
\title{Get the tuning table}
\usage{
get_tuning_table(
  replist,
  fleets,
  option,
  digits = 6,
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{fleets}{A vector of fleet numbers}

\item{option}{Which type of tuning: 'none', 'Francis', 'MI', 'DM', or
'Francis_via_DM'}

\item{digits}{Number of digits to round numbers to}

\item{write}{Write suggested tunings to a file 'suggested_tunings.ss'}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Get the tuning table
}
