% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readctl_3.30.R
\name{get_tv_parlabs}
\alias{get_tv_parlabs}
\title{Get time varying parameter labels}
\usage{
get_tv_parlabs(full_parms, block_design)
}
\arguments{
\item{full_parms}{the dataframe with the full parameter lines in the control
file as read in by r4ss.}

\item{block_design}{The block design in the control file as read in by r4ss.}
}
\description{
function to add get the names of short time varying parameter lines
}
