% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fwf4.r
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Utility function to test if x is "numerically" integer wrt machine epsilon
taken from exaple section of help of is.integer}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{value to check if it is "integer"}

\item{tol}{tolerace}
}
\description{
Utility function to test if x is "numerically" integer wrt machine epsilon
taken from exaple section of help of is.integer
}
