% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{iterate_jitter}
\alias{iterate_jitter}
\title{Execute a single jittered model run}
\usage{
iterate_jitter(
  i,
  dir = getwd(),
  printlikes = TRUE,
  exe = "ss",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{i}{Index of the jitter iteration.}

\item{dir}{Directory of the base model to be jittered}

\item{printlikes}{}

\item{exe}{Executable name. Can be just the name of the executable
file if it is in the specified directory or in the user's PATH. Can
also include the absolute path or a path relative to the specified
directory. Needs to be a single character string, not a vector.
On Windows, \code{exe} can optionally have the \code{.exe} extension appended;
on Unix-based systems (i.e., Mac and Linux), no extension should be
included.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{...}{Additional arguments passed to \code{\link[=run]{run()}}}
}
\value{
Performs one jittered model run and returns its negative log-likelihood
}
\description{
Execute a single jittered model run
}
\author{
James T. Thorson, Kelli F. Johnson, Ian G. Taylor,
Kathryn L. Doering, Kiva L. Oken
}
