% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_multifig.R
\name{make_multifig}
\alias{make_multifig}
\title{Create multi-figure plots.}
\usage{
make_multifig(
  ptsx,
  ptsy,
  yr,
  linesx = 0,
  linesy = 0,
  ptsSD = 0,
  sampsize = 0,
  effN = 0,
  showsampsize = TRUE,
  showeffN = TRUE,
  sampsize_label = "N=",
  effN_label = "effN=",
  sampsizeround = 1,
  maxrows = 6,
  maxcols = 6,
  rows = 1,
  cols = 1,
  fixdims = TRUE,
  main = "",
  cex.main = 1,
  xlab = "",
  ylab = "",
  size = 1,
  cexZ1 = 1.5,
  bublegend = TRUE,
  maxsize = NULL,
  do.sqrt = TRUE,
  minnbubble = 8,
  allopen = TRUE,
  xbuffer = c(0.1, 0.1),
  ybuffer = c(0, 0.15),
  yupper = NULL,
  ymin0 = TRUE,
  xlas = 0,
  ylas = NULL,
  axis1 = NULL,
  axis2 = NULL,
  axis1labs = NULL,
  linepos = 1,
  type = "o",
  polygons = TRUE,
  bars = FALSE,
  barwidth = NULL,
  ptscex = 1,
  ptscol = 1,
  ptscol2 = 1,
  colvec = c(rgb(1, 0, 0, 0.7), rgb(0, 0, 1, 0.7), rgb(0.1, 0.1, 0.1, 0.7)),
  linescol = c(rgb(0, 0.8, 0, 0.7), rgb(1, 0, 0, 0.7), rgb(0, 0, 1, 0.7)),
  lty = 1,
  lwd = 2,
  pch = 1,
  nlegends = 3,
  legtext = list("yr", "sampsize", "effN"),
  legx = NULL,
  legy = NULL,
  legadjx = NULL,
  legadjy = NULL,
  legsize = c(1.2, 1),
  legfont = c(2, 1),
  venusmars = TRUE,
  sampsizeline = FALSE,
  effNline = FALSE,
  sampsizemean = NULL,
  effNmean = NULL,
  ipage = 0,
  scalebins = FALSE,
  sexvec = NULL,
  multifig_colpolygon = grey(c(0.6, 0.8, 0.7), alpha = 0.7),
  multifig_oma = NULL,
  ...
)
}
\arguments{
\item{ptsx}{vector of x values for points or bars}

\item{ptsy}{vector of y values for points or bars of same length as ptsx}

\item{yr}{vector of category values (years) of same length as ptsx}

\item{linesx}{optional vector of x values for lines}

\item{linesy}{optional vector of y values for lines}

\item{ptsSD}{optional vector of standard deviations used to plot error bars
on top of each point under the assumption of normally distributed error}

\item{sampsize}{optional sample size vector of same length as ptsx}

\item{effN}{optional effective sample size vector of same length as ptsx}

\item{showsampsize}{show sample size values on plot?}

\item{showeffN}{show effective sample size values on plot?}

\item{sampsize_label}{label on sampsize}

\item{effN_label}{label on effN}

\item{sampsizeround}{rounding level for sample size values}

\item{maxrows}{maximum (or fixed) number or rows of panels in the plot}

\item{maxcols}{maximum (or fixed) number or columns of panels in the plot}

\item{rows}{number or rows to return to as default for next plots to come or
for single plots}

\item{cols}{number or cols to return to as default for next plots to come or
for single plots}

\item{fixdims}{fix the dimensions at maxrows by maxcols or resize based on
number of elements in \code{yr} input.}

\item{main}{title of plot}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{size}{vector of bubbles sizes if making a bubble plot}

\item{cexZ1}{Character expansion (cex) for point associated with value of 1.}

\item{bublegend}{Add legend with example bubble sizes to bubble plots.}

\item{maxsize}{maximum size of bubbles}

\item{do.sqrt}{scale bubbles based on sqrt of size vector. see ?bubble3 for
more info.}

\item{minnbubble}{number of unique x values before adding buffer. see
?bubble3 for more info.}

\item{allopen}{should all bubbles be open? see ?bubble3 for more info.}

\item{xbuffer}{extra space around points on the left and right as fraction
of total width of plot}

\item{ybuffer}{extra space around points on the bottom and top as fraction
of total height of plot}

\item{yupper}{upper limit on ymax (applied before addition of ybuffer)}

\item{ymin0}{fix minimum y-value at 0?}

\item{xlas}{label style (las) input for x-axis. Default 0 has horizontal
labels, input 2 would provide vertical labels.}

\item{ylas}{label style (las) input for y-axis. Default NULL has horizontal
labels when all labels have fewer than 6 characters and vertical otherwise.
Input 0 would force vertical labels, and 1 would force horizontal.}

\item{axis1}{optional position of bottom axis values}

\item{axis2}{optional position of left size axis values}

\item{axis1labs}{optional vector of labels for axis1 (either NULL or needs to
match length of axis1)}

\item{linepos}{should lines be added on top of points (linepos=1) or behind
(linepos=2)? A value of linepos = 0 will result in no line.}

\item{type}{type of line/points used for observed values (see 'type' in
?plot for details) on top of a grey polygon. Default is "o" for overplotting
points on lines.}

\item{polygons}{should polygons be added to the
(turning off is required for sex-ratio plot)}

\item{bars}{should the ptsx/ptsy values be bars instead of points
(TRUE/FALSE) NOT CURRENTLY FUNCTIONAL}

\item{barwidth}{width of bars in barplot, default method chooses based on
quick and dirty formula also, current method of plot(...type='h') could be
replaced with better approach}

\item{ptscex}{character expansion factor for points (default=1)}

\item{ptscol}{color for points/bars}

\item{ptscol2}{color for negative value points in bubble plots}

\item{colvec}{Vector of length 3 with colors for females, males, unsexed fish}

\item{linescol}{color for lines}

\item{lty}{line type}

\item{lwd}{Line width for plot elements.}

\item{pch}{point character type}

\item{nlegends}{number of lines of text to add as legends in each plot}

\item{legtext}{text in legend, a list of length=nlegends. values may be any
of 1.  "yr", 2. "sampsize", 3. "effN", or a vector of length = ptsx.}

\item{legx}{vector of length=nlegends of x-values of legends (default is
first one on left, all after on right)}

\item{legy}{vector of length=nlegends of y-values of legends (default is top
for all plots)}

\item{legadjx}{left/right adjustment of legends around legx}

\item{legadjy}{left/right adjustment of legends around legy}

\item{legsize}{font size for legends. default=c(1.2,1.0) (larger for year
and normal for others)}

\item{legfont}{font type for legends, same as "font" under ?par}

\item{venusmars}{Label females and males with venus and mars symbols?}

\item{sampsizeline}{show line for input sample sizes on top of conditional
age-at-length plots (TRUE/FALSE/scalar, still in development)}

\item{effNline}{show line for effective sample sizes on top of conditional
age-at-length plots (TRUE/FALSE/scalar, still in development)}

\item{sampsizemean}{mean input sample size value (used when sampsizeline=TRUE)}

\item{effNmean}{mean effective sample size value (used when effNline=TRUE)}

\item{ipage}{which page of plots when covering more than will fit within
maxrows by maxcols.}

\item{scalebins}{Rescale expected and observed proportions by dividing by
bin width for models where bins have different widths? Caution!: May not
work correctly in all cases.}

\item{sexvec}{vector of sex codes if more than one present (otherwise NULL)}

\item{multifig_colpolygon}{vector of polygon fill colors of length 3
(for females, males, and unsexed fish). Can be input to SS_plots and will be
passed to this function via the ... argument.}

\item{multifig_oma}{vector of outer margins. Can be input to SS_plots and will be
passed to this function via the ... argument.}

\item{\dots}{additional arguments passed to \code{par}.}
}
\description{
Function created as an alternative to lattice package for multi-figure plots
of composition data and fits from Stock Synthesis output.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}},\code{\link[=SSplotComps]{SSplotComps()}}
}
\author{
Ian Taylor
}
