% Generated by roxygen2 (4.0.1): do not edit by hand
\name{movepars}
\alias{movepars}
\title{explore movement parameterizations}
\usage{
movepars(nareas = 4, accuage = 40, getpars = T, getrates = T)
}
\arguments{
\item{nareas}{Number of areas}

\item{accuage}{Accumulator age}

\item{getpars}{T/F switch to get chosen parameters as output}

\item{getrates}{T/F switch to get derived movement rates as output}
}
\description{
A function to visualize parameterization of movement in Stock Synthesis. It
creates a GUI interface for movement exploration.
}
\author{
Ian Taylor
}
\keyword{dplot}
\keyword{dynamic}
\keyword{hplot}

