% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_comps.R
\name{plot_tunings}
\alias{plot_tunings}
\title{Plot the comp tunings from tune_comps() by iteration}
\usage{
plot_tunings(weights, dir = NULL)
}
\arguments{
\item{weights}{The list of variance adjustment tables created by
\code{\link[=tune_comps]{tune_comps()}}. This can either by the list of both "tuning_table_list"
and "weights" returned by \code{\link[=tune_comps]{tune_comps()}} or just the "weights" list.}

\item{dir}{Directory in which to save the plot as a png file.
If \code{dir == NULL} then the plot will be created in the default
graphics device.}
}
\description{
Plot the comp tunings from tune_comps() by iteration
}
