% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCI.R
\name{plotCI}
\alias{plotCI}
\title{Plot points with confidence intervals.}
\usage{
plotCI(x, y = NULL, uiw, liw = uiw, ylo = NULL, yhi = NULL, ...,
  sfrac = 0.01, ymax = NULL, add = FALSE, col = "black")
}
\arguments{
\item{x}{The x coordinates of points in the plot}

\item{y}{The y coordinates of the points in the plot.}

\item{uiw}{The width of the upper portion of the confidence region.}

\item{liw}{The width of the lower portion of the confidence region.}

\item{ylo}{Lower limit of y range.}

\item{yhi}{Upper limit of y range.}

\item{\dots}{Additional inputs that will be passed to the function
\code{plot(x,y,ylim=ylim,...)}}

\item{sfrac}{Fraction of width of plot to be used for bar ends.}

\item{ymax}{Additional input for Upper limit of y range.}

\item{add}{Add points and intervals to existing plot? Default=FALSE.}

\item{col}{Color for the points and lines.}
}
\description{
Given a set of x and y values and upper and lower bounds, this function
plots the points with error bars. This was Written by Venables and modified
to add access to ylim and contents.
}
\author{
Bill Venables, Ian Stewart, Ian Taylor, John Wallace
}
