% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngfun.R
\name{pngfun}
\alias{pngfun}
\title{Open png device and return info on the file being created}
\usage{
pngfun(
  plotinfo,
  file,
  plotdir,
  pwidth,
  pheight,
  punits,
  res,
  ptsize,
  caption = NA,
  alt_text = NA
)
}
\arguments{
\item{plotinfo}{table of information about all plots}

\item{file}{filename to write to (including .png extension)}

\item{plotdir}{directory where plots will be written}

\item{pwidth}{Width of plots printed to files in units of \code{punits}.}

\item{pheight}{Height of plots printed to files in units of \code{punits}.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{caption}{caption for the image}

\item{alt_text}{alternative text for screen readers
(if left as NA then will be set based on the caption)}
}
\description{
this was previously contained within each of the SSplotXXX() functions.
it (1) translates the not-quite-matching specifications for the image to the
values needed by png(), then return (2) returns the plotinfo data.frame
(which exists within each function which calls this) after adding a row
with the filename and caption for each new plot
Note: this just opens the png device which needs to be closed via dev.off()
outside this function
}
\author{
Ian G. Taylor
}
