\name{r4ss-package}
\alias{r4ss}
\alias{r4ss-package}
\docType{package}
\title{r4ss: Current Releases on GitHub}
\description{
The actively maintained r4ss package is distributed from GitHub. See
\code{\link{r4ss_repository}} for the canonical repository URL.
}
\seealso{
\url{https://github.com/r4ss/r4ss}
}
\keyword{internal}