% Generated by roxygen2 (4.1.0): do not edit by hand
\name{read.admbFit}
\alias{read.admbFit}
\title{Read ADMB .par and .cor files.}
\usage{
read.admbFit(file)
}
\arguments{
\item{file}{Name of ADMB executable such that files to read will
have format file.par and file.cor.}
}
\value{
List of various things from these files.
}
\description{
This function will parse the .par and .cor files to provide
things like parameter estimates, standard deviations,
and correlations. Required for Jim Thorson's Laplace
Approximation but likely useful for other purposes.
}
\author{
James Thorson
}
\seealso{
\code{\link{getADMBHessian}}, \code{\link{NegLogInt_Fn}}
}

