% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfit.R
\name{selfit}
\alias{selfit}
\title{A function to visual parameterization of double normal and double logistic
selectivity in Stock Synthesis}
\usage{
selfit(minLength = 10, maxLength = 65, silent = FALSE, init = NULL)
}
\arguments{
\item{minLength}{Minimum size to show}

\item{maxLength}{Maximum size to show}

\item{silent}{T/F switch to return fit at the end}

\item{init}{Optional initial values for the parameters}
}
\description{
A GUI interface for exploring selectivity.
}
\examples{

\dontrun{
selfit()
}

}
\author{
Tommy Garrison
}
\seealso{
\code{\link{sel.line}}
}
\keyword{dplot}
\keyword{dynamic}
\keyword{hplot}

