% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfit_spline.R
\name{selfit_spline}
\alias{selfit_spline}
\title{visualize parameterization of cubic spline selectivity in SS}
\usage{
selfit_spline(n = 4, minBin = 10, maxBin = 65, knots = NULL,
  slopevec = c(0.01, -0.01), params = NULL, dir = getwd(),
  silent = FALSE)
}
\arguments{
\item{n}{Number of knots.}

\item{minBin}{Minimum length or age to show.}

\item{maxBin}{Maximum length or age to show.}

\item{knots}{Vector giving location of each knot.}

\item{slopevec}{Optional initial values parameters controlling slope at
first and last knot.}

\item{params}{Optional initial values for the parameters controlling
selectivity at each knot.}

\item{dir}{Directory in which the spline_selex executable is located
(default = working directory).}

\item{silent}{TRUE/FALSE switch to return fit at the end.}
}
\description{
A GUI interface for exploring spline selectivity.
}
\examples{

\dontrun{
selfit_spline()
}

}
\seealso{
\code{\link{selfit}}
}
\author{
Ian Taylor, Yukio Takeuchi
}
