% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selShapes.R
\name{selShapes}
\alias{selShapes}
\title{Launch a shiny app that displays various selectivity curves}
\usage{
selShapes()
}
\description{
This app is hosted at https://connect.fisheries.noaa.gov/shiny-selex-ss3/
}
\author{
Allan Hicks, Andrea Havron, Ian Taylor,

inspired by tcl/tk code written by Tommy Garrison
}
