% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pars.R
\name{signif_string}
\alias{signif_string}
\title{Convert a numeric vector to character with chosen significant digits}
\usage{
signif_string(x, digits = 3)
}
\arguments{
\item{x}{Vector of numeric values}

\item{digits}{Number of significant digits}
}
\description{
Convert a numeric vector to character with chosen significant digits
}
\details{
The \code{signif()} function applies the same number of decimal
to all values in a vector, but parameter tables often include very
different scales, like Wtlen_1_Fem_GP_1 = 2e-06 and
Wtlen_2_Fem_GP_1 = 3, so this applies \code{signif()} separately to each
value and then converts to a string.
}
\author{
Ian G. Taylor
}
