% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_add_age_comp.r
\name{sim_add_age_comp}
\alias{sim_add_age_comp}
\title{this function adds a new line for an age composition observation in the agecomp data structure for the SS DAT file}
\usage{
sim_add_age_comp(dat_struct = NULL, comp_year = -1, comp_seas = -1,
  comp_fleet = -1, comp_Nsamp = -1, comp_matrix = NULL)
}
\arguments{
\item{dat_struct}{- DAT structure to be edited}

\item{comp_year}{- year for the age comp observation}

\item{comp_seas}{- season for the age comp observation}

\item{comp_fleet}{- fleet for the age comp observation}

\item{comp_Nsamp}{- sample size for the age comp observation}

\item{comp_matrix}{- age comp observation matrix, by sex}
}
\value{
edited DAT structure
}
\description{
this function adds a new line for an age composition observation in the agecomp data structure for the SS DAT file
}

