% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_add_catch}
\alias{sim_add_catch}
\title{this function adds new lines for catch in the catch dataframe in the data structure
for the SS DAT file}
\usage{
sim_add_catch(dat_struct = NULL, catch_mat = NULL, catch_year = -1)
}
\arguments{
\item{dat_struct}{- DAT file data structure to be edited}

\item{catch_mat}{- catches for each season (rows) and fleet (cols) (matrix)}

\item{catch_year}{- year that the catches will be applied}
}
\value{
edited DAT structure
}
\description{
this function adds new lines for catch in the catch dataframe in the data structure
for the SS DAT file
}

