% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_add_CPUE.r
\name{sim_add_CPUE}
\alias{sim_add_CPUE}
\title{this function adds a new line for a CPUE/survey observation in the CPUE data structure for the SS DAT file}
\usage{
sim_add_CPUE(
  dat_struct = NULL,
  CPUE_year = -1,
  CPUE_seas = -1,
  CPUE_fleet = -1,
  CPUE_obs = -999,
  CPUE_std_err = 999
)
}
\arguments{
\item{dat_struct}{- DAT structure to be edited}

\item{CPUE_year}{- year for the CPUE observation}

\item{CPUE_seas}{- season for the CPUE observation}

\item{CPUE_fleet}{- fleet for the CPUE observation}

\item{CPUE_obs}{- CPUE observation}

\item{CPUE_std_err}{- standard error for the CPUE observation}
}
\value{
edited DAT structure
}
\description{
this function adds a new line for a CPUE/survey observation in the CPUE data structure for the SS DAT file
}
