% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_add_discard.r
\name{sim_add_discard}
\alias{sim_add_discard}
\title{this function adds a new line for an observation in the discard data structure for the SS DAT file}
\usage{
sim_add_discard(dat_struct = NULL, dis_year = -1, dis_seas = -1,
  dis_fleet = -1, dis_obs = -999, dis_std_err = 999)
}
\arguments{
\item{dat_struct}{- DAT structure to be edited}

\item{dis_year}{- year for the discard observation}

\item{dis_seas}{- season for the discard observation}

\item{dis_fleet}{- fleet for the discard observation}

\item{dis_obs}{- discard observation}

\item{dis_std_err}{- standard error for the discard observation}
}
\value{
edited DAT structure
}
\description{
this function adds a new line for an observation in the discard data structure for the SS DAT file
}

