% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_add_index_as_CPUE}
\alias{sim_add_index_as_CPUE}
\title{this function adds a new line for an index observation in the CPUE dataframe in the data structure for the SS DAT file}
\usage{
sim_add_index_as_CPUE(dat_struct = NULL, index_year = -1, seas = -1,
  index_fleet = -1, index_obs = -999, index_std_err = 999)
}
\arguments{
\item{dat_struct}{- DAT structure to be edited}

\item{index_year}{- year for the index observation}

\item{seas}{- season for the index observation}

\item{index_fleet}{- fleet for the index observation}

\item{index_obs}{- CPUE or index observation}

\item{index_std_err}{- standard error for the index observation}
}
\value{
edited DAT structure
}
\description{
this function adds a new line for an index observation in the CPUE dataframe in the data structure for the SS DAT file
}

