% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_add_length_comp.r
\name{sim_add_length_comp}
\alias{sim_add_length_comp}
\title{this function adds a new line for a length composition observation in the lencomp data structure for the SS DAT file}
\usage{
sim_add_length_comp(dat_struct = NULL, comp_year = -1, comp_seas = -1,
  comp_fleet = -1, comp_Nsamp = -1, comp_gender = -1, comp_part = -1,
  comp_matrix = NULL)
}
\arguments{
\item{dat_struct}{- DAT structure to be edited}

\item{comp_year}{- year for the length comp observation}

\item{comp_seas}{- season for the length comp observation}

\item{comp_fleet}{- fleet for the length comp observation}

\item{comp_Nsamp}{- sample size for the length comp observation}

\item{comp_gender}{- gender indicator for the length comp observation (0 combine M & F, 1 F only, 2 M only, 3 separate M & F)}

\item{comp_part}{- partition indicator for the length comp observation (0 combined, 1 discard, 2 retained)}

\item{comp_matrix}{- length comp observation matrix, by sex}
}
\value{
edited DAT structure
}
\description{
this function adds a new line for a length composition observation in the lencomp data structure for the SS DAT file
}

