% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_generate_age_comp_from_expected.r
\name{sim_generate_age_comp_from_expected}
\alias{sim_generate_age_comp_from_expected}
\title{this function generates an age composition observation given the expected values for the fleet in the data structure for the SS DAT file}
\usage{
sim_generate_age_comp_from_expected(exp_dat_struct = NULL,
  exp_rep_struct = NULL, gen_year = -1, gen_seas = -1, gen_fleet = -1,
  apply_error = FALSE)
}
\arguments{
\item{exp_dat_struct}{- DAT structure with the expected values}

\item{exp_rep_struct}{- structure returned by SS_output()}

\item{gen_year}{- year for the generated composition observation}

\item{gen_seas}{- season number for the generated composition observation}

\item{gen_fleet}{- fleet number for the generated composition observation}

\item{apply_error}{- apply error to the generated composition observation}
}
\value{
generated age composition matrix with sample size
}
\description{
this function generates an age composition observation given the expected values for the fleet in the data structure for the SS DAT file
}

