% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_generate_CPUE}
\alias{sim_generate_CPUE}
\title{this function generates a CPUE observation for the CPUE dataframe in the data structure for the SS DAT file}
\usage{
sim_generate_CPUE(dat_struct = NULL, rep_struct = NULL, CPUE_fleet = -1,
  CPUE_fleet_name = "", CPUE_year = -1, apply_err = FALSE)
}
\arguments{
\item{dat_struct}{- DAT structure}

\item{rep_struct}{- structure returned by SS_output()}

\item{CPUE_fleet}{- fleet number of the CPUE fleet}

\item{CPUE_fleet_name}{- fleet name of the CPUE fleet}

\item{CPUE_year}{- year for the generated CPUE observation}

\item{apply_error}{- apply error to the generated observation (NOT IMPLEMENTED)}
}
\value{
generated CPUE_obs and CPUE_std_err
}
\description{
this function generates a CPUE observation for the CPUE dataframe in the data structure for the SS DAT file
}

