% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_generate_length_comp.r
\name{sim_generate_length_comp}
\alias{sim_generate_length_comp}
\title{this function generates a length composition observation for the fleet in the data structure for the SS DAT file}
\usage{
sim_generate_length_comp(dat_struct = NULL, rep_struct = NULL,
  gen_year = -1, gen_seas = -1, gen_fleet = -1, apply_error = FALSE)
}
\arguments{
\item{dat_struct}{- DAT structure}

\item{rep_struct}{- structure returned by SS_output()}

\item{gen_year}{- year for the generated composition observation}

\item{gen_seas}{- season number}

\item{gen_fleet}{- fleet number}

\item{apply_error}{- apply error to the generated observation}
}
\value{
generated length composition matrix with sample size
}
\description{
this function generates a length composition observation for the fleet in the data structure for the SS DAT file
}

