% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_get_age_selex}
\alias{sim_get_age_selex}
\title{this function retrieves the most recent selectivity-at-age from the SS report file structure}
\usage{
sim_get_age_selex(rep_struct = NULL, fleet_num = -1)
}
\arguments{
\item{rep_struct}{- structure returned by SS_output()}

\item{fleet_num}{- fleet number}
}
\value{
structure with selectivity-at-age
}
\description{
this function retrieves the most recent selectivity-at-age from the SS report file structure
}

