% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_get_forecast_catch_by_fleet.R
\name{sim_get_forecast_catch_by_fleet}
\alias{sim_get_forecast_catch_by_fleet}
\title{this function gets the catch for the first forecast year from the Forecast-report.sso file}
\usage{
sim_get_forecast_catch_by_fleet(forecast_report_path, num_fleets = 0,
  num_seasons = 0, num_fc_years = 0)
}
\arguments{
\item{forecast_report_path}{- full path for Forecast-report.sso file}

\item{num_fleets}{- number of fishing fleets, specified in the DAT file}

\item{num_seasons}{- number of seasons, specified in the DAT file}

\item{num_fc_years}{- number of forecast years, specified in the forecast.ss file}
}
\value{
matrix of catch by fleet (columns) and seasons (rows)
}
\description{
this function gets the catch for the first forecast year from the Forecast-report.sso file
}

