% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_get_length_selex.r
\name{sim_get_length_selex}
\alias{sim_get_length_selex}
\title{this function retrieves the most recent selectivity-at-length from the SS report file structure}
\usage{
sim_get_length_selex(dat_struct = NULL, rep_struct = NULL, fleet_num = -1)
}
\arguments{
\item{dat_struct}{- structure for DAT file}

\item{rep_struct}{- structure returned by SS_output()}

\item{fleet_num}{- fleet number}
}
\value{
structure with selectivity-at-length
}
\description{
this function retrieves the most recent selectivity-at-length from the SS report file structure
}

