% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_get_N_at_length.r
\name{sim_get_N_at_length}
\alias{sim_get_N_at_length}
\title{this function retrieves the numbers-at-length for the year, season, and fleet}
\usage{
sim_get_N_at_length(dat_struct = NULL, rep_struct = NULL, ret_year = -1,
  ret_seas = -1, fleet_num = -1)
}
\arguments{
\item{dat_struct}{- DAT structure}

\item{rep_struct}{- structure returned by SS_output()}

\item{ret_year}{- year for the numbers-at-length}

\item{ret_seas}{- season for the numbers-at-length}

\item{fleet_num}{- fleet number}
}
\value{
structure with numbers-at-length
}
\description{
this function retrieves the numbers-at-length for the year, season, and fleet
}

