% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_map_bins_to_bins.r
\name{sim_map_bins_to_bins}
\alias{sim_map_bins_to_bins}
\title{this function maps values from population bins to values for data bins}
\usage{
sim_map_bins_to_bins(pop_vector = NULL, data_vector = NULL,
  comp_matrix = NULL)
}
\arguments{
\item{pop_vector}{- vector of bin values for input values}

\item{data_vector}{- vector of bin values for output values}

\item{comp_matrix}{- input values (matrix)}
}
\value{
output values (matrix)
}
\description{
this function maps values from population bins to values for data bins
}

