% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_map_pop_age_to_data_age.r
\name{sim_map_pop_age_to_data_age}
\alias{sim_map_pop_age_to_data_age}
\title{this function maps a population age bins age composition observation to a data age bins age composition observation}
\usage{
sim_map_pop_age_to_data_age(dat_struct = NULL, comp_matrix = NULL)
}
\arguments{
\item{dat_struct}{- DAT structure}

\item{comp_matrix}{- age composition structure to be edited}
}
\value{
edited age composition structure
}
\description{
this function maps a population age bins age composition observation to a data age bins age composition observation
}

