% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_map_pop_len_to_data_len.r
\name{sim_map_pop_len_to_data_len}
\alias{sim_map_pop_len_to_data_len}
\title{this function maps a population length bins length composition observation to a data length bins length composition observation}
\usage{
sim_map_pop_len_to_data_len(dat_struct = NULL, comp_matrix = NULL)
}
\arguments{
\item{dat_struct}{- DAT structure}

\item{comp_matrix}{- length composition structure to be edited}
}
\value{
edited length composition structure
}
\description{
this function maps a population length bins length composition observation to a data length bins length composition observation
}

