% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_set_endyr.r
\name{sim_set_endyr}
\alias{sim_set_endyr}
\title{this function sets the variable endyr in the data structure for the SS DAT file}
\usage{
sim_set_endyr(dat_struct = NULL, endyr = -1)
}
\arguments{
\item{dat_struct}{- DAT structure to be edited}

\item{endyr}{- value to replace endyr}
}
\value{
edited DAT structure
}
\description{
this function sets the variable endyr in the data structure for the SS DAT file
}
