% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_data_to_fims.R
\name{ss3_data_to_fims}
\alias{ss3_data_to_fims}
\title{Convert SS3 data into format required by FIMS}
\usage{
ss3_data_to_fims(
  ss3_dir = NULL,
  ss_new = TRUE,
  ss3_inputs = NULL,
  ss3_output = NULL,
  EWAA = NULL,
  fleets = NULL,
  maxage = NULL,
  lengths = NULL
)
}
\arguments{
\item{ss3_dir}{Directory containing the SS3 input and output model
files. Alternatively the SS3 input and output lists can be input
directly using \code{ss3_inputs} and \code{ss3_output}}

\item{ss_new}{Logical indicating whether to read in the .ss_new files
instead of the original input files. Beneficial for dealing with
negative years in the wtatage file which can't yet be processed by
this function.}

\item{ss3_inputs}{A list containing \code{dat} such as that created by
\code{r4ss::SS_read()}. When \code{EWAA} is \code{TRUE}, the list must also contain
\code{wtatage}. Only required if \code{ss3_dir} is not provided.}

\item{ss3_output}{A list created by \code{r4ss::SS_output()}. Only required for
age-to-length conversion if \code{ss3_dir} is not provided.}

\item{EWAA}{Logical indicating whether to use empirical weight-at-age
and age_to_length conversion tables from the SS3 model or to rely on
the FIMS growth model. If NULL then it will follow the setting in the
SS3 control file (\code{ctl[['EmpiricalWAA']]})}

\item{fleets}{Which fleets to include in the processed output.
Note that the only start year population weight-at-age is read from the
\code{wtatage} element (fleet = 0). NULL will default to including all fleets
from the SS3 model.}

\item{maxage}{Max age to include in the processed output.
NULL will default to using age 0 up to the maximum age data bin in
the SS3 model.}

\item{lengths}{Vector of lengths to include in the processed output.
Needs to be a subset of the length bins in the SS3 model.
NULL will default to using all length data bins from the SS3 model.}
}
\value{
A data frame that can be passed to \code{FIMS::FIMSFrame()}
}
\description{
Uses output from \code{r4ss::SS_read()} and does
filtering, simplifying, and reformatting.
}
\seealso{
\code{\link[=SS_write]{SS_write()}} can be used to write the input files using the list
}
\author{
Ian G. Taylor, Megumi Oshima, Kelli F. Johnson
}
