% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_add_parameter_line.R
\name{SS_add_parameter_line}
\alias{SS_add_parameter_line}
\title{Add parameter line}
\usage{
SS_add_parameter_line(
  par_df,
  row_to_copy = 1,
  row_before,
  newval_df = data.frame(rowname = "new_parameter")
)
}
\arguments{
\item{par_df}{Parameter data frame from the r4ss control list that you would like to add an additional parameter line to. The parameter
table can by short or long.}

\item{row_to_copy}{Row number or row name to copy from the existing parameter data frame. Matching is done by \code{grep()} so only a
uniquely identifiable string is necessary, not the full row name. This can be a string or an integer.}

\item{row_before}{Row number of row name you would like to place the new parameter line after. Matching is done by \code{grep()} so only a
uniquely identifiable string is necessary, not the full row name. This can be a string or an integer.}

\item{newval_df}{A data frame of values you would like to change from the row that was copied. The only required column is "rowname".
All other column names must match existing column names in \code{par_df}.}
}
\value{
A parameter data frame with the new parameter line inserted in the specified location.
}
\description{
Add an additional parameter line to an existing parameter data frame in a control list. The function will copy an existing line,
change the values in columns that the user specifies (e.g., PHASE, INIT), and place the new line below an existing line. Currently the new
parameter cannot be in the first row. The user must also name the new row. The name does not have to follow any naming conventions, but it
must be unique. Parameters can be added to either style of parameter data frame (short or long).
}
\examples{

inputs <- SS_read(
dir = system.file("extdata", "simple_small", package = "r4ss")
)
# Details for new row
newval_df <- data.frame(
  rowname = "P_3_FISHERY",
  INIT = 5,
  PHASE = 4,
  Block = 1,
  Block_Fxn = 1
)

size_selex_new <- SS_add_parameter_line(
  par_df = inputs[["ctl"]][["size_selex_parms"]],
  row_to_copy = 1,
  row_before = "P_2_FISH",
  newval_df = newval_df
)

size_selex_new

}
\author{
Kiva L. Oken
}
