% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_decision_table_stuff.R
\name{SS_decision_table_stuff}
\alias{SS_decision_table_stuff}
\title{Extract total catch, spawning output, and fraction unfished from forecast years}
\usage{
SS_decision_table_stuff(replist, yrs = 2025:2036, digits = c(0, 0, 3))
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{yrs}{Range of years from which to extract values}

\item{digits}{Vector of number of digits to round to in table for
\itemize{
\item 1 catch
\item 2 spawning output
\item 3 fraction unfished (column is called "depl")
}}
}
\value{
A tibble with columns for year, total catch (dead biomass),
spawning output, and fraction unfished.
}
\description{
Values of total catch, spawning output, and fraction unfished are extracted
from the forecast years of a time series table for inclusion in a decision table.
}
\seealso{
\code{\link[=SS_ForeCatch]{SS_ForeCatch()}}
}
\author{
Ian G. Taylor
}
