% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_makeHTMLdiagnostictable.R
\name{SS_makeHTMLdiagnostictable}
\alias{SS_makeHTMLdiagnostictable}
\title{Creates html tables that show diagnostic outputs, including status checks, gradients, and correlations.}
\usage{
SS_makeHTMLdiagnostictable(replist, gradmax = 0.001)
}
\arguments{
\item{replist}{List item representing stock assessment model output list (\code{SS_output})}

\item{gradmax}{the largest gradient value for estimated parameter; the default is 1E-3}
}
\value{
a three-element vector; the first element is the name of the html table file, the second is the table caption, and the third is the category of output type
}
\description{
Creates html tables that show diagnostic outputs, including status checks, gradients, and correlations.
}
\seealso{
\code{\link{SS_plots}}, \code{\link{SS_output}, \code{\link{SS_html}}}
}
\author{
Christine Stawitz
}
