% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_output.R
\name{SS_output}
\alias{SS_output}
\title{A function to create a list object for the output from Stock Synthesis}
\usage{
SS_output(
  dir = getwd(),
  dir.mcmc = NULL,
  repfile = "Report.sso",
  compfile = "CompReport.sso",
  covarfile = "covar.sso",
  forefile = "Forecast-report.sso",
  wtfile = "wtatage.ss_new",
  warnfile = "warning.sso",
  ncols = lifecycle::deprecated(),
  forecast = TRUE,
  warn = TRUE,
  covar = TRUE,
  readwt = TRUE,
  verbose = TRUE,
  printstats = FALSE,
  hidewarn = FALSE,
  NoCompOK = TRUE,
  aalmaxbinrange = 4,
  SpawnOutputLabel = "Spawning output"
)
}
\arguments{
\item{dir}{A file path to the directory of interest.
When omitted, functions typically default to the current working
directory or the model directory stored in \code{replist}.}

\item{dir.mcmc}{Optional directory containing MCMC output. This can either be
relative to \code{dir}, such that \code{file.path(dir, dir.mcmc)}
will end up in the right place, or an absolute path.}

\item{repfile}{Name of the big report file (could be renamed by user).}

\item{compfile}{Name of the composition report file.}

\item{covarfile}{Name of the covariance output file.}

\item{forefile}{Name of the forecast file.}

\item{wtfile}{Name of the file containing weight at age data.}

\item{warnfile}{Name of the file containing warnings.}

\item{ncols}{Deprecated. This value is now calculated automatically.}

\item{forecast}{Read the forecast-report file?}

\item{warn}{Read the Warning.sso file?}

\item{covar}{Read covar.sso?}

\item{readwt}{Read the weight-at-age file?}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{printstats}{Print summary statistics about the output to the R GUI?
Defaults to FALSE.}

\item{hidewarn}{Hides some warnings output from the R GUI.}

\item{NoCompOK}{Allow the function to work without a CompReport file.}

\item{aalmaxbinrange}{The largest length bin range allowed for composition
data to be considered as conditional age-at-length data.}

\item{SpawnOutputLabel}{An alternative to "Spawning output" for use in
figure axis labels and table headers for models that include a fecundity
relationship. This provides an option to provide the units, e.g.
\code{SpawnOutputLabel = "Spawning output (trillions of eggs)"}.
This needs to be a user input because the units depend on the choice of
fecundity parameters which are calculated outside of the SS3 model.}
}
\value{
Many values are returned. Complete list would be quite long, but
should probably be created at some point in the future.
}
\description{
Reads the Report.sso and (optionally) the covar.sso, CompReport.sso and
other files produced by Stock Synthesis and formats the important
content of these files into a list in the R workspace. A few statistics
unavailable elsewhere are taken from the .par file. Summary
information and statistics can be returned to the R console or just
contained within the list produced by this function.
}
\examples{
\dontrun{
# read model output
myreplist <- SS_output(dir = system.file("extdata", "simple_small", package = "r4ss"))
# make a bunch of plots
SS_plots(myreplist)

# read model output and also read MCMC results (if run), which in
# this case would be stored in a subdirectory named "mcmc"
myreplist <- SS_output(
  dir = system.file("extdata", "simple_small_mcmc", package = "r4ss"),
  dir.mcmc = "mcmc"
)
}

}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}
}
\author{
Ian Stewart, Ian Taylor
}
