% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_plots.R
\name{SS_plots}
\alias{SS_plots}
\title{plot many quantities related to output from Stock Synthesis}
\usage{
SS_plots(
  replist = NULL,
  plot = 1:26,
  pdf = FALSE,
  png = TRUE,
  html = png,
  printfolder = "plots",
  dir = replist[["inputs"]][["dir"]],
  fleets = "all",
  areas = "all",
  fleetnames = replist[["FleetNames"]],
  fleetcols = NULL,
  fleetlty = 1,
  fleetpch = 1,
  lwd = 1,
  areacols = NULL,
  areanames = NULL,
  verbose = TRUE,
  uncertainty = TRUE,
  forecastplot = FALSE,
  datplot = TRUE,
  Natageplot = TRUE,
  samplesizeplots = TRUE,
  compresidplots = TRUE,
  comp.yupper = 0.4,
  sprtarg = replist[["sprtarg"]],
  btarg = replist[["btarg"]],
  minbthresh = replist[["minbthresh"]],
  pntscalar = NULL,
  bub.scale.pearson = 1.5,
  bub.scale.dat = 3,
  pntscalar.nums = 2.6,
  pntscalar.tags = 2.6,
  minnbubble = 8,
  aalyear = -1,
  aalbin = -1,
  aalresids = TRUE,
  maxneff = 5000,
  cohortlines = c(),
  smooth = TRUE,
  showsampsize = TRUE,
  showeffN = TRUE,
  sampsizeline = FALSE,
  effNline = FALSE,
  showlegend = TRUE,
  pwidth = 6.5,
  pheight = 4,
  pheight_tall = 6.5,
  punits = "in",
  ptsize = 10,
  res = 300,
  mainTitle = FALSE,
  cex.main = 1,
  selexlines = 1:6,
  rows = 1,
  cols = 1,
  maxrows = 6,
  maxcols = 4,
  maxrows2 = 4,
  maxcols2 = 4,
  andrerows = 4,
  tagrows = 3,
  tagcols = 3,
  parrows = 4,
  parcols = 2,
  fixdims = TRUE,
  new = TRUE,
  SSplotDatMargin = 8,
  filenotes = NULL,
  catchasnumbers = NULL,
  catchbars = TRUE,
  legendloc = "topleft",
  minyr = -Inf,
  maxyr = Inf,
  sexes = "all",
  scalebins = FALSE,
  scalebubbles = FALSE,
  tslabels = NULL,
  catlabels = NULL,
  maxsize = 1,
  showmle = TRUE,
  showpost = TRUE,
  showprior = TRUE,
  showinit = TRUE,
  showdev = FALSE,
  fitrange = FALSE,
  ...
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{plot}{Plot sets to be created, see list of plots below.  Use to
specify only those plot sets of interest, e.g., c(1,2,5,10). Plots for data
not available in the model run will automatically be skipped, whether called
or not.
Current grouping of plots is as follows:
\enumerate{
\item Biology
\item Selectivity and retention
\item Timeseries
\item Recruitment deviations
\item Recruitment bias adjustment
\item Spawner-recruit
\item Catch
\item SPR
\item Discards
\item Mean weight
\item Indices
\item Numbers at age
\item Length comp data (and generalized size comp data)
\item Age comp data
\item Conditional age-at-length data
\item Length comp fits (and generalized size comp fits)
\item Age comp fits
\item Conditional age-at-length fits
\item Francis and Punt conditional age-at-length comp fits
\item Mean length-at-age and mean weight-at-age
\item Tags
\item Yield
\item Movement
\item Data range
\item Parameter distributions
\item Diagnostic tables
}}

\item{pdf}{Send plots to PDF file instead of R GUI?}

\item{png}{Send plots to PNG files instead of R GUI?}

\item{html}{Run \code{\link[=SS_html]{SS_html()}} on completion? By default has same
value as \code{png}.}

\item{printfolder}{The sub-directory under 'dir' (see below) in which the
PNG files will be located.  The default sub-directory is "plots".
The directory will be created if it doesn't exist.
If 'printfolder' is set to "", it is ignored and the PNG files will be located
in the directory specified by 'dir'.}

\item{dir}{The directory in which a PDF file (if requested) will be created
and within which the printfolder sub-directory (see above) will be created
if png=TRUE. By default it will be the same directory that the report file
was read from by the \code{SS_output} function. Alternatives to the default
can be either relative (to the working directory) or absolute paths.
The function will attempt to create the directory it doesn't exist, but it
does not do so recursively.}

\item{fleets}{Either the string "all", or a vector of numerical values, like
c(1,3), listing fleets or surveys to be included in the plot.}

\item{areas}{Either the string "all", or a vector of numerical values, like
c(1,3), listing areas for which plots should be made in a multi-area model.
By default, plots will be made for all areas (excepting cases where the
function has not yet been updated for multi-area models). Default="all".}

\item{fleetnames}{Optional replacement for fleetnames used in data file.}

\item{fleetcols}{Optional vector of colors to use for each fleet. If NULL,
colors are generated automatically.}

\item{fleetlty}{Vector of line types used for each fleet in some plots.
Default=1.}

\item{fleetpch}{Vector of point types used for each fleet in some plots.
Default=1.}

\item{lwd}{Line width for plot elements.}

\item{areacols}{Optional vector of colors for each area if model has
multiple areas. NULL value will be replaced by a default set of areas.}

\item{areanames}{Optional vector of names for each area used in titles.
If NULL, names are generated automatically where needed.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{uncertainty}{Include values in plots showing estimates of uncertainty
(requires positive definite hessian in model?  Default=TRUE.}

\item{forecastplot}{Include forecast years in the timeseries plots and
plots of time-varying quantities?}

\item{datplot}{Plot the data by itself? This is useful in document
preparation, but doesn't change across alternative model runs with the same
data, so can be committed to save time once the plots have been created once.
Setting datplot=FALSE is equivalent to leaving off plots 15 and 16.
Default=TRUE.}

\item{Natageplot}{Plot the expected numbers at age bubble plots and mean-age
time series?  Default=TRUE.}

\item{samplesizeplots}{Show sample size plots?  Default=TRUE.}

\item{compresidplots}{Show residuals for composition plots?}

\item{comp.yupper}{Upper limit on ymax for polygon/histogram composition
plots. This avoids scaling all plots to have max=1 if there is a vector
with only a single observed fish in it. Default=0.4.}

\item{sprtarg}{Specify the F/SPR proxy target. Default=0.4.}

\item{btarg}{Target \%unfished to be used in plots showing \%unfished. May be
omitted by setting to NA.}

\item{minbthresh}{Threshold depletion to be used in plots showing depletion.
May be omitted by setting to NA.}

\item{pntscalar}{This scalar defines the maximum bubble size for bubble
plots. This option is still available but a better choice is to use
bub.scale.pearson and bub.scale.dat, which are allow the same scaling
throughout all plots.}

\item{bub.scale.pearson}{Character expansion (cex) value for a proportion of
1.0 in bubble plot of Pearson residuals. Default=1.5.}

\item{bub.scale.dat}{Character expansion (cex) value for a proportion of 1.0
in bubble plot of composition data. Default=3.}

\item{pntscalar.nums}{This scalar defines the maximum bubble size for
numbers-at-age and numbers-at-length plots.}

\item{pntscalar.tags}{This scalar defines the maximum bubble size for
tagging plots.}

\item{minnbubble}{This defines the minimum number of years below which blank
years will be added to bubble plots to avoid cropping.  Default=8.}

\item{aalyear}{Years to plot multi-panel conditional age-at-length fits for
all length bins; must be in a "c(YYYY,YYYY)" format. Useful for checking the
fit of a dominant year class, critical time period, etc. Default=-1.}

\item{aalbin}{The length bin for which multi-panel plots of the fit to
conditional age-at-length data will be produced for all years.  Useful to
see if growth curves are ok, or to see the information on year classes move
through the conditional data. Default=-1.}

\item{aalresids}{Plot the full set of conditional age-at-length Pearson
residuals? Turn to FALSE if plots are taking too long and you don't want
them.}

\item{maxneff}{The maximum value to include on plots of input and effective
sample size. Occasionally a calculation of effective N blows up to very
large numbers, rendering it impossible to observe the relationship for other
data. Default=5000.}

\item{cohortlines}{Optional vector of birth years for cohorts for which to
add growth curves to numbers at length bubble plots.  Default=c().}

\item{smooth}{Add loess smoother to observed vs. expected index plots and
input vs. effective sample size? Default=TRUE.}

\item{showsampsize}{Display sample sizes on composition plots?  Default=TRUE.}

\item{showeffN}{Display effective sample sizes on composition plots?
Default=TRUE.}

\item{sampsizeline}{Show line for input sample sizes on top of conditional
age-at-length plots (TRUE/FALSE, still in development).}

\item{effNline}{Show line for effective sample sizes on top of conditional
age-at-length plots (TRUE/FALSE, still in development).}

\item{showlegend}{Display legends in various plots?}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{pheight_tall}{Height of tall plots printed to png files in units of
\code{punits}, where the tall plots are a subset of the plots which typically
work best in a taller format.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{ptsize}{Point size for plotted text in plots printed to files. See
\code{help("png")} in R for details.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{mainTitle}{Logical indicating if a title should be included at the top
(not yet implemented for all plots).}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{selexlines}{Vector controlling which lines should be shown on
selectivity plots if the model includes retention. Default=1:5.}

\item{rows}{Number of rows to use for single panel plots. Default=1.}

\item{cols}{Number of columns to use for single panel plots. Default=1.}

\item{maxrows}{Maximum number of rows to for multi-panel plots.}

\item{maxcols}{Maximum number of columns for multi-panel plots.}

\item{maxrows2}{Maximum number of rows for conditional age-at-length
multi-panel plots.}

\item{maxcols2}{Maximum number of rows for conditional age-at-length
multi-panel plots.}

\item{andrerows}{Number of rows of Andre's conditional age-at-length plots
within each page.}

\item{tagrows}{Number of rows for tagging-related plots.}

\item{tagcols}{Number of columns for tagging-related plots.}

\item{parrows}{Number of rows for parameter distribution plots.}

\item{parcols}{Number of columns for parameter distribution plots.}

\item{fixdims}{Control whether multi-panel plots all have dimensions equal
to maxrows by maxcols, or resized within those limits to fit number of
plots. Default=TRUE.}

\item{new}{Open a new window or add to existing plot windows.  Default=TRUE.}

\item{SSplotDatMargin}{Size of right-hand margin in data plot (may be too
small if fleet names are long).}

\item{filenotes}{Optional vector of character strings to be added to intro
HTML page (if created) with notes about the model.}

\item{catchasnumbers}{Is catch input in numbers instead of biomass?
Default=FALSE.}

\item{catchbars}{Show catch by fleet as barplot instead of stacked polygons.
Default=TRUE.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a
vector of two numeric values representing the fraction of the maximum in
the x and y dimensions, respectively. See \code{help("legend")} for more info
on the string options.}

\item{minyr}{First year to show in time-series and time-varying plots.}

\item{maxyr}{Last year to show in time-series and time-varying plots. This
can either be an alternative to, or redundant with, the forecastplot input.}

\item{sexes}{Which sexes to show in composition plots. Default="all".}

\item{scalebins}{Rescale expected and observed proportions in composition
plots by dividing by bin width for models where bins have different widths?
Caution: May not work correctly in all cases.}

\item{scalebubbles}{Scale data-only bubbles by sample size, not just
proportion within sample? Default=FALSE.}

\item{tslabels}{Either NULL to have default labels for timeseries plots or
a vector of appropriate length with labels for each figure.}

\item{catlabels}{Either NULL to have default labels for catch plots or
a vector of appropriate length with labels for each figure.}

\item{maxsize}{The size of the largest bubble in the datasize
plot. Default is 1.0.}

\item{showmle}{Show MLE estimate and asymptotic variance estimate with blue
lines in the parameter distribution plots?}

\item{showpost}{Show posterior distribution as bar graph in parameter
distribution plots? Requires MCMC results to be available in \code{replist}.}

\item{showprior}{Show prior distribution as black line in the parameter
distribution plots?}

\item{showinit}{Show initial value as red triangle in the parameter
distribution plots?}

\item{showdev}{Include devs in the parameter distribution plots?}

\item{fitrange}{Fit range in parameter distribution plots tightly around MLE
and posterior distributions instead of full parameter range?}

\item{\dots}{Additional arguments that will be passed to some subfunctions.}
}
\description{
Creates a user-chosen set of plots, including biological quantities, time
series, and fits to data.  Plots are sent to R GUI, single PDF file, or
multiple PNG files. This is now just a wrapper which calls on separate
functions to make all the plots.
}
\details{
Functions to make groups of plots (e.g., \code{SSplotBiology()}) are
called by \code{SS_plots()}. These functions invisibly return a table with
details on the plots created. This table is  manipulated in \code{SS_plots()} and
written to a CSV file called \code{plotInfoTable_*.csv}, where * is replaced by
the date that the plots were created. One column of info is \code{category},
which includes a category for the plot (e.g., biology plots will all be
labeled as the \code{bio} category.)

If \code{html = TRUE}, \code{SS_html()} is called at the end of \code{SS_plots()} to create
the html output. A new tab in the html output for each unique value in the
\code{category} column of \code{plotInfoTable_*.csv}.

A new html file for each category is created with a name set by
\code{paste("_SS_output_", category, ".html", sep = "")} which includes
information on all the plots or tables associated with that category in the
order they appear in the CSV files. The tabbed output at the top of each
html file includes references to all the other files. If there are
duplicated file names in the CSV files (e.g., from multiple calls to
\code{SS_plots()} with different inputs), the one with the more recent \code{png_time}
is used. To view the html output, open \verb{_SS_output.html} in a browser.

Note that \code{SS_plots()} will avoid overwriting old plots created from running
\code{SS_plots()} by renaming old plotting output with the date appended to the
end of the file name (e.g., if \code{plots} is the \code{plotdir}, then it would be
renamed to something like \code{plots_StartTime_Tue_Jun_22_17.34.28_2021}). The
new output will be added to the specified \code{plotdir}.

Users can add their own figures to the html view. This can be done by either
multiple calls to \code{SS_plots()} with different arguments, or generating plots
outside of r4ss, adding details about these plots to a new
\code{plotInfoTable_*.csv} file, and calling \code{SS_html()} directly. Creating a
plotInfoTable_*.csv filename that is alphabetically earlier than the ones
created by \code{r4ss} should put the extra plots above the other r4ss plots
within that category. Adding any value for \code{category} other than the
existing defaults creates a new tab.

Multiple calls to \code{SS_plots()} with different arguments is possible, but if
adding more panels to a given plot, file may be generated that are no longer
needed, e.g., a lingering "page3" file when all the files now fit on pages 1
and 2.
}
\references{
Walters, Hilborn, and Christensen, 2008, Surplus production
dynamics in declining and recovering fish populations. Can. J. Fish. Aquat.
Sci. 65: 2536-2551.
}
\seealso{
\code{\link[=SS_output]{SS_output()}}, \code{\link[=SSplotBiology]{SSplotBiology()}},
\code{\link[=SSplotCatch]{SSplotCatch()}}, \code{\link[=SSplotComps]{SSplotComps()}},
\code{\link[=SSplotDiscard]{SSplotDiscard()}}, \code{\link[=SSplotIndices]{SSplotIndices()}},
\code{\link[=SSplotMnwt]{SSplotMnwt()}}, \code{\link[=SSplotNumbers]{SSplotNumbers()}},
\code{\link[=SSplotRecdevs]{SSplotRecdevs()}}, \code{\link[=SSplotSelex]{SSplotSelex()}},
\code{\link[=SSplotSpawnrecruit]{SSplotSpawnrecruit()}}, \code{\link[=SSplotSPR]{SSplotSPR()}},
\code{\link[=SSplotTags]{SSplotTags()}}, \code{\link[=SSplotTimeseries]{SSplotTimeseries()}},
\code{\link[=SSplotYield]{SSplotYield()}}
}
\author{
Ian Stewart, Ian Taylor
}
