% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{SS_profile}
\alias{SS_profile}
\title{Deprecated function to run a likelihood profile, renamed to profile().}
\usage{
SS_profile(...)
}
\arguments{
\item{...}{Any arguments associated with the now-deprecated functions.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
SS_profile() has been renamed as \code{\link[=profile]{profile()}}. See
https://github.com/r4ss/r4ss/issues/723 for more details.
}
\seealso{
\code{\link[=profile]{profile()}}
}
\author{
Ian G. Taylor
}
