% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_read.R
\name{SS_read}
\alias{SS_read}
\title{Read all Stock Synthesis input files for a model}
\usage{
SS_read(
  dir = getwd(),
  ss_new = FALSE,
  read_wtatage = FALSE,
  read_par = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dir}{A file path to the directory of interest or a raw github URL (see
example). The default is the current working directory, \code{dir = getwd()}.}

\item{ss_new}{A logical that controls if the \code{.ss_new} files or
the original input files are read in.
The default is to read the original files.}

\item{read_wtatage}{A logical that controls if the weight-at-age file is read in
for models that use parametric growth. For such models, it will read the
wtatage.ss_new file regardless of the setting of \code{ss_new} = TRUE/FALSE
because wtatage.ss typically doesn't exist for these models.}

\item{read_par}{A logical that controls if the .par file is read in for models that
don't require a .par file. The default is FALSE.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
An invisible list is returned.
The first element (\code{dir}) is the directory that was provided in the
argument \code{dir}.
The second element (\code{path}) is the result of \code{normalizePath(dir)},
which gives the full path.
The remaining four to six elements are list objects from reading in
the following input files:
\itemize{
\item data
\item control
\item starter
\item forecast
\item wtatage (may be absent if not required by the model and
\code{read_wtatage = FALSE})
\item par (will be NULL if not required by model or if control and par
do not match)
}
}
\description{
Read all the input files for a Stock Synthesis model into R as a list object.
These files will be in a single directory on your machine, i.e., \code{dir}.
Functionality comes from the \verb{r4ss::SS_read*()} functions.
This function simplifies the number of lines of code you need to write by
using all of the read functions to read in the
starter, control, data, and forecast files and if requested, the
weight-at-age file. The starter file is helpful because it provides names for
the control and data files.
}
\examples{
# Read in the 'simple' example model stored in {r4ss}
inputs <- SS_read(
  dir = system.file("extdata", "simple_small", package = "r4ss")
)
# Read in an example from GitHub stored in ss3-user-examples,
# wrapped in `dontrun` because it requires an Internet connection
\dontrun{
webexample <- SS_read(dir = file.path(
  "https://raw.githubusercontent.com",
  "nmfs-ost",
  "ss3-user-examples",
  "main",
  "model_files",
  "simple_long_wtatage"
))
}
}
\seealso{
\code{\link[=SS_write]{SS_write()}} can be used to write the input files using the list
created by this function.

Other read/write functions:
\code{\link[=SS_readctl]{SS_readctl()}},
\code{\link[=SS_readdat]{SS_readdat()}},
\code{\link[=SS_readforecast]{SS_readforecast()}},
\code{\link[=SS_readstarter]{SS_readstarter()}},
\code{\link[=SS_write]{SS_write()}},
\code{\link[=SS_writectl]{SS_writectl()}},
\code{\link[=SS_writedat]{SS_writedat()}},
\code{\link[=SS_writeforecast]{SS_writeforecast()}},
\code{\link[=SS_writestarter]{SS_writestarter()}}
}
\author{
Ian G. Taylor, Kelli F. Johnson
}
\concept{read/write functions}
