% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_read_summary.R
\name{SS_read_summary}
\alias{SS_read_summary}
\title{read ss_summary file}
\usage{
SS_read_summary(file = "ss_summary.sso", verbose = FALSE)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{verbose}{A logical value specifying if output should be printed
to the screen. The default is \code{verbose = TRUE}.}
}
\value{
Output will be a list with four elements, \code{header},
\code{likelihoods}, \code{parameters}, and \code{derived_quants}.
Each is a data frame with rownames indicating the quantity shown in each row.
}
\description{
read Stock Synthesis ss_summary.sso file into list object in R
}
\examples{

  \dontrun{
    summary <- SS_read_summary(file='c:/mymodel/ss_summary.sso')
  }

}
\seealso{
\code{\link{SS_output}}, \code{\link{SS_readforecast}},
\code{\link{SS_readdat}}, \code{\link{SS_readstarter}}
}
\author{
Ian Taylor
}
