% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readdat.R
\name{SS_readdat}
\alias{SS_readdat}
\title{read data file from SS}
\usage{
SS_readdat(version = "3.24", ...)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{verbose}{Should there be verbose output while running the file?
Default=TRUE.}

\item{echoall}{Debugging tool (not fully implemented) of echoing blocks of
data as it is being read.}

\item{section}{Which data set to read. Only applies for a data.ss_new file
created by Stock Synthesis. Allows the choice of either expected values
(section=2) or bootstrap data (section=3+). Leaving default of section=NULL
will read input data, (equivalent to section=1).}
}
\description{
Read Stock Synthesis data file into list object in R. This function is a
wrapper which calls either SS_readdat_3.24 or SS_readdat_3.30
(and potentially additional functions in the future). This setup allows those
functions to be cleaner (if somewhat redundant) than a single function that
attempts to do everything.
}
\author{
Ian G. Taylor
}
\seealso{
\code{\link{SS_readdat_3.24}}, \code{\link{SS_readdat_3.30}}
\code{\link{SS_readstarter}}, \code{\link{SS_readforecast}},
\code{\link{SS_readctl}}, \code{\link{SS_writestarter}},
\code{\link{SS_writeforecast}}, \code{\link{SS_writedat}},
\code{\link{SS_writectl}}
}
\keyword{data}

