% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readdat_2.00.R
\name{SS_readdat_2.00}
\alias{SS_readdat_2.00}
\title{read data file from SS version 2.00}
\usage{
SS_readdat_2.00(
  file,
  verbose = TRUE,
  echoall = lifecycle::deprecated(),
  section = NULL
)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.

See the formal arguments for a possible default filename.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{echoall}{Deprecated.}

\item{section}{Which data set to read. Only applies for a data.ss_new file
created by Stock Synthesis. Allows the choice of either expected values
(section=2) or bootstrap data (section=3+). Leaving default of section=NULL
will read input data, (equivalent to section=1). ## needs to be added}
}
\description{
Read Stock Synthesis (version 2.00) data file into list object in R.
This function was formerly called SS_readdat. That name is now used
for a wrapper function that calls either SS_readdat_2.00 SS_readdat_3.00
SS_readdat_3.24 or SS_readdat_3.30 (and potentially additional functions in the future).
}
\seealso{
\code{\link[=SS_readdat]{SS_readdat()}}, \code{\link[=SS_readdat_3.30]{SS_readdat_3.30()}}
\code{\link[=SS_readstarter]{SS_readstarter()}}, \code{\link[=SS_readforecast]{SS_readforecast()}},
\code{\link[=SS_writestarter]{SS_writestarter()}},
\code{\link[=SS_writeforecast]{SS_writeforecast()}}, \code{\link[=SS_writedat]{SS_writedat()}}
}
\author{
Ian G. Taylor, Yukio Takeuchi, Z. Teresa A'mar, Neil L. Klaer
}
