% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readdat_3.30.R
\name{SS_readdat_3.30}
\alias{SS_readdat_3.30}
\title{read data file from SS version 3.30 (function still in development)}
\usage{
SS_readdat_3.30(file, verbose = TRUE, echoall = FALSE, section = NULL)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{verbose}{Should there be verbose output while running the file?
Default=TRUE.}

\item{echoall}{Debugging tool (not fully implemented) of echoing blocks of
data as it is being read.}

\item{section}{Which data set to read. Only applies for a data.ss_new file
created by Stock Synthesis. Allows the choice of either expected values
(section=2) or bootstrap data (section=3+). Leaving default of section=NULL
will read input data, (equivalent to section=1).}
}
\description{
Read Stock Synthesis (version 3.30) data file into list object in R.
This function was formerly called SS_readdat. That name is now used
for a wrapper function that calls either SS_readdat_3.24 or SS_readdat_3.30
(and potentially additional functions in the future).
}
\seealso{
\code{\link{SS_readdat}}, \code{\link{SS_readdat_3.30}}
\code{\link{SS_readstarter}}, \code{\link{SS_readforecast}},
\code{\link{SS_writestarter}},
\code{\link{SS_writeforecast}}, \code{\link{SS_writedat}}
}
\author{
Ian G. Taylor, Yukio Takeuchi, Z. Teresa A'mar, Chris J. Grandin
}
