% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readpar_3.24.R
\name{SS_readpar_3.24}
\alias{SS_readpar_3.24}
\title{Deprecated: read ss.par file from SS version 3.24}
\usage{
SS_readpar_3.24(parfile, datsource, ctlsource, verbose = TRUE)
}
\arguments{
\item{parfile}{Filename either with full path or relative to working directory.}

\item{datsource}{list or character. If list, should be a list produced
from \code{\link[=SS_writedat]{SS_writedat()}}. If character, should be the full file location of an
SS data file.}

\item{ctlsource}{list or character. If list, should be a list produced
from \code{\link[=SS_writectl]{SS_writectl()}}. If character, should be the full file location of an
SS control file.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Read Stock Synthesis (version 3.24) parameter file into list object in R.
}
\details{
Support for 3.24 models within the r4ss \verb{SS_read*} and \verb{SS_write*()}
functions is ending, so please update models to 3.30.
}
\seealso{
\code{\link[=SS_readctl]{SS_readctl()}}, \code{\link[=SS_readdat]{SS_readdat()}}
\code{\link[=SS_readdat_3.24]{SS_readdat_3.24()}},\code{\link[=SS_readdat_3.24]{SS_readdat_3.24()}}
\code{\link[=SS_readctl_3.24]{SS_readctl_3.24()}},
\code{\link[=SS_readstarter]{SS_readstarter()}}, \code{\link[=SS_readforecast]{SS_readforecast()}},
\code{\link[=SS_writestarter]{SS_writestarter()}},
\code{\link[=SS_writeforecast]{SS_writeforecast()}}, \code{\link[=SS_writedat]{SS_writedat()}}
}
\author{
Nathan R. Vaughan
}
