% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{SS_RunJitter}
\alias{SS_RunJitter}
\title{Deprecated function to run jitters, renamed to jitter()}
\usage{
SS_RunJitter()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
SS_RunJitter() has been renamed as \code{\link[=jitter]{jitter()}}. See
https://github.com/r4ss/r4ss/issues/723 for more details.
}
\seealso{
\code{\link[=jitter]{jitter()}}
}
\author{
Ian G. Taylor
}
