% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_splitdat.R
\name{SS_splitdat}
\alias{SS_splitdat}
\title{Split apart bootstrap data to make input file.}
\usage{
SS_splitdat(
  inpath = "working_directory",
  outpath = "working_directory",
  inname = "data.ss_new",
  outpattern = "BootData",
  number = FALSE,
  verbose = TRUE,
  fillblank = TRUE,
  MLE = TRUE,
  inputs = FALSE,
  notes = ""
)
}
\arguments{
\item{inpath}{Directory containing the input file. By default the working
directory given by getwd() is used. Default="working_directory".}

\item{outpath}{Directory into which the output file will be written.
Default="working_directory".}

\item{inname}{File name of input data file to be split.
Default="Data.SS_New".}

\item{outpattern}{File name of output data file. Default="BootData".}

\item{number}{Append bootstrap number to the file name chosen in
\code{outpattern}? Default=F.}

\item{verbose}{Provide richer command line info of function progress?
Default=TRUE.}

\item{fillblank}{Replace blank lines with "#". Helps with running on linux.
Default=TRUE.}

\item{MLE}{Grab the maximum likelihood values from the second block in
Data.SS_New (instead of bootstrap values or copies of inputs)? Default=TRUE.}

\item{inputs}{Grab the copy of the input values values from the first block
in Data.SS_New (instead of MLE or bootstrap values)? Default=F.}

\item{notes}{Notes to the top of the new file (comment indicator "#C" will
be added). Default="".}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{SS_splitdat()} is being deprecated because it is no longer needed in as of
Stock Synthesis 3.30.19, where the files are already split. If needing to
split older version of SS3, this could be done by reading in the file to
\code{SS_readdat()}, specifying the section argument to grab the desired part of
the file, and then writing out the file using. `SS_writedat()'.
}
\author{
Ian Taylor
}
\keyword{internal}
