% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_write_length.fit.R
\name{SS_write_length.fit}
\alias{SS_write_length.fit}
\title{Write length.fit file to be used by the MFCL length-comp viewer.}
\usage{
SS_write_length.fit(replist = NULL, outfile = "length.fit",
  compfile = "CompReport.sso", dir = "default", overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{replist}{List created by \code{SS_output}}

\item{outfile}{Name of file to create.}

\item{compfile}{SS output file with composition data info.}

\item{dir}{Directory where stuff happens. Defaults to directory where model
was run.}

\item{overwrite}{Overwrite existing file?}

\item{verbose}{More verbose info on progress of the function?}
}
\description{
Writes files in the format used by the MFCL length-composition viewer.
Inspired by Simon Hoyle's demonstration. Still needs work.
}
\author{
Ian Taylor
}
\references{
\url{http://www.multifan-cl.org/},
\url{http://www.spc.int/OceanFish/en/ofpsection/sam/research/272-mfcl-viewer}
}
\keyword{data}
\keyword{manip}

