% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writectl.R
\name{SS_writectl}
\alias{SS_writectl}
\title{Write Stock Synthesis control file}
\usage{
SS_writectl(
  ctllist,
  outfile,
  version = "3.30",
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{ctllist}{List object created by \code{\link[=SS_readdat]{SS_readdat()}}.}

\item{outfile}{Filename for where to write new control file.}

\item{version}{SS version number. Currently "3.24" or "3.30" are supported,
either as character or numeric values (noting that numeric 3.30  = 3.3).
\code{version = NULL} is no longer the default or an allowed entry.
The default is \code{version = "3.30"}.}

\item{overwrite}{A logical value specifying if the existing file(s)
should be overwritten. The default value is \code{overwrite = FALSE}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Write Stock Synthesis control file from list object in R which was probably
created using \code{\link[=SS_readctl]{SS_readctl()}}. This function is a
wrapper which calls SS_writectl_3.30() (previously also SS_writectl_3.24,
but that function has been deprecated).
}
\seealso{
Other read/write functions: 
\code{\link{SS_read}()},
\code{\link{SS_readctl}()},
\code{\link{SS_readdat}()},
\code{\link{SS_readforecast}()},
\code{\link{SS_readstarter}()},
\code{\link{SS_write}()},
\code{\link{SS_writedat}()},
\code{\link{SS_writeforecast}()},
\code{\link{SS_writestarter}()}
}
\author{
Ian G. Taylor, Yukio Takeuchi, Gwladys I. Lambert,
Kathryn L. Doering, Nathan R. Vaughan
}
\concept{read/write functions}
