% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writectl_3.24.R
\name{SS_writectl_3.24}
\alias{SS_writectl_3.24}
\title{Deprecated: write 3.24 control file}
\usage{
SS_writectl_3.24(ctllist, outfile, overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{ctllist}{List object created by \code{\link[=SS_readctl]{SS_readctl()}}.}

\item{outfile}{Filename for where to write new data file.}

\item{overwrite}{A logical value specifying if the existing file(s)
should be overwritten. The default value is \code{overwrite = FALSE}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
write Stock Synthesis control file from list object in R which was probably
created using \code{\link[=SS_readctl]{SS_readctl()}}
}
\details{
Support for 3.24 models within the r4ss \verb{SS_read*} and \verb{SS_write*()}
functions is ending, so please update models to 3.30.
}
\seealso{
\code{\link[=SS_readctl]{SS_readctl()}}, \code{\link[=SS_readctl_3.24]{SS_readctl_3.24()}},\code{\link[=SS_readstarter]{SS_readstarter()}},
}
\author{
Yukio Takeuchi, Kathryn L. Doering, Nathan R. Vaughan
}
