% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writectl_3.30.R
\name{SS_writectl_3.30}
\alias{SS_writectl_3.30}
\title{write control file for SS version 3.30}
\usage{
SS_writectl_3.30(ctllist, outfile, overwrite = FALSE, verbose)
}
\arguments{
\item{ctllist}{List object created by \code{\link{SS_readctl}}.}

\item{outfile}{Filename for where to write new data file.}

\item{overwrite}{Should existing files be overwritten? Default=FALSE.}

\item{verbose}{Should there be verbose output while running the file?}
}
\description{
write Stock Synthesis control file from list object in R which was created
  using \code{\link{SS_readctl}}.This function is designed to be called 
  using \code{\link{SS_writectl}} and should not be called directly.
}
\seealso{
\code{\link{SS_readctl}}, \code{\link{SS_readctl_3.30}},\code{\link{SS_readstarter}},
\code{\link{SS_readforecast}},
\code{\link{SS_writestarter}}, \code{\link{SS_writeforecast}},
\code{\link{SS_writedat}}
}
\author{
Kathryn Doering, Yukio Takeuchi, Neil Klaer, Watal M. Iwasaki
}
