% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writedat_3.24.R
\name{SS_writedat_3.24}
\alias{SS_writedat_3.24}
\title{write data file for SS version 3.24}
\usage{
SS_writedat_3.24(
  datlist,
  outfile,
  overwrite = FALSE,
  faster = lifecycle::deprecated(),
  verbose = TRUE
)
}
\arguments{
\item{datlist}{List object created by \code{\link[=SS_readdat]{SS_readdat()}}.}

\item{outfile}{Filename for where to write new data file.}

\item{overwrite}{Should existing files be overwritten? Default=FALSE.}

\item{faster}{Deprecated. Speed up writing by writing length and age comps without aligning
the columns (by using write.table instead of print.data.frame)}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Write Stock Synthesis data file from list object in R which was probably
created using \code{\link[=SS_readdat]{SS_readdat()}} (which would have called on
\code{\link[=SS_readdat_3.24]{SS_readdat_3.24()}}).
}
\seealso{
\code{\link[=SS_writedat]{SS_writedat()}}, \code{\link[=SS_writedat_3.30]{SS_writedat_3.30()}},
\code{\link[=SS_readdat]{SS_readdat()}},
\code{\link[=SS_readstarter]{SS_readstarter()}}, \code{\link[=SS_writestarter]{SS_writestarter()}},
\code{\link[=SS_readforecast]{SS_readforecast()}}, \code{\link[=SS_writeforecast]{SS_writeforecast()}}
}
\author{
Ian G. Taylor, Yukio Takeuchi, Gwladys I. Lambert, Kelli F. Johnson,
Chantel R. Wetzel
}
