% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writestarter.R
\name{SS_writestarter}
\alias{SS_writestarter}
\title{write starter file}
\usage{
SS_writestarter(
  mylist,
  dir = getwd(),
  file = "starter.ss",
  overwrite = FALSE,
  verbose = TRUE,
  warn = lifecycle::deprecated()
)
}
\arguments{
\item{mylist}{List object created by \code{\link[=SS_readstarter]{SS_readstarter()}}.}

\item{dir}{A file path to the directory of interest.
When omitted, functions typically default to the current working
directory or the model directory stored in \code{replist}.}

\item{file}{Filename for new starter file. Default="starter.ss".}

\item{overwrite}{A logical value specifying if the existing file(s)
should be overwritten. The default value is \code{overwrite = FALSE}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{warn}{Deprecated.}
}
\description{
write Stock Synthesis starter file from list object in R which was probably
created using \code{\link[=SS_readstarter]{SS_readstarter()}}
}
\seealso{
Other read/write functions:
\code{\link[=SS_read]{SS_read()}},
\code{\link[=SS_readctl]{SS_readctl()}},
\code{\link[=SS_readdat]{SS_readdat()}},
\code{\link[=SS_readforecast]{SS_readforecast()}},
\code{\link[=SS_readstarter]{SS_readstarter()}},
\code{\link[=SS_write]{SS_write()}},
\code{\link[=SS_writectl]{SS_writectl()}},
\code{\link[=SS_writedat]{SS_writedat()}},
\code{\link[=SS_writeforecast]{SS_writeforecast()}}
}
\author{
Ian G. Taylor, Kelli F. Johnson, Kathryn R. Doering
}
\concept{read/write functions}
