% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writewtatage.R
\name{SS_writewtatage}
\alias{SS_writewtatage}
\title{Write weight-at-age file}
\usage{
SS_writewtatage(
  mylist,
  dir = getwd(),
  file = "wtatage.ss",
  overwrite = FALSE,
  verbose = TRUE,
  warn = lifecycle::deprecated()
)
}
\arguments{
\item{mylist}{Object created by \code{\link[=SS_readwtatage]{SS_readwtatage()}}.}

\item{dir}{A file path to the directory of interest.
When omitted, functions typically default to the current working
directory or the model directory stored in \code{replist}.}

\item{file}{Filename for new weight-at-age file, which
will be appended to \code{dir} to create a full file path.
Default="wtatage.ss".}

\item{overwrite}{A logical value specifying if the existing file(s)
should be overwritten. The default value is \code{overwrite = FALSE}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{warn}{Deprecated.}
}
\description{
Write Stock Synthesis weight-at-age file from R object that was probably
created using \code{\link[=SS_readwtatage]{SS_readwtatage()}}
}
\seealso{
\code{\link[=SS_readwtatage]{SS_readwtatage()}}
}
\author{
Kelli F. Johnson
}
