% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSbiologytables.R
\name{SSbiologytables}
\alias{SSbiologytables}
\title{Deprecated function to create a table of biology for assessment reporting:
length, weight, \% mature, fecundity, and selectivity}
\usage{
SSbiologytables(
  replist = NULL,
  printfolder = "tables",
  dir = replist[["inputs"]][["dir"]],
  fleetnames = replist[["FleetNames"]],
  selexyr = replist[["endyr"]]
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{printfolder}{The sub-directory under 'dir' (see below) in which the
PNG files will be located.  The default sub-directory is "plots".
The directory will be created if it doesn not exist.
If 'printfolder' is set to "", it is ignored and the PNG files will
be located in the directory specified by 'dir'.}

\item{dir}{A file path to the directory of interest.
When omitted, functions typically default to the current working
directory or the model directory stored in \code{replist}.}

\item{fleetnames}{Optional replacement for fleetnames used in data file.}

\item{selexyr}{The year to summarize selectivity, the default is the final
model year.}
}
\value{
A csv files containing biology and selectivity tables
}
\description{
Replaced by \code{\link[=table_biology]{table_biology()}}. Takes the object created by SS_output to create table for reporting
for West Coast groundfish.  Works with Stock Synthesis versions 3.30.12
and later.
}
\author{
Chantel Wetzel
}
