% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSbootstrap.R
\name{SSbootstrap}
\alias{SSbootstrap}
\title{Fit models to parametric bootstraps}
\usage{
SSbootstrap()
}
\description{
Run a series of models fit to parametric bootstrap data taken from
data.ss_new.  This is not yet a generalized function, just some example code
for how to do a parametric bootstrap such as was done for the Pacific hake
model in 2006.
}
\note{
Thanks to Nancie Cummings for inspiration.
}
\author{
Ian Taylor
}
