% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{SSdiagsTime2Year}
\alias{SSdiagsTime2Year}
\title{Convert Time-Steps}
\usage{
SSdiagsTime2Year(ss3out, time.steps = 0.25, end.time)
}
\arguments{
\item{ss3out}{outputs from \code{\link[=SS_output]{SS_output()}} or \code{\link[=SSsummarize]{SSsummarize()}}}

\item{time.steps}{time steps behind yrs e.g. 0.25 for quarterly}

\item{end.time}{last time step e.g. 2018.75 with a cpue observation}
}
\value{
Reformatted Rep file outputs
}
\description{
Function to convert non-annual into annual time-steps for retros and cpue residuals
}
\keyword{rep}
\keyword{retro}
\keyword{retrocomps}
\keyword{utils}
