% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSFishGraph.R
\name{SSFishGraph}
\alias{SSFishGraph}
\title{A function for converting Stock Synthesis output to the format used by
FishGraph}
\usage{
SSFishGraph(replist, title = "SSv3 output", species = "some kind of fish")
}
\arguments{
\item{replist}{Object created by SS_output}

\item{title}{Title of output}

\item{species}{Species name}
}
\description{
Only skeleton of a function right now, needs work. Intended as a translator
to convert the output from object created by \code{\link{SS_output}} to the
format used by FishGraph.
}
\author{
Ian Taylor
}
\references{
A website related to FishGraph is
\url{http://r-forge.r-project.org/projects/fishgraph/}
}
\keyword{data}
\keyword{manip}

