% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmakeMmatrix.R
\name{SSmakeMmatrix}
\alias{SSmakeMmatrix}
\title{Convert a matrix of natural mortality values into
inputs for Stock Synthesis}
\usage{
SSmakeMmatrix(mat, startyr, outfile = NULL, overwrite = FALSE,
  yrs.in.columns = TRUE)
}
\arguments{
\item{mat}{a matrix of natural mortality by year and age, starting with age 0}

\item{startyr}{the first year of the natural mortality values (no missing years)}

\item{outfile}{optional file to which the results will be written}

\item{overwrite}{if 'outfile' is provided and exists, option to overwrite or not}

\item{yrs.in.columns}{an indicator of whether the matrix has years in columns or rows}
}
\value{
Prints inputs with option to write to chosen file
}
\description{
Inspired by Valerio Bartolino and North Sea herring
}
\author{
Ian Taylor
}
