% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{sspar}
\alias{sspar}
\title{Allow Multi-Plots
Set the par() to options suitable for ss3diags multi plots.}
\usage{
sspar(
  mfrow = c(1, 1),
  plot.cex = 1,
  mai = c(0.55, 0.6, 0.1, 0.1),
  omi = c(0, 0, 0, 0) + 0.1,
  labs = TRUE
)
}
\arguments{
\item{mfrow}{determines plot frame set up}

\item{plot.cex}{cex graphic option}

\item{mai}{graphical par for plot margins}

\item{omi}{Outer margins in lines of text.}

\item{labs}{if TRUE margins are narrow}
}
\description{
See \link[graphics]{par} for more details on each parameter.
}
\keyword{ssplot}
\keyword{utils}
