% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotCohortCatch.R
\name{SSplotCohortCatch}
\alias{SSplotCohortCatch}
\title{Plot cumulative catch by cohort.}
\usage{
SSplotCohortCatch(
  replist,
  subplots = 1:2,
  add = FALSE,
  plot = TRUE,
  print = FALSE,
  cohortcols = NULL,
  cohortfrac = 1,
  cohortvec = NULL,
  cohortlabfrac = 0.1,
  cohortlabvec = NULL,
  lwd = 3,
  plotdir = replist[["inputs"]][["dir"]],
  xlab = "Year",
  labels = c("Age", "Cumulative catch by cohort (in numbers x1000)",
    "Cumulative catch by cohort (x1000 mt)"),
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{subplots}{Vector controlling which subplots to create}

\item{add}{Add to existing plot? (not yet implemented)}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{cohortcols}{Vector of colors to show for each cohort. Default is range
of colors shade indicating time period.}

\item{cohortfrac}{What fraction of the cohorts to include in plot. If value
< 1 is used, then cohorts are filtered to only include those with the
highest maximum cumulative catch. Value will be overridden by
\code{cohortvec}.}

\item{cohortvec}{Optional vector of birth years for cohorts to include in
plot. Value overrides \code{cohortfrac}.}

\item{cohortlabfrac}{What fraction of the cohorts to label in plot. By
default, top 10\% of cohorts are labeled. Value will be overridden by
\code{cohortlabvec}.}

\item{cohortlabvec}{Optional vector of birth years for cohorts to label in
plot. Value overrides \code{cohortlabfrac}.}

\item{lwd}{Line width for plot elements.}

\item{plotdir}{Directory where PNG files will be written.}

\item{xlab}{x-label for all plots}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files. See
\code{help("png")} in R for details.}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Cumulative catch contributions for each cohort are plotted based on
estimated catch-at-age matrix and weight-at-age values by fleet.  Curves are
shown in units of both numbers and biomass.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SS_output]{SS_output()}}
}
\author{
Ian Taylor
}
