% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotComparisons.R
\name{SSplotComparisons}
\alias{SSplotComparisons}
\title{plot model comparisons}
\usage{
SSplotComparisons(
  summaryoutput,
  subplots = 1:20,
  plot = TRUE,
  print = FALSE,
  png = print,
  pdf = FALSE,
  models = "all",
  endyrvec = NULL,
  indexfleets = NULL,
  indexUncertainty = TRUE,
  indexQlabel = TRUE,
  indexQdigits = 4,
  indexSEvec = NULL,
  indexPlotEach = FALSE,
  labels = c("Year", "Spawning biomass (t)", "Fraction of unfished spawning biomass",
    "Age-0 recruits (1,000s)", "Recruitment deviations", "Index", "Log index",
    "SPR-related quantity", "Density", "Management target",
    "Minimum stock size threshold", "Spawning output", "Harvest rate",
    "Summary biomass (t)", "Age X+ biomass (t)"),
  col = NULL,
  shadecol = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  spacepoints = 10,
  staggerpoints = 1,
  initpoint = 0,
  tickEndYr = TRUE,
  shadeForecast = TRUE,
  xlim = NULL,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  type = "o",
  uncertainty = TRUE,
  shadealpha = 0.1,
  legend = TRUE,
  legendlabels = summaryoutput[["modelnames"]],
  legendloc = "topright",
  legendorder = NULL,
  legendncol = 1,
  sprtarg = NULL,
  btarg = NULL,
  minbthresh = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  plotdir = NULL,
  filenameprefix = "",
  densitynames = c("SSB_Virgin", "SR_LN(R0)"),
  densityxlabs = NULL,
  rescale = TRUE,
  densityscalex = 1,
  densityscaley = 1,
  densityadjust = 1,
  densitysymbols = TRUE,
  densitytails = TRUE,
  densitymiddle = FALSE,
  densitylwd = 1,
  fix0 = TRUE,
  new = TRUE,
  add = FALSE,
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  mcmcVec = FALSE,
  show_equilibrium = TRUE,
  ...
)
}
\arguments{
\item{summaryoutput}{List created by \code{SSsummarize}}

\item{subplots}{Vector of subplots to be created
Numbering of subplots is as follows:
\itemize{
\item 1  spawning biomass
\item 2  spawning biomass with uncertainty intervals
\item 3  biomass ratio (hopefully equal to fraction of unfished)
\item 4  biomass ratio with uncertainty
\item 18  summary biomass
\item 19  summary biomass with uncertainty
\item 5  SPR ratio
\item 6  SPR ratio with uncertainty
\item 7  F value
\item 8  F value with uncertainty
\item 9  recruits
\item 10  recruits with uncertainty
\item 11  recruit devs
\item 12  recruit devs with uncertainty
\item 13  index fits
\item 14  index fits on a log scale
\item 15  phase plot
\item 16  densities
\item 17  cumulative densities
}}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{png}{Has same result as \code{print}, included for consistency with
\code{SS_plots}.}

\item{pdf}{Write output to PDF file? Can't be used in conjunction with
\code{png} or \code{print}.}

\item{models}{Optional subset of the models described in
\code{summaryoutput}.  Either "all" or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model. If the number of models is subset using
the \code{models} input then \code{endyr} needs to be shortened as well.}

\item{indexfleets}{Fleet numbers for each model to compare
indices of abundance. Can take different forms:
\itemize{
\item NULL: (default) create a separate plot for each index as long as the fleet
numbering is the same across all models.
\item integer: create a single comparison plot for the chosen index
\item vector of length equal to number of models: a single fleet number
for each model to be compared in a single plot
\item list: list of fleet numbers associated with indices within each
model to be compared, where the list elements are each a vector of the
same length but the names of the list elements don't matter and can be
absent.
}}

\item{indexUncertainty}{Show uncertainty intervals on index data?
Default=FALSE because if models have any extra standard deviations added,
these intervals may differ across models.}

\item{indexQlabel}{Add catchability to legend in plot of index fits
(TRUE/FALSE)?}

\item{indexQdigits}{Number of significant digits for catchability in legend
(if \code{indexQlabel = TRUE})}

\item{indexSEvec}{Optional replacement for the SE values in
\code{summaryoutput[["indices"]]} to deal with the issue of differing uncertainty by
models described above.}

\item{indexPlotEach}{TRUE plots the observed index for each model with
colors, or FALSE just plots observed once in black dots.}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{col}{Optional vector of colors to be used for lines. Input NULL
makes use of \code{rich.colors.short} function.}

\item{shadecol}{Optional vector of colors to be used for shading uncertainty
intervals. The default (NULL) is to use the same colors provided by
\code{col} (either the default or a user-chosen input) and make them
more transparent by applying the \code{shadealpha} input as an alpha
transparency value (using the \code{adjustcolor()} function)}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{spacepoints}{Number of years between points shown on top of lines (for
long timeseries, points every year get mashed together)}

\item{staggerpoints}{Number of years to stagger the first point (if
\code{spacepoints > 1}) for each line (so that adjacent lines have points in
different years)}

\item{initpoint}{Year value for first point to be added to lines.
Points added to plots are those that satisfy
(Yr-initpoint)\%\%spacepoints == (staggerpoints*iline)\%\%spacepoints}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis mark at final
year in timeseries plots.}

\item{shadeForecast}{TRUE/FALSE switch to turn on off shading of years beyond
the maximum ending year of the models}

\item{xlim}{Optional x limits}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space
to fit legend if necessary. Default=1.05.}

\item{xaxs}{Choice of xaxs parameter (see ?par for more info)}

\item{yaxs}{Choice of yaxs parameter (see ?par for more info)}

\item{type}{Type parameter passed to points (default 'o' overplots points on
top of lines)}

\item{uncertainty}{Show plots with uncertainty intervals? Either a single
TRUE/FALSE value, or a vector of TRUE/FALSE values for each model,
or a set of integers corresponding to the choice of models.}

\item{shadealpha}{Transparency adjustment used to make default shadecol
values (implemented as \code{adjustcolor(col=col, alpha.f=shadealpha)})}

\item{legend}{Add a legend?}

\item{legendlabels}{Optional vector of labels to include in legend. Default
is \code{summaryoutput[["modelnames"]]}.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a
vector of two numeric values representing the fraction of the maximum in
the x and y dimensions, respectively. See \code{help("legend")} for more info
on the string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have
the legend display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend.}

\item{sprtarg}{Target value for SPR-ratio where line is drawn in the SPR
plots and phase plot.}

\item{btarg}{Target biomass value at which to show a line (set to 0 to
remove)}

\item{minbthresh}{Minimum biomass threshold at which to show a line (set to
0 to remove)}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files. See
\code{help("png")} in R for details.}

\item{plotdir}{Directory where PNG files will be written.}

\item{filenameprefix}{Additional text to append to PNG or PDF file names.
It will be separated from default name by an underscore.}

\item{densitynames}{Vector of names (or subset of names) of parameters or
derived quantities contained in \code{summaryoutput[["pars"]][["Label"]]} or
\code{summaryoutput[["quants"]][["Label"]]} for which to make density plots}

\item{densityxlabs}{Optional vector of x-axis labels to use in the density
plots (must be equal in length to the printed vector of quantities that
match the \code{densitynames} input)}

\item{rescale}{TRUE/FALSE control of automatic rescaling of units into
thousands, millions, or billions}

\item{densityscalex}{Scalar for upper x-limit in density plots (values below
1 will cut off the right tail to provide better contrast among narrower
distributions}

\item{densityscaley}{Scalar for upper y-limit in density plots (values below
1 will cut off top of highest peaks to provide better contrast among broader
distributions}

\item{densityadjust}{Multiplier on bandwidth of kernel in density function
used for smoothing MCMC posteriors. See 'adjust' in ?density for details.}

\item{densitysymbols}{Add symbols along lines in density plots. Quantiles
are \code{c(0.025,0.1,0.25,0.5,0.75,0.9,0.975)}.}

\item{densitytails}{Shade tails outside of 95\% interval darker in
density plots?}

\item{densitymiddle}{Shade middle inside of 95\% interval darker in
density plots?}

\item{densitylwd}{Line width for density plots}

\item{fix0}{Always include 0 in the density plots?}

\item{new}{Create new empty plot window}

\item{add}{Allows single plot to be added to existing figure. This needs to
be combined with specific 'subplots' input to make sure only one thing gets
added.}

\item{par}{list of graphics parameter values passed to the \code{par}
function}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{mcmcVec}{Vector of TRUE/FALSE values (or single value) indicating
whether input values are from MCMC or to use normal distribution around
MLE}

\item{show_equilibrium}{Whether to show the equilibrium values for
SSB. For some model comparisons, these might not be comparable and thus
useful to turn off. Defaults to TRUE.}

\item{...}{Additional arguments passed to \code{\link[=add_legend]{add_legend()}}, which
can include \code{legendcex}, \code{legendsp}, and \code{pt.cex}.}
}
\description{
Creates a user-chosen set of plots comparing model output from a summary of
multiple models, where the collection was created using the
\code{SSsummarize} function.
}
\examples{
\dontrun{
# directories where models were run need to be defined
dir1 <- system.file("extdata", "simple_small", package = "r4ss")
dir2 <- system.file("extdata", "simple_small_nohess", package = "r4ss")

# read two models
mod1 <- SS_output(dir = dir1)
mod2 <- SS_output(dir = dir2)

# create list summarizing model results
mod.sum <- SSsummarize(list(mod1, mod2))

# plot comparisons
SSplotComparisons(mod.sum, legendlabels = c("First model", "Second model"))

# Example showing comparison of MLE to MCMC results where the mcmc would have
# been run in the subdirectory "mcmc"
mod1 <- SS_output(
  dir = system.file("extdata", "simple_small_mcmc", package = "r4ss"),
  dir.mcmc = "mcmc"
)
# pass the same model twice to SSsummarize in order to plot it twice
mod.sum <- SSsummarize(list(mod1, mod1))
# compare MLE to MCMC
SSplotComparisons(mod.sum,
  legendlabels = c("MCMC", "MLE"),
  mcmcVec = c(TRUE, FALSE)
)
}

}
\seealso{
Other model comparison functions:
\code{\link[=SSgetoutput]{SSgetoutput()}},
\code{\link[=SSsummarize]{SSsummarize()}},
\code{\link[=SStableComparisons]{SStableComparisons()}}
}
\author{
Ian G. Taylor, John R. Wallace
}
\concept{model comparison functions}
