% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotDiscard.R
\name{SSplotDiscard}
\alias{SSplotDiscard}
\title{Plot fit to discard fraction.}
\usage{
SSplotDiscard(
  replist,
  subplots = 1:4,
  plot = TRUE,
  print = FALSE,
  plotdir = replist[["inputs"]][["dir"]],
  fleets = "all",
  fleetnames = replist[["FleetNames"]],
  datplot = FALSE,
  labels = c("Year", "Discard fraction", "Total discards", "for"),
  yhi = 1,
  ymax = NULL,
  col1 = "blue",
  col2 = "black",
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{subplots}{Vector of which plots to make
\itemize{
\item 1 data only
\item 2 data with fit
\item 3 data only (log scale)
\item 4 data with fit (log scale)
}
If \code{plotdat = FALSE} then subplots 1 and 3 are not created, regardless of
choice of \code{subplots}.}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{plotdir}{Directory where PNG files will be written.}

\item{fleets}{Either the string "all", or a vector of numerical values, like
c(1,3), listing fleets or surveys to be included in the plot.}

\item{fleetnames}{Optional replacement for fleetnames used in data file.}

\item{datplot}{Make data-only plot of discards? This can override the choice
of \code{subplots}.}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{yhi}{Maximum y-value which will always be included in the plot
(all data included regardless). Default = 1 so that discard fractions are always
plotted on a 0-1 range, but total discard amounts which are greater than this value
will exceed it.}

\item{ymax}{Optional maximum y-value to include (useful if upper tails on
discard amounts are very high)}

\item{col1}{First color to use in plot (for expected values)}

\item{col2}{Second color to use in plot (for observations and intervals)}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files. See
\code{help("png")} in R for details.}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Plot fit to discard fraction from Stock Synthesis output file.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}
}
\author{
Ian G. Taylor, Ian J. Stewart, Robbie L. Emmet
}
