% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotMnwt.R
\name{SSplotMnwt}
\alias{SSplotMnwt}
\title{Plot mean weight data and fits.}
\usage{
SSplotMnwt(
  replist,
  subplots = 1:2,
  ymax = NULL,
  plot = TRUE,
  print = FALSE,
  fleets = "all",
  fleetnames = replist[["FleetNames"]],
  datplot = FALSE,
  labels = c("Year", "discard", "retained catch", "whole catch",
    "Mean individual body weight (kg)", "Mean weight in", "for"),
  col1 = "blue",
  col2 = "black",
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = replist[["inputs"]][["dir"]],
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{subplots}{Vector of which plots to make (1 = data only, 2 = with fit).
If \code{plotdat = FALSE} then subplot 1 is not created, regardless of
choice of \code{subplots}.}

\item{ymax}{Optional input to override default ymax value.}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{fleets}{Either the string "all", or a vector of numerical values, like
c(1,3), listing fleets or surveys to be included in the plot.}

\item{fleetnames}{Optional replacement for fleetnames used in data file.}

\item{datplot}{Make data-only plot of discards? This can override the choice
of \code{subplots}.}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{col1}{first color to use in plot (for expected values)}

\item{col2}{second color to use in plot (for observations and intervals)}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files. See
\code{help("png")} in R for details.}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{plotdir}{Directory where PNG files will be written.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Plot mean weight data and fits from Stock Synthesis output. Intervals are
based on T-distributions as specified in model.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SS_output]{SS_output()}}
}
\author{
Ian Taylor, Ian Stewart
}
