% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotTags.R
\name{SSplotTags}
\alias{SSplotTags}
\title{Plot tagging data and fits}
\usage{
SSplotTags(
  replist = replist,
  subplots = 1:10,
  latency = NULL,
  taggroups = NULL,
  rows = 1,
  cols = 1,
  tagrows = 3,
  tagcols = 3,
  plot = TRUE,
  print = FALSE,
  pntscalar = 2.6,
  minnbubble = 8,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  ptsize = 10,
  res = 300,
  cex.main = 1,
  col1 = rgb(0, 0, 1, 0.7),
  col2 = "red",
  col3 = "grey95",
  col4 = "grey70",
  labels = c("Year", "Frequency", "Tag Group", "Fit to tag recaptures by tag group",
    "Post-latency tag recaptures aggregated across tag groups",
    "Observed tag recaptures by year and tag group",
    "Residuals for post-latency tag recaptures: (obs-exp)/sqrt(exp)",
    "Observed and expected post-latency tag recaptures by year and tag group",
    "Summarized observed and expected numbers of recaptures by fleet",
    "Pearson residuals by tag group"),
  plotdir = replist[["inputs"]][["dir"]],
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{subplots}{vector controlling which subplots to create}

\item{latency}{period of tag mixing to exclude from plots (in future could
be included in SS output)}

\item{taggroups}{which tag groups to include in the plots. Default=NULL
causes all groups to be included.}

\item{rows}{number or rows of panels for regular plots}

\item{cols}{number or columns of panels for regular plots}

\item{tagrows}{number or rows of panels for multi-panel plots}

\item{tagcols}{number or columns of panels for multi-panel plots}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{pntscalar}{maximum bubble size for balloon plots; each plot scaled
independently based on this maximum size and the values plotted. Often some
plots look better with one value and others with a larger or smaller value.
Default=2.6}

\item{minnbubble}{minimum number of years below which blank years will be
added to bubble plots to avoid cropping}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{ptsize}{Point size for plotted text in plots printed to files. See
\code{help("png")} in R for details.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{col1}{color for bubbles}

\item{col2}{color for lines with expected values}

\item{col3}{shading color for observations within latency period}

\item{col4}{shading color for observations after latency period}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{plotdir}{Directory where PNG files will be written.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Plot observed and expected tag recaptures in aggregate and by tag group.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SS_output]{SS_output()}}
}
\author{
Andre E. Punt, Ian G. Taylor, Ashleigh J. Novak
}
