% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackpoly.R
\name{stackpoly}
\alias{stackpoly}
\title{modified from "stackpoly" by Jim Lemon from "plotrix" package}
\usage{
stackpoly(x, y, main = "", xlab = "", ylab = "", xat = NA,
  xaxlab = NA, xlim = NA, ylim = NA, lty = 1, border = NA, col = NA,
  axis4 = F, x.hash = NULL, density = 20, ...)
}
\arguments{
\item{x}{A numeric data frame or matrix with the 'x' values. If 'y' is NULL,
these will become the 'y' values and the 'x' positions will be the integers
from 1 to dim(x)[1].}

\item{y}{The 'y' values.}

\item{main}{The title for the plot.}

\item{xlab}{x axis labels for the plot.}

\item{ylab}{y axis labels for the plot.}

\item{xat}{Where to put the optional xaxlabs.}

\item{xaxlab}{Optional labels for the x positions.}

\item{xlim}{Optional x limits.}

\item{ylim}{Optional y limits.}

\item{lty}{Line type for the polygon borders.}

\item{border}{Color for the polygon borders.}

\item{col}{Color to fill the polygons. If NULL, 'rainbow' will be called to
generate the colors. If NA, the polygons will not be filled.}

\item{axis4}{option to add an axis on the right hand side.}

\item{x.hash}{values from x for which the bars have hash marks instead of solid fill}

\item{density}{density value for hashed areas}

\item{\dots}{Additional arguments passed to 'plot'.}
}
\description{
Plot one or more columns of numeric values as the top edges of polygons
instead of lines.
}
\references{
\url{https://cran.r-project.org/package=plotrix}
}
\author{
Jim Lemon, Ian Taylor
}
