% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_compweight.R
\name{table_compweight}
\alias{table_compweight}
\title{Table of length and age comp Francis weights
(doesn't yet include other data types)}
\usage{
table_compweight(
  replist,
  dir = replist[["inputs"]][["dir"]],
  caption = paste("Data weightings applied to compositions",
    "according to the `Francis` method. `Obs.` refers to the number of unique",
    "composition vectors included in the likelihood. `N input` and `N adj.`",
    "refer to the sample sizes of those vectors before and after being adjusted",
    "by the the weights."),
  caption_CAAL = "`CAAL` is conditional age-at-length data.",
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{caption}{A character string for the caption.}

\item{caption_CAAL}{Additional text added to the caption for models with conditional age at length data.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
A list containing the table and caption.
}
\description{
Table of length and age comp Francis weights
(doesn't yet include other data types)
}
\examples{
\dontrun{
# Load the model output
output <- r4ss::SS_output()
# Create the table
table_compweight(output)
}

}
\seealso{
Other table functions:
\code{\link[=table_all]{table_all()}},
\code{\link[=table_biology]{table_biology()}},
\code{\link[=table_config]{table_config()}},
\code{\link[=table_exec_summary]{table_exec_summary()}},
\code{\link[=table_parcounts]{table_parcounts()}},
\code{\link[=table_pars]{table_pars()}},
\code{\link[=table_ts]{table_ts()}}
}
\author{
Kelli F. Johnson, Ian G. Taylor
}
\concept{table functions}
