% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pars.R
\name{table_pars}
\alias{table_pars}
\title{Table of parameters}
\usage{
table_pars(
  replist,
  dir = replist[["inputs"]][["dir"]],
  rows = NULL,
  caption =
    "Parameter estimates, estimation phase, parameter bounds, estimation status, estimated standard deviation (SD), prior information [distribution(mean, SD)] used in the base model.",
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{rows}{Which rows to include from the \code{parameters} table created by
\code{\link[=SS_output]{SS_output()}}. NULL will cause all rows to be included.}

\item{caption}{A character string for the caption.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Table of parameters
}
\seealso{
Other table functions:
\code{\link[=table_all]{table_all()}},
\code{\link[=table_biology]{table_biology()}},
\code{\link[=table_compweight]{table_compweight()}},
\code{\link[=table_config]{table_config()}},
\code{\link[=table_exec_summary]{table_exec_summary()}},
\code{\link[=table_parcounts]{table_parcounts()}},
\code{\link[=table_ts]{table_ts()}}
}
\author{
Kelli F. Johnson
}
\concept{table functions}
