% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pars.R
\name{table_pars_split}
\alias{table_pars_split}
\title{Multiple tables of parameters (split to avoid long tables)}
\usage{
table_pars_split(
  output,
  rows = NULL,
  rows_per_table = 40,
  caption =
    "Parameter estimates, estimation phase, parameter bounds, estimation status, estimated standard deviation (SD), prior information [distribution(mean, SD)] used in the base model.",
  label = "table-pars-base"
)
}
\arguments{
\item{output}{A list from \code{\link[=SS_output]{SS_output()}}.}

\item{rows}{Which rows to include from the \code{parameters} table created by
\code{\link[=SS_output]{SS_output()}}. NULL will cause all rows to be included.}

\item{rows_per_table}{Number of rows to include in each table.}

\item{caption}{A character string for the caption. The caption will be
appended to include the parameter number in each table
(e.g. "Table 1 of 3 showing parameters 1-25.").}

\item{label}{A character string with the label for the table.
No underscores allowed. The label will be appended the table number if the
table is split (e.g. "table-pars-base" will be changed to
"table-pars-base-1")}
}
\description{
Multiple tables of parameters (split to avoid long tables)
}
\seealso{
\code{\link[=table_pars]{table_pars()}}
}
\author{
Ian G. Taylor
}
