% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ts.R
\name{table_ts}
\alias{table_ts}
\title{Table of estimated time series}
\usage{
table_ts(
  replist,
  dir = replist[["inputs"]][["dir"]],
  caption = "Time series of population estimates for the base model.",
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{caption}{A character string for the caption.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Adapted from Lingcod 2021:
https://github.com/pfmc-assessments/lingcod/blob/main/R/table_ts.R
}
\examples{
\dontrun{
table_ts(model)
}
}
\seealso{
Other table functions:
\code{\link[=table_all]{table_all()}},
\code{\link[=table_biology]{table_biology()}},
\code{\link[=table_compweight]{table_compweight()}},
\code{\link[=table_config]{table_config()}},
\code{\link[=table_exec_summary]{table_exec_summary()}},
\code{\link[=table_parcounts]{table_parcounts()}},
\code{\link[=table_pars]{table_pars()}}
}
\author{
Kelli F. Johnson, Ian G. Taylor
}
\concept{table functions}
