% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readctl_3.30.R
\name{translate_3.30_to_3.24_Q_setup}
\alias{translate_3.30_to_3.24_Q_setup}
\title{Use 3.30 q options to create the 3.24 q setup}
\usage{
translate_3.30_to_3.24_Q_setup(
  Q_options,
  Nfleets,
  fleetnames = seq_len(Nfleets)
)
}
\arguments{
\item{Q_options}{The Q options list element in the 3.30
control file r4ss list output generated from \link{SS_readctl}.}

\item{Nfleets}{Number of fleets in the model}

\item{fleetnames}{Name of the fleets. Defaults to fleet numbers, in the order}
}
\value{
A dataframe containing the 3.24 Q setup.
}
\description{
Use 3.30 q options to create the 3.24 q setup
}
