% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readctl_3.30.R
\name{translate_3.30_to_3.24_var_adjust}
\alias{translate_3.30_to_3.24_var_adjust}
\title{Use 3.30 variance adjustments to create the 3.24 formatting}
\usage{
translate_3.30_to_3.24_var_adjust(
  Variance_adjustment_list = NULL,
  Nfleets,
  fleetnames = seq_len(Nfleets)
)
}
\arguments{
\item{Variance_adjustment_list}{The Variance_adjustments_list element
in the control file r4ss list output generated from \link{SS_readctl}.
Defaults to NULL, which can be the case if no variance adjustments were
included in the model.}

\item{Nfleets}{Number of fleets in the model}

\item{fleetnames}{Name of the fleets. Defaults to fleet numbers, in the order
defined in the model.}
}
\value{
A dataframe of 3.24 variance adjustments.
}
\description{
This functionality used to be in SS_readctl_3.30, but ware removed to avoid
confusion.
}
