% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss3_exe.r
\name{tryCatch.W.E}
\alias{tryCatch.W.E}
\title{tryCatch with warning and error}
\usage{
tryCatch.W.E(expr)
}
\arguments{
\item{expr}{function to evaluate}
}
\value{
A warning if something went wrong with the function evaluated
}
\description{
Get warning and error from tryCatch
}
\author{
Kelli Johnson
}
