% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fwf4.r
\name{write_fwf4}
\alias{write_fwf4}
\title{Function to write formatted table similar to
table written by gdata::write.fwf from data.frame or matrix
This function does not accept columns or logical with factor}
\usage{
write_fwf4(
  x,
  file = "",
  append = FALSE,
  quote = FALSE,
  sep = " ",
  na = "NA",
  rownames = FALSE,
  colnames = TRUE,
  rowCol = NULL,
  justify = "left",
  width = NULL,
  eol = "\\n",
  qmethod = c("escape", "double"),
  digits = 6,
  checkNA = TRUE,
  checkInfty = TRUE,
  checkError = TRUE
)
}
\arguments{
\item{x}{data.frame or matrix the object to be written}

\item{file}{either a character string naming a file or a
connection open for writing. "" indicates output to the console.}

\item{append}{logical, append to existing data in \code{file}}

\item{quote}{logical, quote data in output}

\item{sep}{character, separator between columns in output}

\item{na}{character, the string to use for missing values i.e. \code{NA} in
the output}

\item{rownames}{logical, print row names}

\item{colnames}{logical, print column names}

\item{rowCol}{character, rownames column name}

\item{justify}{character, alignment of character columns; see
\code{\link[=format]{format()}}}

\item{width}{numeric, width of the columns in the output}

\item{eol}{the character(s) to print at the end of each line (row).  For
example, 'eol="\\r\\n"' will produce Windows' line endings on a Unix-alike OS,
and 'eol="\\r"' will produce files as expected by Mac OS Excel 2004.}

\item{qmethod}{a character string specifying how to deal with embedded
double quote characters when quoting strings.  Must be one of '"escape"'
(default), in which case the quote character is escaped in C style by a
backslash, or '"double"', in which case it is doubled.  You can specify just
the initial letter.}

\item{digits}{Used for signif}

\item{checkNA}{logical if TRUE, function will stop when NA is found}

\item{checkInfty}{logical if TRUE, function will stop when Infinity is found}

\item{checkError}{logical if TRUE both, set checkNA and checkInftr TRUE}
}
\description{
Function to write formatted table similar to
table written by gdata::write.fwf from data.frame or matrix
This function does not accept columns or logical with factor
}
\author{
Yukio Takeuchi
}
