% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{writeComment}
\alias{writeComment}
\title{Add a comment line to the input files}
\usage{
writeComment(text, con, ...)
}
\arguments{
\item{text}{Comment to write}

\item{con}{File to write to (passed to \code{con} input to \code{writeLines()})}

\item{...}{Additional arguments passed to \code{writeLines()}}
}
\description{
Used by the SS_write* functions.
}
