% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{add_legend}
\alias{add_legend}
\title{Add legend to plots}
\usage{
add_legend(
  legendlabels,
  legendloc = "topleft",
  legendorder = 1,
  cumulative = FALSE
)
}
\arguments{
\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a
vector of two numeric values representing the fraction of the maximum in
the x and y dimensions, respectively. See \code{help("legend")} for more info on
the string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have
the legend display the model names in an order that is different than that
which is represented in the summary input object.}

\item{cumulative}{default for legend location set to topleft}
}
\description{
ss3diags function to add legend to plots
}
