% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{add_legend}
\alias{add_legend}
\title{Add legend to plots}
\usage{
add_legend(
  legendlabels,
  cumulative = FALSE,
  legendloc = "topleft",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.9,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  type = "l"
)
}
\arguments{
\item{legendlabels}{Optional vector of labels to include in legend.}

\item{cumulative}{default for legend location set to topleft}

\item{legendloc}{Location of legend. Either a string like "topleft" or a
vector of two numeric values representing the fraction of the maximum in
the x and y dimensions, respectively. See \code{help("legend")} for more info on
the string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have
the legend display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend.}

\item{legendcex}{Allows to adjust legend cex. Defaults to 1.}

\item{legendsp}{Space between legend labels}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{type}{Type parameter passed to points (default 'o' overplots points on
top of lines)}
}
\description{
ss3diags function to add legend to plots
}
