% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dir.R
\name{check_dir}
\alias{check_dir}
\title{Directory check}
\usage{
check_dir(dir, verbose = TRUE)
}
\arguments{
\item{dir}{A file path to the directory of interest.
The default value is \code{dir = NULL}, which leads to using
the current working directory.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Directory check
}
\details{
Check that
\enumerate{
\item The user knows that the data will not be saved if \code{dir = NULL}.
\item The directory exists if it can be created.
\item The function fails if the directory cannot be created.
}

Note: this function was copied from the {nwfscSurvey} package:
https://github.com/pfmc-assessments/nwfscSurvey/blob/main/R/check_dir.R
rather than adding an additional dependency on that package.
}
\examples{
check_dir(getwd(), verbose = FALSE)
# See more output
check_dir(getwd())

}
\author{
Chantel R. Wetzel
}
