% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exe.R
\name{check_exe}
\alias{check_exe}
\title{Find location of executable within path or specified directory}
\usage{
check_exe(exe = "ss", dir = getwd(), verbose = FALSE)
}
\arguments{
\item{exe}{Executable name, optionally including the path to the
executable if not in the specified directory or in the path.}

\item{dir}{The directory where \code{exe} is located (if not in path). Can
also be a vector of directories.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
A list containing the cleaned \code{exe} name based on the input
argument and the path (or paths if \code{dir} is a vector) to where it is found.
}
\description{
Check that the executable name provided in \code{model},
an input argument to numerous \code{r4ss} functions,
does not contain the extension and is available.
}
\author{
Kelli F. Johnson, Ian G. Taylor
}
