% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check input argument \code{model}}
\usage{
check_model(model, mydir = getwd())
}
\arguments{
\item{model}{Executable name of the \code{.exe} file in \code{mydir}
without the extension, e.g., \code{"ss"}.}

\item{mydir}{The directory where \code{model} is located.}
}
\value{
A cleaned \code{model} name based on the input argument.
}
\description{
Check that the executable name provided in \code{model},
an input argument to numerous \code{r4ss} functions,
does not contain the extension and is available.
}
\author{
Kelli Faye Johnson
}
