% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp2long.R
\name{comp2long}
\alias{comp2long}
\title{Convert composition data to long format}
\usage{
comp2long(x, measure = NULL, zero = TRUE)
}
\arguments{
\item{x}{A data frame containing length or generalized size composition data,
or a list containing such a data frame.}

\item{measure}{An optional string indicating the type of measurement, e.g.,
\code{"weight"}.}

\item{zero}{Whether zero frequencies should be included in the output.}
}
\value{
Data frame containing the columns
\preformatted{year  month  fleet  sex  part  Nsamp  length  freq}
for length composition data, or
\preformatted{method  year  month  fleet  sex  part  Nsamp  size  freq}
for generalized size composition data.

The second to last column can be named explicitly via the measure argument.
}
\description{
Convert data frame containing length or generalized size composition data
from wide to long format.
}
\details{
If \code{x} is not a data frame, the function will look for data frames
called \code{x[["lencomp"]]} or \code{x[["dat"]][["lencomp"]]}.
}
\note{
In r4ss list objects, length composition data are stored as \code{lencomp},
while generalized size composition data are stored as
\code{sizefreq_data_list[[i]]}.
}
\examples{
# Read composition data from example model
path <- system.file("extdata", "simple_small", package = "r4ss")
inputs <- SS_read(path)
dat <- SS_readdat(file.path(path, "data.ss"), verbose = FALSE)
x <- dat[["lencomp"]]

# Main argument can be a list or data frame
head(comp2long(inputs))
head(comp2long(dat))
head(comp2long(x))

# Rename second to last column
head(comp2long(x, measure = "weight"))

# Shrink output by omitting zero frequencies
nrow(comp2long(x))
nrow(comp2long(x, zero = FALSE))

}
\seealso{
\code{\link{SS_read}} and \code{\link{SS_readdat}} read composition data from
files into r4ss list objects.
}
