% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RebuildPlot.R
\name{DoProjectPlots}
\alias{DoProjectPlots}
\title{Deprecated function to make plots from Andre Punt's Rebuilder program.}
\usage{
DoProjectPlots(...)
}
\arguments{
\item{...}{Any arguments associated with the now-deprecated functions.}
}
\description{
The function has been moved to
https://github.com/pfmc-assessments/rebuilder.
This function was rarely used because it was specific to U.S. west
coast groundfish stocks that were overfished and in a rebuilding
plan. Therefore there's no need to have it available to all r4ss
users.
}
\author{
Ian G. Taylor
}
