% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_models.R
\name{download_models}
\alias{download_models}
\title{Download SS3 test models}
\usage{
download_models(
  dir = file.path("inst", "extdata"),
  branch = "main",
  overwrite = FALSE
)
}
\arguments{
\item{dir}{A file path where the downloaded \code{"models"} subfolder will be written to.}

\item{branch}{A string specifying the desired branch of
\href{https://github.com/nmfs-ost/ss3-test-models#test-models}{nmfs-ost/ss3-test-models repository}
that you want to download. The default is \code{"main"}, which is the
stable/default branch.}

\item{overwrite}{A logical value specifying if the existing file(s)
should be overwritten. The default value is \code{overwrite = FALSE}.}
}
\value{
Invisibly return a logical revealing whether the files were copied
(TRUE) or not (FALSE). This function is used for its side effects of
downloading SS3 test models.
}
\description{
Download and unzip the models folder stored on GitHub within the
nmfs-ost/ss3-test-models repository.
}
\examples{
download_models(dir = getwd())
model_name <- list.files("models") # see the model names
# remove files
unlink(file.path("models"), recursive = TRUE)
}
\references{
\href{https://github.com/nmfs-ost/ss3-test-models#test-models}{nmfs-ost/ss3-test-models repository}
}
\author{
Kathryn Doering
}
