% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{get_dat_new_name}
\alias{get_dat_new_name}
\title{Get the name of the data .ss_new file in a directory}
\usage{
get_dat_new_name(dir)
}
\arguments{
\item{dir}{Relative or absolute path to a directory}
}
\value{
A string with the name of the data .ss_new file. If not found, will
be NA.
}
\description{
Get the name of the data .ss_new file in a directory
}
