% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{get_par_name}
\alias{get_par_name}
\title{Get the name of the .par file in a directory}
\usage{
get_par_name(dir, verbose = TRUE)
}
\arguments{
\item{dir}{A file path to the directory of interest.
The default value is \code{dir = NULL}, which leads to using
the current working directory.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
A string with the name of the .par file. If not found, will
be NA. If multiple files exist, preference is given to ss3.par
(default as of 3.30.22.1), followed by ss.par, followed by the most recently
modified file with a *.par extension (choosing the first if two were modified
at the same time).
}
\description{
In previous versions of Stock Synthesis,
}
\seealso{
\link{get_dat_new_name}
}
