% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SIS_info.R
\name{get_SIS_info}
\alias{get_SIS_info}
\title{Gather information for the NOAA Species Information System (SIS)}
\usage{
get_SIS_info(
  model,
  dir = model[["inputs"]][["dir"]],
  writecsv = TRUE,
  stock = "StockName",
  assessment_type = "Operational",
  final_year = model[["endyr"]] + 1,
  data_year = model[["endyr"]],
  month,
  sciencecenter = "NWFSC",
  Mgt_Council = "PFMC",
  SpawnOutputLabel = model[["SpawnOutputLabel"]],
  contact = "first.last@noaa.gov",
  review_result = "XXXX",
  catch_input_data = "XXXX",
  abundance_input_data = "XXXX",
  bio_input_data = "XXXX",
  comp_input_data = "XXXX",
  ecosystem_linkage = "XXXX"
)
}
\arguments{
\item{model}{Output from SS_output}

\item{dir}{Directory in which to write the CSV files.}

\item{writecsv}{Write results to a CSV file (where the name will have the
format "[stock]_2019_SIS_info.csv" where \code{stock}
is an additional input}

\item{stock}{String to prepend id info to filename for CSV file}

\item{assessment_type}{"Operational" or "Stock Monitoring Updates"
(or perhaps additional options as well)}

\item{final_year}{Year of assessment and reference points
(typically will be \code{model[["endyr"]] + 1})}

\item{data_year}{Last year of of timeseries data}

\item{sciencecenter}{Origin of assessment report}

\item{Mgt_Council}{Council jurisdiction. Currently only
\code{"PFMC"} (Pacific Fishery Management Council) and \code{"GM"}
(Gulf of Mexico) are the only options.}

\item{SpawnOutputLabel}{Units for spawning output if not in mt
(e.g. "Millions of eggs"). In the future this can be included in the
\code{model} list created by \code{r4ss::SS_output()}}

\item{contact}{Name and/or email address for lead author.}

\item{review_result}{Something like "Full Acceptance"}

\item{catch_input_data}{Qualitative category for catch input data}

\item{abundance_input_data}{Qualitative category for abundance input data}

\item{bio_input_data}{Qualitative category for biological input data}

\item{comp_input_data}{Qualitative category for size/age composition input
data}

\item{ecosystem_linkage}{Qualitative category for ecosystem linkage}
}
\description{
Processes model results contained in the list created by
\code{\link[=SS_output]{SS_output()}} in a format that is more convenient for submission
to SIS. Currently the results are returned invisibly as a list of two tables
and written to a CSV file from which results could be copied into SIS.
In the future some more direct link could be explored to avoid the manual
copy step.
}
\examples{
\dontrun{
# read the model output
model <- SS_output(dir = system.file("extdata/simple_small", package = "r4ss"),
  printstats = FALSE, verbose = FALSE)
# run get_SIS_info:
info <- get_SIS_info(model, stock = "SimpleExample", month = 1)
}

}
\seealso{
\code{\link[=SS_output]{SS_output()}}
}
\author{
Ian G. Taylor, Andi Stephens, LaTreese S. Denson
}
