% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Function to pull values and calculate confidence intervals from model output}
\usage{
get_values(replist, label, yrs, ci_value, single = FALSE)
}
\arguments{
\item{replist}{r4ss model output list created by \code{SS_output()}}

\item{label}{description}

\item{yrs}{Vector of years.}

\item{ci_value}{To calculate confidence intervals, the desired interval must
be specified. The default is 0.95.}

\item{single}{Default is FALSE}
}
\value{
data frame with point estimate and confidence interval low and high values
}
\description{
Function to pull values and calculate confidence intervals from model output
}
\author{
Chantel R. Wetzel, Kelli F. Johnson, Ian G. Taylor
}
