% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getADMBHessian}
\alias{getADMBHessian}
\title{Read admodel.hes file}
\usage{
getADMBHessian(File, FileName)
}
\arguments{
\item{File}{Directory in which .hes file is located.}

\item{FileName}{Nmae of .hes file.}
}
\value{
A list with elements num.pars, hes, hybrid_bounded_flag, and scale.
}
\description{
This function reads in all of the information contained in the
admodel.hes file. Some of this is needed for relaxing the
covariance matrix, and others just need to be recorded and
rewritten to file so ADMB "sees" what it's expecting.
}
\author{
Jim?, Cole?
}

