% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{iterate_jitter}
\alias{iterate_jitter}
\title{Execute a single jittered model run}
\usage{
iterate_jitter(
  i,
  printlikes = TRUE,
  exe = "ss",
  verbose = FALSE,
  init_values_src = 0,
  ...
)
}
\arguments{
\item{i}{Index of the jitter iteration.}

\item{printlikes}{A logical value specifying if the likelihood values should
be printed to the console.}

\item{exe}{Executable name. Can be just the name of the executable
file if it is in the specified directory or in the user's PATH. Can
also include the absolute path or a path relative to the specified
directory. Needs to be a single character string, not a vector.
On Windows, \code{exe} can optionally have the \code{.exe} extension appended;
on Unix-based systems (i.e., Mac and Linux), no extension should be
included.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{init_values_src}{Either zero or one, specifying if the initial values to
jitter should be read from the control file or from the par file, respectively.
Cannot be \code{NULL}. Defaults to zero (initial values read from control file).}

\item{...}{Additional arguments passed to \code{\link[=run]{run()}}}
}
\value{
Negative log-likelihood of one jittered model
}
\description{
Assumes based model for the jitter is in the current working directory.
}
\author{
James T. Thorson, Kelli F. Johnson, Ian G. Taylor,
Kathryn L. Doering, Kiva L. Oken
}
