% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movepars.R
\name{movepars}
\alias{movepars}
\title{Explore movement parameterizations in a GUI}
\usage{
movepars(nareas = 4, accuage = 40, season.duration = 1,
  min.move.age = 0.5)
}
\arguments{
\item{nareas}{Number of areas}

\item{accuage}{Accumulator age}

\item{season.duration}{Length of season (annual rates are scaled by
this value in SS).}

\item{min.move.age}{Minimum age of movement.}
}
\description{
A function to visualize parameterization of movement in Stock Synthesis. It
creates a GUI interface for movement exploration. Based on selectivity GUI
by Tommy Garrison
}
\author{
Ian Taylor
}

