% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PinerPlot.R
\name{PinerPlot}
\alias{PinerPlot}
\title{Make plot of likelihood contributions by fleet}
\usage{
PinerPlot(
  summaryoutput,
  plot = TRUE,
  print = FALSE,
  component = "Length_like",
  main = "Changes in length-composition likelihoods by fleet",
  models = "all",
  fleets = "all",
  fleetnames = summaryoutput[["FleetNames"]][[1]],
  profile.string = "R0",
  profile.label = expression(log(italic(R)[0])),
  exact = FALSE,
  ylab = "Change in -log-likelihood",
  col = NULL,
  pch = NULL,
  lty = 1,
  lty.total = 1,
  lwd = 2,
  lwd.total = 3,
  cex = 1,
  cex.total = 1.5,
  xlim = NULL,
  ymax = NULL,
  xaxs = "r",
  yaxs = "r",
  type = "o",
  legend = TRUE,
  legendloc = "topright",
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  add_cutoff = FALSE,
  cutoff_prob = 0.95,
  verbose = TRUE,
  fleetgroups = NULL,
  likelihood_type = "raw_times_lambda",
  minfraction = 0.01,
  conv_criteria = 0.01
)
}
\arguments{
\item{summaryoutput}{List created by the function
\code{\link[=SSsummarize]{SSsummarize()}}.}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{component}{Which likelihood component to plot. Default is "Length_like".}

\item{main}{Title for plot. Should match component.}

\item{models}{Optional subset of the models described in
\code{summaryoutput}. Can be "all", "converged", or a vector of numbers indicating
columns in summary tables. The default "all" will include all models.
The "converged" option will include only models with a maximum gradient less
than or equal to the specified convergence criterion \code{conv_criteria}.}

\item{fleets}{Either the string "all", or a vector of numerical values, like
c(1,3), listing fleets or surveys to be included in the plot.}

\item{fleetnames}{Optional replacement for fleetnames used in data file.}

\item{profile.string}{Character string used to find parameter over which the
profile was conducted. If \code{exact=FALSE}, this can be a substring of
one of the SS parameter labels found in the Report.sso file.
For instance, the default input 'R0'
matches the parameter 'SR_LN(R0)'. If \code{exact=TRUE}, then
profile.string needs to be an exact match to the parameter label.}

\item{profile.label}{Label for x-axis describing the parameter over which
the profile was conducted.}

\item{exact}{Should the \code{profile.string} have to match the parameter
label exactly, or is a substring OK.}

\item{ylab}{Label for y-axis. Default is "Change in -log-likelihood".}

\item{col}{Optional vector of colors for each line.}

\item{pch}{Optional vector of plot characters for the points.}

\item{lty}{Line total for the likelihood components.}

\item{lty.total}{Line type for the total likelihood.}

\item{lwd}{Line width for plot elements.}

\item{lwd.total}{Line width for the total likelihood.}

\item{cex}{Character expansion for the points representing the likelihood
components.}

\item{cex.total}{Character expansion for the points representing the total
likelihood.}

\item{xlim}{Range for x-axis. Change in likelihood is calculated relative to
values within this range.}

\item{ymax}{Maximum y-value. Default is 10\\% greater than largest value
plotted.}

\item{xaxs}{The style of axis interval calculation to be used for the x-axis
(see ?par for more info)}

\item{yaxs}{The style of axis interval calculation to be used for the y-axis
(see ?par for more info).}

\item{type}{Line type (see ?plot for more info).}

\item{legend}{Add a legend?}

\item{legendloc}{Location of legend. Either a string like "topleft" or a
vector of two numeric values representing the fraction of the maximum in
the x and y dimensions, respectively. See \code{help("legend")} for more info
on the string options.}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{plotdir}{Directory where PNG files will be written.}

\item{add_cutoff}{Add dashed line at ~1.92 to indicate 95\% confidence interval
based on common cutoff of half of chi-squared of p=.95 with 1 degree of
freedom: \code{0.5*qchisq(p=cutoff_prob, df=1)}. The probability value
can be adjusted using the \code{cutoff_prob} below.}

\item{cutoff_prob}{Probability associated with \code{add_cutoff} above.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{fleetgroups}{Optional character vector, with length equal to
the number of declared fleets, where fleets with the same value are
aggregated}

\item{likelihood_type}{choice of "raw" or "raw_times_lambda" (the default)
determines whether or not likelihoods plotted are adjusted by lambdas
(likelihood weights)}

\item{minfraction}{Minimum change in likelihood (over range considered) as a
fraction of change in total likelihood for a component to be included in the
figure.}

\item{conv_criteria}{Maximum gradient for a model to be considered converged,
used when \code{models = "converged"}.}
}
\description{
This style of plot was officially named a "Piner Plot" at the
CAPAM Selectivity Workshop, La Jolla March 2013. This is in
honor of Kevin Piner's contributions to interpreting likelihood
profiles. He's surely not the first person to make such a plot
but the name seems to have stuck.
}
\references{
Kevin Piner says that he's not the originator of this idea so
Athol Whitten is going to add a reference here.
}
\seealso{
Other profile functions:
\code{\link[=SSplotProfile]{SSplotProfile()}},
\code{\link[=profile]{profile()}}
}
\author{
Ian G. Taylor, Kevin R. Piner, James T. Thorson
}
\concept{profile functions}
