% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngfun.R
\name{pngfun}
\alias{pngfun}
\title{Open png device and return info on the file being created}
\usage{
pngfun(
  plotinfo,
  file,
  plotdir,
  pwidth,
  pheight,
  punits,
  res,
  ptsize,
  caption = NA,
  alt_text = NA,
  filenameprefix = "",
  par = NULL
)
}
\arguments{
\item{file}{filename to write to (including .png extension)}

\item{caption}{caption for the image}

\item{alt_text}{alternative text for screen readers (may match caption in some cases)}
}
\description{
this was previously contained within each of the SSplotXXX() functions.
it (1) translates the not-quite-matching specifications for the image to the
values needed by png(), then return (2) returns the plotinfo data.frame
(which exists within each function which calls this) after adding a row
with the filename and caption for each new plot
Note: this just opens the png device which needs to be closed via dev.off()
outside this function
}
\author{
Ian G. Taylor
}
