% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_multiple_folders.R
\name{populate_multiple_folders}
\alias{populate_multiple_folders}
\title{Populate multiple Stock Synthesis folders with input files}
\usage{
populate_multiple_folders(
  outerdir.old,
  outerdir.new,
  create.dir = TRUE,
  overwrite = FALSE,
  use_ss_new = FALSE,
  copy_par = FALSE,
  exe.dir = NULL,
  exe.file = "ss3",
  verbose = TRUE
)
}
\arguments{
\item{outerdir.old}{Location of existing outer directory containing
subdirectories for each model.}

\item{outerdir.new}{New outer directory into which the subfolders
should be created.}

\item{create.dir}{Create new outer directory if it doesn't exist already?}

\item{overwrite}{A logical value specifying if the existing file(s)
should be overwritten. The default value is \code{overwrite = FALSE}.}

\item{use_ss_new}{Use .ss_new files instead of original inputs?}

\item{copy_par}{Copy any .par files found in the individual directories?}

\item{exe.dir}{Where to get executable to copy to each new subfolder.
Options are
\itemize{
\item FALSE to not copy any executables,
\item TRUE to copy executables found in each existing subfolder to the
corresponding new subfolder,
\item a path to a central location containing an executable to copy into
each new subfolder.
}}

\item{exe.file}{Filename of executable to copy into all the subfolders.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
Returns a table of results indicating which directories were
successfully populated with the model input files and/or executables.
}
\description{
Creates a set of multiple folders and populates each with SS3 input files
such as for the purpose of running a new version of SS3 for an existing
set of test models.
}
\examples{
\dontrun{
populate_multiple_folders(
  outerdir.old = system.file("extdata", package = "r4ss"),
  outerdir.new = file.path(tempdir(), "test")
)
}

}
\seealso{
Other run functions: 
\code{\link{copy_SS_inputs}()},
\code{\link{jitter}()},
\code{\link{profile}()},
\code{\link{retro}()},
\code{\link{run}()},
\code{\link{tune_comps}()}
}
\author{
Ian G. Taylor, Kelli F. Johnson
}
\concept{run functions}
