% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4ss-package.R
\docType{package}
\name{r4ss-package}
\alias{r4ss}
\alias{r4ss-package}
\title{r4ss: R Code for Stock Synthesis}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A collection of R functions for use with Stock Synthesis, a fisheries stock assessment modeling platform written in ADMB by Dr. Richard D. Methot at the NOAA Northwest Fisheries Science Center. The functions include tools for summarizing and plotting results, manipulating files, visualizing model parameterizations, and various other common stock assessment tasks. This version of 'r4ss' is compatible with Stock Synthesis versions 3.24 through 3.30 (specifically version 3.30.25, from June 2026). Support for 3.24 models is only through the core functions for reading output and plotting.
}
\examples{
\dontrun{
# it's useful to create a variable for the directory with the model output
mydir <- file.path(
  path.package("r4ss"),
  file.path("extdata", "simple_small")
)

# read the model output and print diagnostic messages
replist <- SS_output(
  dir = mydir,
  verbose = TRUE,
  printstats = TRUE
)

# plots the results
SS_plots(replist)
}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ss/r4ss}
  \item \url{https://r4ss.github.io/r4ss/}
  \item Report bugs at \url{https://github.com/r4ss/r4ss/issues}
}

}
\author{
\strong{Maintainer}: Ian G. Taylor \email{Ian.Taylor@noaa.gov}

Authors:
\itemize{
  \item Ian G. Taylor \email{Ian.Taylor@noaa.gov}
  \item Ian J. Stewart
  \item Allan C. Hicks
  \item Tommy M. Garrison
  \item Andre E. Punt
  \item R.I.C. Chris Francis
  \item John R. Wallace
  \item Chantel R. Wetzel
  \item James T. Thorson
  \item Yukio Takeuchi
  \item Kotaro Ono
  \item Cole C. Monnahan
  \item Christine C. Stawitz
  \item Z. Teresa A'mar
  \item Athol R. Whitten
  \item Kelli F. Johnson
  \item Robbie L. Emmet
  \item Sean C. Anderson
  \item Gwladys I. Lambert
  \item Megan M. Stachura
  \item Andrew B. Cooper
  \item Andi Stephens
  \item Neil L. Klaer
  \item Carey R. McGilliard
  \item Iago Mosqueira
  \item Watal M. Iwasaki
  \item Kathryn L. Doering
  \item Andrea M. Havron
  \item Nathan R. Vaughan
  \item LaTreese S. Denson
  \item Ashleigh J. Novak
  \item Henning Winker
  \item Lee Qi
  \item Megumi Oshima
  \item Eric Fletcher
  \item Elizabeth Perl
  \item Kiva L. Oken
  \item Arni Magnusson
  \item Peter Kuriyama
}

}
\keyword{internal}
