% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4ss-package.R
\docType{package}
\name{r4ss-package}
\alias{r4ss}
\alias{r4ss-package}
\title{r4ss: R Code for Stock Synthesis}
\description{
A collection of R functions for use with Stock Synthesis, a fisheries stock assessment modeling platform written in ADMB by Dr. Richard D. Methot at the NOAA Northwest Fisheries Science Center. The functions include tools for summarizing and plotting results, manipulating files, visualizing model parameterizations, and various other common stock assessment tasks. This version of r4ss is compatible with Stock Synthesis versions 3.24 through 3.30 (specifically version 3.30.19.01, from April 2022).
}
\examples{
\dontrun{
# read in the report file using SS_output
myreplist <- SS_output(dir = "c:/SS/simple/")

# make a collection of plots using SS_plots
SS_plots(replist = myreplist)
}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ss/r4ss}
  \item Report bugs at \url{https://github.com/r4ss/r4ss/issues}
}

}
\author{
Ian G. Taylor, Ian J. Stewart, Allan C. Hicks, Tommy M. Garrison,
Andre E. Punt, John R. Wallace, Chantel R. Wetzel, James T. Thorson,
Yukio Takeuchi, Kotaro Ono, Cole C. Monnahan, Christine C. Stawitz,
Z. Teresa A'mar, Athol R. Whitten, Kelli F. Johnson, Robbie L. Emmet,
Sean C. Anderson, Gwladys I. Lambert, Megan M. Stachura,
Andrew B. Cooper, Andi Stephens, Neil L. Klaer, Carey R. McGilliard,
Iago Mosqueira, Watal M. Iwasaki, Kathryn L. Doering, Andrea M. Havron,
Nathan Vaughan, LaTreese S. Denson, Ashleigh J. Novak, and Henning Winker

Package maintainer: Ian G. Taylor \href{mailto:Ian.Taylor@noaa.gov}{Ian.Taylor@noaa.gov}
}
\keyword{internal}
