<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r4ss_logo.R
\name{r4ss_logo}
\alias{r4ss_logo}
\title{Make a simple logo for r4ss organization on GitHub}
\usage{
r4ss_logo()
}
\description{
I was tired of the automatically generated symbol
that appeared by default.
}
\author{
Ian Taylor
}

=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r4ss_logo.R
\name{r4ss_logo}
\alias{r4ss_logo}
\title{Make a simple logo for r4ss organization on GitHub}
\usage{
r4ss_logo()
}
\description{
I was tired of the automatically generated symbol
that appeared by default.
}
\author{
Ian Taylor
}

>>>>>>> 0cd3a9f8038ee7641644193b1ef48cae5076aaa8
