% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich.colors.short.R
\name{rich.colors.short}
\alias{rich.colors.short}
\title{Make a vector of colors.}
\usage{
rich.colors.short(n, alpha = 1)
}
\arguments{
\item{n}{Number of colors to generate.}

\item{alpha}{Alpha transparency value for all colors in vector. Value is
passed to rgb function.}
}
\description{
A subset of rich.colors by Arni Magnusson from the gplots package, with the
addition of alpha transparency (which is now available in the gplots version
as well)
}
\author{
Arni Magnusson, Ian Taylor
}
\keyword{color}

