% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run multiple Stock Synthesis models}
\usage{
run(
  dir = getwd(),
  exe = "ss",
  extras = "",
  skipfinished = TRUE,
  show_in_console = TRUE,
  console_output_file = "console.output.txt",
  verbose = TRUE
)
}
\arguments{
\item{dir}{Directory, or vector of directories, containing the
model input files.}

\item{exe}{Executable name, optionally including the path to the
executable if not in the specified directory or in the path.}

\item{extras}{Additional ADMB command line arguments passed
to the executable, such as "-nohess"}

\item{skipfinished}{Skip any folders that already contain a Report.sso file.
This can be helpful if the function is interrupted.}

\item{show_in_console}{Show output in the R console? If FALSE,
then it is saved to a file.}

\item{console_output_file}{File to store console output (if
show_in_console = FALSE)}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
Returns table showing which directories had model run and which
had errors like missing executable or Report.sso already present.
The three possible messages are "model run failed", "ran model", and
"unknown run status".
}
\description{
First searches path for the specified executable name
and returns the location of that file if found. If not found in path
it then searches the specified directory. Linux systems have an
existing executable utility \verb{/usr/sbin/ss} in the path. If \code{exe = "ss"} and this file is found by this function, it will be ignored
based on the smaller file size relative to the SS3 executable. Linux
users who want to use the workflow of having an SS3 executable in
their path should name the file something else, such as \code{ss3} or
\code{ss_linux}.
}
\details{
Loops over a vector of directories and iteratively runs SS in each one
}
\examples{
\dontrun{
dir <- file.path(system.file("extdata", package = "r4ss"), "simple_small")
r4ss::run(dir = dir)
}

}
\seealso{
\code{\link[=copy_SS_inputs]{copy_SS_inputs()}},
\code{\link[=populate_multiple_folders]{populate_multiple_folders()}}
}
\author{
Ian Taylor, Kathryn Doering
}
