% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a Stock Synthesis model}
\usage{
run(
  dir = getwd(),
  exe = "ss3",
  extras = "",
  skipfinished = TRUE,
  show_in_console = FALSE,
  console_output_file = "console.output.txt",
  verbose = TRUE
)
}
\arguments{
\item{dir}{Directory containing the model input files.}

\item{exe}{Executable name. Can be just the name of the executable
file if it is in the specified directory or in the user's PATH. Can
also include the absolute path or a path relative to the specified
directory. Needs to be a single character string, not a vector.
On Windows, \code{exe} can optionally have the \code{.exe} extension appended;
on Unix-based systems (i.e., Mac and Linux), no extension should be
included.}

\item{extras}{Additional ADMB command line arguments passed
to the executable, such as "-nohess"}

\item{skipfinished}{Skip any folders that already contain a Report.sso file.
This can be helpful if the function is interrupted while running iteratively.}

\item{show_in_console}{Show output in the R console? If FALSE,
then the console output is saved to a file (specified by
\code{console_output_file}) at the end of the model run.}

\item{console_output_file}{File to store console output (if
show_in_console = FALSE).}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
Returns one of five messages:
"ran model", "model run failed", "unknown run status", "not a
directory", or "contained Report.sso".
}
\description{
The \code{run()} function checks for the executable via
\code{\link[=check_exe]{check_exe()}}. This involves first checking the
specified
directory \code{dir} for the specified SS3 executable name. If it is not
found in the specified
directory, then it checks the PATH. Linux systems may have an
existing executable utility \verb{/usr/sbin/ss} in the path. If \code{exe = "ss"} and this file is found by \code{\link[=check_exe]{check_exe()}}, it will be ignored
based on the smaller file size relative to the SS3 executable. Linux
users who want to use the workflow of having SS3 in their PATH should
name the SS3 file something besides \code{ss}, such as \code{ss3} or
\code{ss_linux}.
}
\details{
Checks for presence of a Stock Synthesis executable and then runs the
model with any additional arguments specified by \code{extras}.
}
\examples{
\dontrun{
dir <- system.file("extdata", "simple_small", package = "r4ss")
r4ss::run(dir = dir)
}

}
\seealso{
Other run functions: 
\code{\link{copy_SS_inputs}()},
\code{\link{jitter}()},
\code{\link{populate_multiple_folders}()},
\code{\link{profile}()},
\code{\link{retro}()},
\code{\link{tune_comps}()}
}
\author{
Ian G. Taylor, Kathryn L. Doering, Kelli F. Johnson
}
\concept{run functions}
