% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run one or more Stock Synthesis models}
\usage{
run(
  dir = getwd(),
  exe = "ss",
  extras = "",
  skipfinished = TRUE,
  show_in_console = TRUE,
  console_output_file = "console.output.txt",
  verbose = TRUE
)
}
\arguments{
\item{dir}{Directory, or vector of directories, containing the
model input files.}

\item{exe}{Executable name. Can be just the name of the executable
file if it is in the specified directory or in the user's PATH. Can
also include the absolute path or a path relative to the specified
directory. Needs to be a single character string, not a vector.
On Windows, \code{exe} can optionally have the \code{.exe} extension appended;
on Unix-based systems (i.e., Mac and Linux), no extension should be
included.}

\item{extras}{Additional ADMB command line arguments passed
to the executable, such as "-nohess"}

\item{skipfinished}{Skip any folders that already contain a Report.sso file.
This can be helpful if the function is interrupted.}

\item{show_in_console}{Show output in the R console? If FALSE,
then it is saved to a file.}

\item{console_output_file}{File to store console output (if
show_in_console = FALSE)}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
Returns a table showing which directories had models run and which
had errors like missing executable or Report.sso already present.
The three possible messages are "model run failed", "ran model", and
"unknown run status".
}
\description{
The \code{run()} function checks for the executable via
\code{\link[=check_exe]{check_exe()}}. This involves first checking the
specified
directory \code{dir} for the specified SS3 executable name. If it is not
found in the specified
directory, then it checks the PATH. Linux systems may have an
existing executable utility \verb{/usr/sbin/ss} in the path. If \code{exe = "ss"} and this file is found by \code{\link[=check_exe]{check_exe()}}, it will be ignored
based on the smaller file size relative to the SS3 executable. Linux
users who want to use the workflow of having SS3 in their PATH should
name the SS3 file something besides \code{ss}, such as \code{ss3} or
\code{ss_linux}.
}
\details{
Checks for presence of a Stock Synthesis executable and then runs the
model with any additional arguments specified by \code{extras}. If input
\code{dir} is a vector, it iteratively runs the SS3 model in each one.
Parallelization is not available at this time.
}
\examples{
\dontrun{
dir <- file.path(system.file("extdata", package = "r4ss"), "simple_small")
r4ss::run(dir = dir)
}

}
\seealso{
\code{\link[=copy_SS_inputs]{copy_SS_inputs()}},
\code{\link[=populate_multiple_folders]{populate_multiple_folders()}}
}
\author{
Ian G. Taylor, Kathryn L. Doering
}
