% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selShapes.R
\name{selShapes}
\alias{selShapes}
\title{A Shiny app that displays various selectivity curves given parameters that would be input into SS
Currently implemented only for 
\enumerate{
  \item logisitic (type 1)
  \item double normal (type 24)
}
This could possibly be hosted on a Shiny server instead of within r4ss}
\usage{
selShapes()
}
\description{
A Shiny app that displays various selectivity curves given parameters that would be input into SS
Currently implemented only for 
\enumerate{
  \item logisitic (type 1)
  \item double normal (type 24)
}
This could possibly be hosted on a Shiny server instead of within r4ss
}
\author{
Allan Hicks, Andrea Havron, Ian Taylor,

inspired by tcl/tk code written by Tommy Garrison
}
