% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_add_catch}
\alias{sim_add_catch}
\title{this function adds a new line for catch in the catch dataframe in the data structure for the SS DAT file}
\usage{
sim_add_catch(dat_struct = NULL, catch_vec = NULL, seas = -1,
  catch_year = -1)
}
\arguments{
\item{dat_struct}{- DAT structure to be edited}

\item{catch_vec}{- catches for each fleet (vector)}

\item{seas}{- season that the catches will be applied}

\item{catch_year}{- year that the catches will be applied}
}
\value{
edited DAT structure
}
\description{
this function adds a new line for catch in the catch dataframe in the data structure for the SS DAT file
}

