% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_calculate_nsamples.r
\name{sim_calculate_nsamples}
\alias{sim_calculate_nsamples}
\title{this function calculates the number of sample sizes needed for
generating a sample from a multinomial distribution with num_bins}
\usage{
sim_calculate_nsamples(num_bins = -1)
}
\arguments{
\item{num_bins}{}
}
\value{
num_samples
}
\description{
this function calculates the number of sample sizes needed for
generating a sample from a multinomial distribution with num_bins
}

