% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{SS_doRetro}
\alias{SS_doRetro}
\title{Deprecated function to run a retrospective analyses, renamed to retro()}
\usage{
SS_doRetro()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
SS_doRetro() has been renamed as \code{\link[=retro]{retro()}}. See
https://github.com/r4ss/r4ss/issues/723 for more details.
}
\seealso{
\code{\link[=retro]{retro()}}
}
\author{
Ian G. Taylor
}
