% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_doRetro.R
\name{SS_doRetro}
\alias{SS_doRetro}
\title{Run retrospective analyses}
\usage{
SS_doRetro(
  dir = getwd(),
  oldsubdir = "",
  newsubdir = "retrospectives",
  subdirstart = "retro",
  years = 0:-5,
  overwrite = TRUE,
  RemoveBlocks = FALSE,
  verbose = FALSE,
  exe = "ss",
  ...
)
}
\arguments{
\item{dir}{Directory where everything takes place.}

\item{oldsubdir}{Subdirectory within \code{masterdir} with existing model
files.}

\item{newsubdir}{Subdirectory within \code{masterdir} where retrospectives
will be run. Default is 'retrospectives'.}

\item{subdirstart}{First part of the pattern of names for the directories in
which the models will actually be run.}

\item{years}{Vector of values to iteratively enter into the starter file for
retrospective year. Should be zero or negative values.}

\item{overwrite}{Overwrite any input files with matching names in the
subdirectories where models will be run.}

\item{RemoveBlocks}{Logical switch determining whether specifications of
blocks is removed from top of control file. Blocks can cause problems for
retrospective analyses, but the method for removing them is overly
simplistic and probably won't work in most cases. Default=FALSE.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{exe}{Executable name, optionally including the path to the
executable if not in the specified directory or in the path.}

\item{...}{Additional arguments passed to r4ss::run(), such as
\code{extras}, and \code{show_in_console}.}

\item{masterdir}{Deprecated. Use \code{dir} instead.}
}
\description{
Do retrospective analyses by creating new directories, copying model files,
and iteratively changing the starter file to set the number of years of data
to exclude. Note that there was a  bug for retrospectives in 3.30.01;
the user should update their model to a newer version of Stock Synthesis to
run retrospectives
}
\examples{
\dontrun{
# note: don't run this in your main directory--make a copy in case something
# goes wrong
mydir <- "C:/Simple"

## retrospective analyses
SS_doRetro(
  masterdir = mydir,
  years = 0:-5
)

retroModels <- SSgetoutput(
  dirvec = file.path(mydir, "retrospectives", paste("retro", 0:-5, sep = ""))
)
retroSummary <- SSsummarize(retroModels)
endyrvec <- retroSummary[["endyrs"]] + 0:-5
SSplotComparisons(retroSummary,
  endyrvec = endyrvec,
  legendlabels = paste("Data", 0:-5, "years")
)
}

}
\seealso{
\code{\link[=SSgetoutput]{SSgetoutput()}}
}
\author{
Ian G. Taylor, James T. Thorson
}
