% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_doRetro.R
\name{SS_doRetro}
\alias{SS_doRetro}
\title{Run retrospective analyses}
\usage{
SS_doRetro(masterdir, oldsubdir, newsubdir = "retrospectives",
  subdirstart = "retro", years = 0:-5, overwrite = TRUE, exefile = "ss",
  extras = "-nox", intern = FALSE, CallType = "system",
  RemoveBlocks = FALSE)
}
\arguments{
\item{masterdir}{Directory where everything takes place.}

\item{oldsubdir}{Subdirectory within \code{masterdir} with existing model
files.}

\item{newsubdir}{Subdirectory within \code{masterdir} where retrospectives
will be run. Default is 'retrospectives'.}

\item{subdirstart}{First part of the pattern of names for the directories in
which the models will actually be run.}

\item{years}{Vector of values to iteratively enter into the starter file for
retrospective year. Should be zero or negative values.}

\item{overwrite}{Overwrite any input files with matching names in the
subdirectories where models will be run.}

\item{exefile}{Executable file found in directory with model files.
On Windows systems, this value will be automatically updated if a single
executable exists in the directory of model files. Input exefile=NULL if
the executable is in your path and doesn't need copying.}

\item{extras}{Additional commands to use when running SS. Default = "-nox"
will reduce the amound of command-line output.}

\item{intern}{Display runtime information from SS in the R console (vs.
saving to a file).}

\item{CallType}{Either "system" or "shell" (choice depends on how you're running
R. Default is "system".}

\item{RemoveBlocks}{Logical switch determining whether specifications of
blocks is removed from top of control file. Blocks can cause problems for
retrospective analyses, but the method for removing them is overly
simplistic and probably won't work in most cases. Default=FALSE.}
}
\description{
Do retrospective analyses by creating new directories, copying model files,
and iteratively changing the starter file to set the number of years of data
to exclude.
}
\examples{

  \dontrun{
    # note: don't run this in your main directory--make a copy in case something goes wrong
    mydir <- "C:/Simple"

    ## retrospective analyses
    SS_doRetro(masterdir=mydir, oldsubdir="", newsubdir="retrospectives", years=0:-5)
    
    retroModels <- SSgetoutput(dirvec=file.path(mydir, "retrospectives",paste("retro",0:-5,sep="")))
    retroSummary <- SSsummarize(retroModels)
    endyrvec <- retroSummary$endyrs + 0:-5
    SSplotComparisons(retroSummary, endyrvec=endyrvec, legendlabels=paste("Data",0:-5,"years"))
  }

}
\seealso{
\code{\link{SSgetoutput}}
}
\author{
Ian Taylor, Jim Thorson
}
