% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_output.R
\name{SS_output}
\alias{SS_output}
\title{A function to create a list object for the output from Stock Synthesis}
\usage{
SS_output(dir = "C:/myfiles/mymodels/myrun/", model = "ss3",
  repfile = "Report.sso", compfile = "CompReport.sso",
  covarfile = "covar.sso", forefile = "Forecast-report.sso",
  wtfile = "wtatage.ss_new", warnfile = "warning.sso", ncols = 200,
  forecast = TRUE, warn = TRUE, covar = TRUE, readwt = TRUE,
  checkcor = TRUE, cormax = 0.95, cormin = 0.01, printhighcor = 10,
  printlowcor = 10, verbose = TRUE, printstats = TRUE,
  hidewarn = FALSE, NoCompOK = FALSE, aalmaxbinrange = 4)
}
\arguments{
\item{dir}{Locates the directory of the files to be read in, double
backslashes (or forwardslashes) and quotes necessary.}

\item{model}{Name of the executable (leaving off the .exe).}

\item{repfile}{Name of the big report file (could be renamed by user).}

\item{compfile}{Name of the composition report file.}

\item{covarfile}{Name of the covariance output file.}

\item{forefile}{Name of the forecast file.}

\item{wtfile}{Name of the file containing weight at age data.}

\item{warnfile}{Name of the file containing warnings.}

\item{ncols}{The maximum number of columns in files being read in.  If this
value is too big the function runs more slowly, too small and errors will
occur.  A warning will be output to the R command line if the value is too
small. It should be bigger than the maximum age + 10 and the number of years
+ 10.}

\item{forecast}{Read the forecast-report file?}

\item{warn}{Read the Warning.sso file?}

\item{covar}{Read covar.sso to get variance information and identify bad
correlations?}

\item{readwt}{Read the weight-at-age file?}

\item{checkcor}{Check for bad correlations?}

\item{cormax}{The specified threshold for defining high correlations.  A
quantity with any correlation above this value is identified.}

\item{cormin}{The specified threshold for defining low correlations.  Only
quantities with all correlations below this value are identified (to find
variables that appear too independent from the model results).}

\item{printhighcor}{The maximum number of high correlations to print to the
R GUI.}

\item{printlowcor}{The maximum number of low correlations to print to the R
GUI.}

\item{verbose}{Return updates of function progress to the R GUI?}

\item{printstats}{Print summary statistics about the output to the R GUI?}

\item{hidewarn}{Hides some warnings output from the R GUI.}

\item{NoCompOK}{Allow the function to work without a CompReport file.}

\item{aalmaxbinrange}{The largest length bin range allowed for composition
data to be considered as conditional age-at-length data.}
}
\value{
Many values are returned in a list. 
Some of the lists items are described below. The first word of the description
is the 'KeyWord' used by Stock Synthesis in the \code{Report.sso} file.
The list elements described below are placed in alphabetical order.
A more complete list is needed in the future as well as documenting 
how each element is generated in the \code{.tpl} files.
\item{endgrowth}{BIOLOGY Biology_at_age_in_endyr_with_ - A data frame
of the biology at age in the final year based on some form of variance
set within Stock Synthesis. The data frame contains the following, where each bullet
is a column name. Pseudo code comes from Stock Synthesis.
\itemize{
\item{Seas - Season}
\item{Morph - Growth morph}
\item{Bio_Pattern - }
\item{Sex - Sex or gender}
\item{Settlement - }
\item{Platoon - }
\item{int_Age - Age}
\item{Real_Age - Real age as defined by \code{real_age(g, ALK_idx, a)}}
\item{Age_Beg - Calendar age as defined by \code{calen_age(g, ALK_idx,a)}}
\item{Age_Mid - Some other age as defined by \code{calen_age(g, ALK_idx_mid, a)}}
\item{M - Natural mortality}
\item{Len_Beg - Average size}
\item{Len_Mid - Average size at mid season as defined by 
\code{Ave_Size(t, mid_subseas, g, a)}}
\item{SD_Beg - \code{Sd_Size_within(ALK_idx,g,a)}}
\item{SD_Mid - \code{Sd_Size_within(ALK_idx_mid,g,a)}}
\item{Wt_Beg - \code{Wt_Age_beg(s,g,a)}}
\item{Wt_Mid - \code{Wt_Age_mid(s,g,a)}}
\item{Len_Mat - \code{ALK(ALK_idx,g,a)*mat_len(GP4(g))}}
\item{Age_Mat - If \code{Maturity_Option <= 2)} 
then \code{SS2out<<mat_age(GP4(g),a)},
\code{Maturity_Option > 2 & sx(g) ==1 && Maturity_Option < 5)} 
then \code{Age_Maturity(GP4(g),a)}, or
\code{else} then \code{-1.0}}
\item{Mat*Fecund - \code{fec(g,a)}}
\item{Mat_F_wtatage - \code{make_mature_bio(g,a)}}
\item{Mat_F_Natage - \code{make_mature_numbers(g,a)}}
\item{Herma_Trans - Optional column if \code{Hermaphro_Option!=0},
\code{Herma_Cum*=(1.0-Hermaphro_val(GP4(g),a-1))} if \code{a > 1}.}
\item{Herma_Cum - Optional column if \code{Hermaphro_Option!=0},
\code{Hermaphro_val(GP4(g),a)}.}
\item{Len:_ - Information for each fleet as defined by the trailing number.}
\item{SelWt:_ - Information for each fleet as defined by the trailing number.}
\item{RetWt:_ - Information for each fleet as defined by the trailing number.}  
}}
\item{likelihoods_by_fleet}{LIKELIHOOD - Fleet-specific components of 
the likelihood.}
\item{likelihoods_used}{LIKELIHOOD - Each component of the likelihood.}
}
\description{
Reads the Report.sso and (optionally) the covar.sso, CompReport.sso and
other files files produced by Stock Synthesis and formats the important
content of these files into a list in the R workspace. A few statistics
unavailable elsewhere are taken from the .par and .cor files. Summary
information and statistics can be returned to the R console or just
contained within the list produced by this function.
}
\examples{

  \dontrun{
    myreplist <- SS_output(dir='c:/SS/SSv3.10b/Simple/')
  }

}
\seealso{
\code{\link{SS_plots}}
}
\author{
Ian Stewart, Ian Taylor
}
