% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SS_readctl}
\alias{SS_readctl}
\title{read control file}
\usage{
SS_readctl(file)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}
}
\description{
read Stock Synthesis control file into list object in R
}
\details{
This function is not fully implemented. The logic to figure out all the
details of a Stock Synthesis control file is very complex, so this function
may be completed in a way that is not totally consistent with the other
similar files. Or it may never be completed at all. The functions
\code{\link{SS_changepars}} and \code{\link{SS_parlines}} offer alternatives
for working with SS control files.
}
\seealso{
\code{\link{SS_changepars}}, \code{\link{SS_parlines}},
\code{\link{SS_readstarter}}, \code{\link{SS_readforecast}},
\code{\link{SS_readdat}}, \code{\link{SS_writestarter}},
\code{\link{SS_writeforecast}}, \code{\link{SS_writedat}},
\code{\link{SS_writectl}}
}
\keyword{data}

