% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readctl_3.24.R
\name{SS_readctl_3.24}
\alias{SS_readctl_3.24}
\title{read control file from SS version 3.24}
\usage{
SS_readctl_3.24(file, verbose = TRUE, echoall = FALSE, nseas = 4,
  N_areas = 1, Nages = 20, Ngenders = 1, Npopbins = NA, Nfish = 2,
  Nsurv = 2, TG_Nrelgrp = NA)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{verbose}{Should there be verbose output while running the file?
Default=TRUE.}

\item{echoall}{Debugging tool (not fully implemented) of echoing blocks of
data as it is being read.}

\item{nseas}{number of season in the model. This information is not
explicitly available in control file}

\item{Nages}{oldest age in the model. This information is also not
explicitly available in control file}

\item{Npopbins}{number of population bins in the model. This information is also not
explicitly available in control file and this information is only required if length based
maturity vector is directly supplied (Maturity option of 6), and not yet tested}

\item{Nfish}{number of fisheries in the model. This information is also not
explicitly available in control file}

\item{Nsurv}{number of survey fleets in the model. This information is also not
explicitly available in control file}

\item{Ns}{number of genders in the model. This information is also not
explicitly available in control file}
}
\description{
Read Stock Synthesis (version 3.24) control file into list object in R.
This function comes with its wrapper function SS_readctl
that calls SS_readctl_3.24 (this function) or SS_readctl_3.30 (to be available in future).
}
\author{
Yukio Takeuchi
}
\seealso{
\code{\link{SS_readctl}}, \code{\link{SS_readdat}}
\code{\link{SS_readdat_3.24}},\code{\link{SS_readdat_3.30}}
\code{\link{SS_readstarter}}, \code{\link{SS_readforecast}},
\code{\link{SS_writestarter}},
\code{\link{SS_writeforecast}}, \code{\link{SS_writedat}}
}

