% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readdat.R
\name{SS_readdat}
\alias{SS_readdat}
\title{read Stock Synthesis data file}
\usage{
SS_readdat(
  file,
  version = "3.30",
  verbose = TRUE,
  echoall = lifecycle::deprecated(),
  section = NULL
)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.

See the formal arguments for a possible default filename.}

\item{version}{SS version number.
Currently "2.00", "3.00", "3.24" or "3.30" are supported, but all
versions prior to "3.30" have been deprecated.
either as character or numeric values (noting that numeric 3.30  = 3.3). If
version is NULL, the version (3.24 or 3.30) will be looked for on the first
line of the file.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{echoall}{Deprecated.}

\item{section}{Which data set to read. Only applies for a data.ss_new file
created by Stock Synthesis. Allows the choice of either expected values
(section=2) or bootstrap data (section=3+). Leaving default of section=NULL
will read input data, (equivalent to section=1).}
}
\description{
Read Stock Synthesis data file into list object in R. This function is a
wrapper which calls SS_readdat_3.30 (previously additional functions,
but they have been deprecated).
}
\seealso{
Other read/write functions: 
\code{\link{SS_read}()},
\code{\link{SS_readctl}()},
\code{\link{SS_readforecast}()},
\code{\link{SS_readstarter}()},
\code{\link{SS_write}()},
\code{\link{SS_writectl}()},
\code{\link{SS_writedat}()},
\code{\link{SS_writeforecast}()},
\code{\link{SS_writestarter}()}
}
\author{
Ian G. Taylor, Allan C. Hicks, Neil L. Klaer, Kelli F. Johnson,
Chantel R. Wetzel, Kathryn L. Doering, Nathan R. Vaughan
}
\concept{read/write functions}
