% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readdat.R
\name{SS_readdat}
\alias{SS_readdat}
\title{read data file}
\usage{
SS_readdat(file, verbose = TRUE, echoall = FALSE, section = NULL)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{verbose}{Should there be verbose output while running the file?
Default=TRUE.}

\item{echoall}{Debugging tool (not fully implemented) of echoing blocks of
data as it is being read.}

\item{section}{Which data set to read. Only applies for a data.ss_new file
created by Stock Synthesis. Allows the choice of either expected values
(section=2) or bootstrap data (section=3+). Leaving default of section=NULL
will read input data, (equivalent to section=1).}
}
\description{
read Stock Synthesis data file into list object in R
}
\author{
Ian Taylor
}
\seealso{
\code{\link{SS_readstarter}}, \code{\link{SS_readforecast}},
\code{\link{SS_readctl}}, \code{\link{SS_writestarter}},
\code{\link{SS_writeforecast}}, \code{\link{SS_writedat}},
\code{\link{SS_writectl}}
}
\keyword{data}

