% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readforecast.R
\name{SS_readforecast}
\alias{SS_readforecast}
\title{read forecast file}
\usage{
SS_readforecast(
  file = "forecast.ss",
  Nfleets = NULL,
  Nareas = NULL,
  nseas = NULL,
  version = "3.30",
  readAll = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.

See the formal arguments for a possible default filename.}

\item{Nfleets}{Number of fleets (not required in 3.30).}

\item{Nareas}{Number of areas (not required in 3.30).}

\item{nseas}{number of seasons (not required in 3.30).}

\item{version}{SS version number. Currently "3.24" or "3.30" are supported,
either as character or numeric values (noting that numeric 3.30  = 3.3).
\code{version = NULL} is no longer the default or an allowed entry.
The default is \code{version = "3.30"}.}

\item{readAll}{Should the function continue even if Forecast = 0 or -1
(at which point SS stops reading)?}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
read Stock Synthesis forecast file into list object in R
}
\seealso{
Other read/write functions: 
\code{\link{SS_read}()},
\code{\link{SS_readctl}()},
\code{\link{SS_readdat}()},
\code{\link{SS_readstarter}()},
\code{\link{SS_write}()},
\code{\link{SS_writectl}()},
\code{\link{SS_writedat}()},
\code{\link{SS_writeforecast}()},
\code{\link{SS_writestarter}()}
}
\author{
Ian G. Taylor, Kelli F. Johnson, Kathryn L. Doering, Nathan R. Vaughan
}
\concept{read/write functions}
