% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readforecast.R
\name{SS_readforecast}
\alias{SS_readforecast}
\title{read forecast file}
\usage{
SS_readforecast(file = "forecast.ss", Nfleets, Nareas, verbose = TRUE)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{Nfleets}{Number of fleets.}

\item{Nareas}{Number of areas.}

\item{verbose}{Should there be verbose output while running the file?}
}
\description{
read Stock Synthesis forecast file into list object in R
}
\author{
Ian Taylor
}
\seealso{
\code{\link{SS_readstarter}}, \code{\link{SS_readdat}},
\code{\link{SS_readctl}}, \code{\link{SS_writestarter}},
\code{\link{SS_writeforecast}}, \code{\link{SS_writedat}},
\code{\link{SS_writectl}}
}
\keyword{data}

