% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readstarter.R
\name{SS_readstarter}
\alias{SS_readstarter}
\title{Read Stock Synthesis starter file as a list}
\usage{
SS_readstarter(file = "starter.ss", verbose = TRUE)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.

See the formal arguments for a possible default filename.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
A list with one element for each line of input values.
List elements containing the name of the control and data file are
particularly helpful, i.e., \code{ctlfile} and \code{datfile}, respectively.
}
\description{
Read Stock Synthesis starter file as a list
}
\examples{
starter_list <- SS_readstarter(
  system.file("extdata", "simple_small", "starter.ss",
    package = "r4ss"
  ),
  verbose = FALSE
)

# The following lines should be TRUE and demonstrate how you can know the
# names of the control and data file given information in the starter file.
starter_list[["ctlfile"]] == "simple_control.ss"
starter_list[["datfile"]] == "simple_data.ss"
}
\seealso{
Other read/write functions: 
\code{\link{SS_readctl}()},
\code{\link{SS_readdat}()},
\code{\link{SS_readforecast}()},
\code{\link{SS_read}()},
\code{\link{SS_writectl}()},
\code{\link{SS_writedat}()},
\code{\link{SS_writeforecast}()},
\code{\link{SS_writestarter}()},
\code{\link{SS_write}()}
}
\author{
Ian G. Taylor, Kathryn L. Doering, Kelli F. Johnson
}
\concept{read/write functions}
