% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readstarter.R
\name{SS_readstarter}
\alias{SS_readstarter}
\title{read starter file}
\usage{
SS_readstarter(file = "starter.ss", verbose = TRUE)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{verbose}{Should there be verbose output while running the file?}
}
\description{
read Stock Synthesis starter file into list object in R
}
\seealso{
\code{\link{SS_readforecast}}, \code{\link{SS_readdat}},
\code{\link{SS_writestarter}},
\code{\link{SS_writeforecast}}, \code{\link{SS_writedat}}
}
\author{
Ian Taylor
}
