% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_RunJitter.R
\name{SS_RunJitter}
\alias{SS_RunJitter}
\title{Iteratively apply the jitter option in SS}
\usage{
SS_RunJitter(
  dir = getwd(),
  mydir = lifecycle::deprecated(),
  Intern = lifecycle::deprecated(),
  Njitter,
  printlikes = TRUE,
  jitter_fraction = NULL,
  init_values_src = NULL,
  exe = "ss",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dir}{Directory where model files are located.}

\item{mydir}{Deprecated. Use \code{dir} instead.}

\item{Intern}{Deprecated. Use \code{show_in_console} instead.}

\item{Njitter}{Number of jitters, or a vector of jitter iterations.
If \code{length(Njitter) > 1} only the iterations specified will be ran,
else \code{1:Njitter} will be executed.}

\item{printlikes}{A logical value specifying if the likelihood values should
be printed to the console.}

\item{jitter_fraction}{The value, typically 0.1, used to define a uniform
distribution in cumulative normal space to generate new initial parameter values.
The default of \code{NULL} forces the user to specify the jitter_fraction
in the starter file, and this value must be greater than zero and
will not be overwritten.}

\item{init_values_src}{Either zero or one, specifying if the initial values to
jitter should be read from the control file or from the par file, respectively.
The default is \code{NULL}, which will leave the starter file unchanged.}

\item{exe}{Executable name. Can be just the name of the executable
file if it is in the specified directory or in the user's PATH. Can
also include the absolute path or a path relative to the specified
directory.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{...}{Additional arguments passed to \code{\link[=run]{run()}}, such as
\code{extras}, \code{show_in_console}, and \code{skipfinished}.}
}
\value{
A vector of likelihoods for each jitter iteration.
}
\description{
Iteratively run a Stock Synthesis model with different jittered starting
parameter values based on the jitter fraction. Output files are renamed
in the format Report1.sso, Report2.sso, etc.
}
\examples{
\dontrun{
#### Run jitter from par file with arbitrary, but common, choice of 0.1
modeldir <- tail(dir(system.file("extdata", package = "r4ss"), full.names = TRUE), 1)
numjitter <- 25
jit.likes <- SS_RunJitter(
  dir = modeldir, Njitter = numjitter,
  jitter_fraction = 0.1, init_value_src = 1
)

#### Read in results using other r4ss functions
# (note that un-jittered model can be read using keyvec=0:numjitter)
profilemodels <- SSgetoutput(dirvec = modeldir, keyvec = 1:numjitter, getcovar = FALSE)
# summarize output
profilesummary <- SSsummarize(profilemodels)
# Likelihoods
profilesummary[["likelihoods"]][1, ]
# Parameters
profilesummary[["pars"]]
}

}
\author{
James T. Thorson, Kelli F. Johnson, Ian G. Taylor
}
