% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_Sensi_plot.R
\name{SS_Sensi_plot}
\alias{SS_Sensi_plot}
\title{Create Cope-Gertseva sensitivity plots}
\usage{
SS_Sensi_plot(
  model.summaries,
  Dir,
  current.year,
  mod.names,
  likelihood.out = c(1, 1, 1),
  Sensi.RE.out = "Sensi_RE_out.DMP",
  CI = 0.95,
  TRP.in = 0.4,
  LRP.in = 0.25,
  sensi_xlab = "Sensitivity scenarios",
  ylims.in = c(-1, 2, -1, 2, -1, 2, -1, 2, -1, 2, -1, 2),
  plot.figs = c(1, 1, 1, 1, 1, 1),
  sensi.type.breaks = NA,
  anno.x = NA,
  anno.y = NA,
  anno.lab = NA
)
}
\arguments{
\item{model.summaries}{Output from \code{\link{SSsummarize}}
summarizing results of models to be included}

\item{Dir}{Directory where plots will be created}

\item{current.year}{Year to report output}

\item{mod.names}{List the names of the sensitivity runs}

\item{likelihood.out}{Note which likelihoods are in the model
(surveys, lengths, ages)}

\item{Sensi.RE.out}{Saved file of relative changes}

\item{CI}{Confidence interval box based on the reference model}

\item{LRP.in}{Limit relative abundance value}

\item{sensi_xlab}{X-axis label}

\item{ylims.in}{Y-axis label}

\item{plot.figs}{Which plots to make/save?}

\item{sensi.type.breaks}{vertical breaks that can separate out types
of sensitivities}

\item{anno.x}{Vertical positioning of the sensitivity types labels}

\item{anno.y}{Horizontal positioning of the sensitivity types labels}

\item{anno.lab}{Sensitivity types labels}

\item{TRP.inTarget}{relative abundance value}
}
\description{
Uses output from \code{\link{SSsummarize}} to make a figure showing
sensitivity of various quantities as described by Cope and Gertseva (2020).
}
\references{
Cope, J. M., V. V. Gertseva. In press. A new way to
visualize and report structural and data uncertainty in stock assessments.
Canadian Journal of Fisheries and Aquatic Sciences.
}
\seealso{
\code{\link{SSsummarize}}
}
\author{
Jason Cope
}
