% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_comps.R
\name{SS_tune_comps}
\alias{SS_tune_comps}
\title{Deprecated function to tune composition data, renamed to tune_comps()}
\usage{
SS_tune_comps()
}
\arguments{
\item{...}{Any arguments associated with the now-deprecated functions.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
SS_tune_comps() has been renamed as \code{\link[=tune_comps]{tune_comps()}}. See
https://github.com/r4ss/r4ss/issues/723 for more details.
}
\seealso{
\code{\link[=tune_comps]{tune_comps()}}
}
\author{
Ian G. Taylor
}
