% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SS_varadjust.R
\name{SS_varadjust}
\alias{SS_varadjust}
\title{Modify variance and sample size adjustments in the control file}
\usage{
SS_varadjust(dir = "C:/myfiles/mymodels/myrun/", ctlfile = "control.ss_new",
  newctlfile = "control_modified.ss", keyword = "Variance_adjustments",
  maxcols = 50, verbose = TRUE)
}
\arguments{
\item{dir}{Directory with control file to change.}

\item{ctlfile}{Control file name. Default="control.ss_new".}

\item{newctlfile}{Name of new control file to be written.
Default="control_modified.ss".}

\item{keyword}{Keyword to use as reference for start of section on
variance adjustments}

\item{maxcols}{Maximum number of columns to search among (may need to
increase from default if you have a huge number of fleets)}

\item{verbose}{TRUE/FALSE switch for amount of detail produced by function.
Default=TRUE.}
}
\description{
Incomplete function in process of development.
}
\author{
Ian Taylor
}
\seealso{
\code{\link{SS_parlines}}, \code{\link{SS_changepars}}
}

