% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writectl_3.24.R
\name{SS_writectl_3.24}
\alias{SS_writectl_3.24}
\title{write control file}
\usage{
SS_writectl_3.24(ctllist, outfile, overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{ctllist}{List object created by \code{\link[=SS_readctl]{SS_readctl()}}.}

\item{outfile}{Filename for where to write new data file.}

\item{overwrite}{Should existing files be overwritten? Default=FALSE.}

\item{verbose}{Should there be verbose output while running the file?
Defaults to FALSE.}
}
\description{
write Stock Synthesis control file from list object in R which was probably
created using \code{\link[=SS_readctl]{SS_readctl()}}
}
\seealso{
\code{\link[=SS_readctl]{SS_readctl()}}, \code{\link[=SS_readctl_3.24]{SS_readctl_3.24()}},\code{\link[=SS_readstarter]{SS_readstarter()}},
}
\author{
Yukio Takeuchi, Kathryn L. Doering, Nathan R. Vaughan
}
