% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writedat.R
\name{SS_writedat}
\alias{SS_writedat}
\title{write data file}
\usage{
SS_writedat(datlist, outfile, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{datlist}{List object created by \code{\link{SS_readdat}}.}

\item{outfile}{Filename for where to write new data file.}

\item{overwrite}{Should existing files be overwritten? Default=FALSE.}

\item{verbose}{Should there be verbose output while running the file?}
}
\description{
write Stock Synthesis data file from list object in R which was probably
created using \code{\link{SS_readdat}}
}
\author{
Ian Taylor, Yukio Takeuchi
}
\seealso{
\code{\link{SS_makedatlist}}, \code{\link{SS_readstarter}},
\code{\link{SS_readforecast}}, \code{\link{SS_readctl}},
\code{\link{SS_writestarter}}, \code{\link{SS_writeforecast}},
\code{\link{SS_writedat}}, \code{\link{SS_writectl}}
}
\keyword{data}
\keyword{manip}

