% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writeforecast.R
\name{SS_writeforecast}
\alias{SS_writeforecast}
\title{write forecast file}
\usage{
SS_writeforecast(
  mylist,
  dir = NULL,
  file = "forecast.ss",
  writeAll = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{mylist}{List object created by \code{\link{SS_readforecast}}.}

\item{dir}{Directory for new forecast file. Default=NULL (working
directory).}

\item{file}{Filename for new forecast file. Default="forecast.ss".}

\item{writeAll}{Should the function continue even if Forecast=0
(at which point SS stops reading, and remaining elements in list may not be
available, depending on settings used in SS_readforecast)}

\item{overwrite}{Should existing files be overwritten? Default=FALSE.}

\item{verbose}{Should there be verbose output while running the file?
Default=TRUE.}
}
\description{
write Stock Synthesis forecast file from list object in R which was probably
created using \code{\link{SS_readforecast}}
}
\seealso{
\code{\link{SS_readstarter}}, \code{\link{SS_readforecast}},
\code{\link{SS_readdat}},
\code{\link{SS_writestarter}}, \code{\link{SS_writedat}}
}
\author{
Ian Taylor
}
