% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writestarter.R
\name{SS_writestarter}
\alias{SS_writestarter}
\title{write starter file}
\usage{
SS_writestarter(
  mylist,
  dir = NULL,
  file = "starter.ss",
  overwrite = FALSE,
  verbose = TRUE,
  warn = lifecycle::deprecated()
)
}
\arguments{
\item{mylist}{List object created by \code{\link[=SS_readstarter]{SS_readstarter()}}.}

\item{dir}{A file path to the directory of interest.
The default value is \code{dir = NULL}, which leads to using
the current working directory.}

\item{file}{Filename for new starter file. Default="starter.ss".}

\item{overwrite}{Should existing files be overwritten? Default=FALSE.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{warn}{Deprecated.}
}
\description{
write Stock Synthesis starter file from list object in R which was probably
created using \code{\link[=SS_readstarter]{SS_readstarter()}}
}
\seealso{
\code{\link[=SS_readstarter]{SS_readstarter()}}, \code{\link[=SS_readforecast]{SS_readforecast()}},
\code{\link[=SS_writestarter]{SS_writestarter()}},
\code{\link[=SS_writeforecast]{SS_writeforecast()}}, \code{\link[=SS_writedat]{SS_writedat()}}
}
\author{
Ian G. Taylor, Kelli F. Johnson, Kathryn R. Doering
}
