% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_writestarter.R
\name{SS_writestarter}
\alias{SS_writestarter}
\title{write starter file}
\usage{
SS_writestarter(mylist, dir = NULL, file = "starter.ss",
  overwrite = FALSE, verbose = TRUE, warn = TRUE)
}
\arguments{
\item{mylist}{List object created by \code{\link{SS_readstarter}}.}

\item{dir}{Directory for new starter file. Default=NULL (working directory).}

\item{file}{Filename for new starter file. Default="starter.ss".}

\item{overwrite}{Should existing files be overwritten? Default=FALSE.}

\item{verbose}{Should there be verbose output while running the file?
Default=TRUE.}

\item{warn}{Print warning if overwriting file?}
}
\description{
write Stock Synthesis starter file from list object in R which was probably
created using \code{\link{SS_readstarter}}
}
\author{
Ian Taylor
}
\seealso{
\code{\link{SS_readstarter}}, \code{\link{SS_readforecast}},
\code{\link{SS_readctl}}, \code{\link{SS_writestarter}},
\code{\link{SS_writeforecast}}, \code{\link{SS_writedat}},
\code{\link{SS_writectl}}
}

