% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSbootstrap.R
\name{SSbootstrap}
\alias{SSbootstrap}
\title{Fit models to parametric bootstraps}
\usage{
SSbootstrap()
}
\description{
Run a series of models fit to parametric bootstrap data taken from
data.ss_new.  This is not yet a generalized function, just some example code
for how to do a parametric bootstrap such as was done for the Pacific hake
model in 2006.
}
\note{
Thanks to Nancie Cummings for inspiration.
}
\author{
Ian Taylor
}
\references{
\url{http://www.pcouncil.org/wp-content/uploads/2006_hake_assessment_FINAL_ENTIRE.pdf}
(A description is on page 41 and Figures 55-56 (pg 139-140) show some
results.)
}
\keyword{data}
\keyword{manip}

