% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSexecutivesummary.r
\name{SSexecutivesummary}
\alias{SSexecutivesummary}
\title{Create executive summary tables from an SS3 Report.sso file}
\usage{
SSexecutivesummary(
  replist,
  plotfolder = "default",
  ci_value = 0.95,
  es_only = FALSE,
  fleetnames = NULL,
  add_text = "model area",
  so_units = "millions of eggs",
  tables = c("a", "b", "c", "d", "e", "f", "g", "h", "i", "catch", "timeseries",
    "numbers", "biomass", "likes"),
  divide_by_2 = FALSE,
  endyr = NULL,
  adopted_ofl = NULL,
  adopted_abc = NULL,
  adopted_acl = NULL,
  forecast_ofl = NULL,
  forecast_abc = NULL,
  format = TRUE,
  match_digits = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{plotfolder}{Directory where a new \code{tables} directory will be created,
which will be used to store the output from this function. The default is
the dir location where the Report.sso file is located.}

\item{ci_value}{To calculate confidence intervals, the desired interval must
be specified. The default is 0.95.}

\item{es_only}{A logical that specifies if only the executive summary tables
should be produced. The default is \code{FALSE}, which leads to all executive
summary and auxiliary tables being produced (see Return).}

\item{fleetnames}{Optional replacement for fleetnames used in data file.}

\item{add_text}{A single character object, where the default is \code{"model area"}. The text will be added to some of the table captions to indicate
what the results apply to. Besides the default, one could use \code{"base model"}, \code{"sub-area model South of Point Conception."}, etc. Just know
that the text will be appended to \code{"for the"}, and thus, the default text
leads to \code{"for the model area."}. Another thing to note is that a full
stop is not needed but can be used because a full stop is appended to the
end of the caption if it does not already exist.}

\item{so_units}{A single character object specifying the unit of measurement
that spawning output is reported in. The default is "millions of eggs".
This text will be used in the table captions. If fecundity is equal to
weight-at-length, then the units are hard-wired to \code{"mt"} regardless of
what is used within this argument.}

\item{tables}{Which tables to produce, where the default is to save all
possible tables. Note that some of the tables depend on calculations
related to previous tables, so only producing some will fail if requested
on their own (e.g., Table 'f' cannot be created without also creating Table
'a').}

\item{divide_by_2}{A logical allowing the output to be based on single sex
values based on the new sex specification (-1) in SS3 for single sex
models. Default value is \code{FALSE}. \code{TRUE} will lead to dividing values by
2.}

\item{endyr}{Optional input to choose a different ending year for tables,
which could be useful for catch-only updates. The default is \code{NULL}, which
leads to using the ending year defined in Report.sso.}

\item{adopted_ofl, adopted_abc, adopted_acl}{Vectors of adopted overfishing
limits (OFL), acceptable biological catch (ABC), and annual catch limits
(ACL) values to be printed in the management performance table. These
vectors \emph{MUST BE} be vectors of length 10. The default of \code{NULL} leads to
the table being filled in with notes that the values need to be changed.}

\item{forecast_ofl, forecast_abc}{Optional input vectors for management
adopted OFL and ABC values for table g. These values will overwrite the
OFL and ABC values in the projection table, rather than the model
estimated OFL values. As an example, \code{c(1500, 1300)} would be viable
input.}

\item{format}{A logical value that controls if the resulting tables are
formatted (e.g., commas added, confidence/credibility intervals separated
with an en dash, i.e., "--"). The formatting is intended to create tables
that can be cut and pasted easily into a word document without additional
formatting. If the tables are being used by LaTeX/Markdown or other
documenting software, having formatting turned on prevents the tables from
being formatted further since the returned objects are no longer numeric.
The default is \code{TRUE}.}

\item{match_digits}{A logical value specifying if the lower and upper
interval values for derived quantities and recruitment deviations should
be reported with the same number of decimal digits as the estimate.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
Individual .csv files for each executive summary table and additional tables
(catch, timeseries, numbers-at-age).
}
\description{
Take the output from \code{SS_output()} and create executive summary .csv files
as required by the current Terms of Reference for U.S. West Coast
groundfish assessments. Additionally, .csv files of historical catches,
time-series, and numbers-at-age are created.
}
\author{
Chantel Wetzel
}
