% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSexecutivesummary.r
\name{SSexecutivesummary}
\alias{SSexecutivesummary}
\title{A function to create a executive summary tables from an SS Report.sso file}
\usage{
SSexecutivesummary(dir, plotdir = "default", quant = 0.95,
  es.only = FALSE)
}
\arguments{
\item{dir}{Locates the directory of the files to be read in, double
backslashes (or forwardslashes) and quotes necessary.}

\item{plotdir}{Directory where the table will be saved.  The default 
saves the table to the dir location where the Report.sso file is located.}

\item{quant}{to calculate confidence intervals, default is set at 0.95}

\item{es.only}{=  only the executive summary tables will be produced, default is false which
will return all executive summary tables, historical catches, and numbers-at-ages}
}
\value{
A csv files containing executive summary tables.
}
\description{
Reads the Report.sso within the directory and creates executive summary
tables as required by the current Terms of Reference for West Coast 
groundfish.  Works with Stock Synthesis versions 3.24U and later.
Additionally, historical catch and numbers at ages tables are created.
}
\author{
Chantel Wetzel
}
