% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSgetoutput.R
\name{SSgetoutput}
\alias{SSgetoutput}
\title{Get output from multiple Stock Synthesis models.}
\usage{
SSgetoutput(
  keyvec = NULL,
  dirvec = NULL,
  getcovar = TRUE,
  getcomp = TRUE,
  forecast = TRUE,
  verbose = TRUE,
  listlists = lifecycle::deprecated(),
  underscore = FALSE,
  save.lists = FALSE,
  SpawnOutputLabel = "Spawning output",
  modelnames = NULL
)
}
\arguments{
\item{keyvec}{A vector of strings that are appended to the output files from
each model if models are all in one directory. Default=NULL.}

\item{dirvec}{A vector of directories (full path or relative to working
directory) in which model output is located. Default=NULL.}

\item{getcovar}{Choice to read or not read covar.sso output (saves time and
memory). Default=TRUE.}

\item{getcomp}{Choice to read or not read CompReport.sso output (saves time
and memory). Default=TRUE.}

\item{forecast}{Choice to read or not read forecast quantities.
Default=FALSE.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{listlists}{Deprecated argument that wasn't working.}

\item{underscore}{Add an underscore '_' between any file names and any keys
in keyvec. Default=FALSE.}

\item{save.lists}{Save each list of parsed output as a .Rdata file (with default
filenaming convention based on iteration and date stamp.}

\item{SpawnOutputLabel}{An alternative to "Spawning output" for use in
figure axis labels and table headers for models that include a fecundity
relationship. This provides an option to provide the units, e.g.
\code{SpawnOutputLabel = "Spawning output (trillions of eggs)"}.
This needs to be a user input because the units depend on the choice of
fecundity parameters which are calculated outside of the SS3 model.}

\item{modelnames}{Optional vector of model names which will can be passed to
downstream functions like \code{SSsummarize()}, \code{SSplotComparisons()} (where it is used
in the default legend), and \code{SStableComparisons()} (where it is used in the
default column names).}
}
\description{
Apply the function \code{\link[=SS_output]{SS_output()}} multiple times and save output as
individual objects or a list of lists.
}
\examples{
# contrived example where the same model is read twice
mydir <- file.path(
  path.package("r4ss"),
  file.path("extdata", "simple_small")
)
models <- SSgetoutput(
  dirvec = c(mydir, mydir),
  modelnames = c("simple_small", "simple_small_again")
)
}
\seealso{
\code{\link[=SS_output]{SS_output()}}

Other model comparison functions:
\code{\link[=SSplotComparisons]{SSplotComparisons()}},
\code{\link[=SSsummarize]{SSsummarize()}},
\code{\link[=SStableComparisons]{SStableComparisons()}}
}
\author{
Ian Taylor
}
\concept{model comparison functions}
