% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmohnsrho.R
\name{SSmohnsrho}
\alias{SSmohnsrho}
\title{Calculate Mohn's Rho values for select quantities
Function calculates a rho value for the ending year for each retrospective
relative to the reference model and a "Wood's Hole Mohn's Rho" which is based
on all years between the reference and the retrospecitive run.}
\usage{
SSmohnsrho(summaryoutput, endyrvec = NULL, startyr = NULL)
}
\arguments{
\item{summaryoutput}{List created by \code{SSsummarize}. The expected order for the 
models are the full refernce model, the retro -1, retro -2, and so forth.}

\item{endyrvec}{Single year or vector of years representing the
final year of values to show for each model.}

\item{startyr}{Single year used to calculate the start of the Wood's Hole
Mohn's Rho value across all years.}
}
\description{
Calculate Mohn's Rho values for select quantities
Function calculates a rho value for the ending year for each retrospective
relative to the reference model and a "Wood's Hole Mohn's Rho" which is based
on all years between the reference and the retrospecitive run.
}
\author{
Chantel R. Wetzel
}
