% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotCatch.R
\name{SSplotCatch}
\alias{SSplotCatch}
\title{Plot catch related quantities.}
\usage{
SSplotCatch(
  replist,
  subplots = 1:16,
  add = FALSE,
  areas = 1,
  plot = TRUE,
  print = FALSE,
  type = "l",
  fleetlty = 1,
  fleetpch = 1,
  fleetcols = "default",
  fleetnames = replist[["FleetNames"]],
  lwd = 3,
  areacols = NULL,
  areanames = "default",
  minyr = -Inf,
  maxyr = Inf,
  annualcatch = TRUE,
  forecastplot = FALSE,
  plotdir = replist[["inputs"]][["dir"]],
  showlegend = TRUE,
  legendloc = "topleft",
  order = "default",
  xlab = "Year",
  labels = c("Harvest rate/Year", "Continuous F", "Landings", "Total catch",
    "Predicted discards", "Discard fraction", "(mt)", "(numbers x1000)",
    "Observed and expected", "aggregated across seasons"),
  catchasnumbers = NULL,
  catchbars = TRUE,
  addmax = TRUE,
  ymax = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{subplots}{Vector controlling which subplots to create
Numbering of subplots is as follows,

\emph{Basic plots for all models}
\itemize{
\item 1 landings
\item 2 landings stacked
\item 3 observed and expected landings (if different)
\item 9 harvest rate
}

\emph{Plots for models with discards}
\itemize{
\item 4 total catch (including discards)
\item 5 total catch (including discards) stacked
\item 6 discards
\item 7 discards stacked plot (depends on multiple fleets)
\item 8 discard fraction
\item 16 landings + dead discards
}

\emph{Plots for seasonal models}
\itemize{
\item 10 landings aggregated across seasons
\item 11 landings aggregated across seasons stacked
\item 12 total catch (if discards present) aggregated across seasons
\item 13 total catch (if discards present) aggregated across seasons stacked
\item 14 discards aggregated across seasons
\item 15 discards aggregated across seasons stacked
}}

\item{add}{Add to existing plot? (not yet implemented)}

\item{areas}{Optional subset of areas to plot for spatial models}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{type}{Type parameter passed to plot function. Default "l" is lines
only.  Other options include "o" for overplotting points on lines.}

\item{fleetlty}{Vector of line type by fleet}

\item{fleetpch}{Vector of plot character by fleet}

\item{fleetcols}{Vector of colors by fleet}

\item{fleetnames}{Optional replacement for fleetnames used in data file.}

\item{lwd}{Line width for plot elements.}

\item{areacols}{Optional vector of colors for each area if model has
multiple areas. NULL value will be replaced by a default set of areas.}

\item{areanames}{Names for areas. Default is to use Area1, Area2,...}

\item{minyr}{Optional input for minimum year to show in plots}

\item{maxyr}{Optional input for maximum year to show in plots}

\item{annualcatch}{Include plot of catch aggregated across seasons within
each year}

\item{forecastplot}{Add points from forecast years}

\item{plotdir}{Directory where PNG files will be written.}

\item{showlegend}{Put legend on plot}

\item{legendloc}{Location of legend. Either a string like "topleft" or a
vector of two numeric values representing the fraction of the maximum in
the x and y dimensions, respectively. See \code{help("legend")} for more info on
the string options.}

\item{order}{Optional input to change the order of fleets in stacked plots.}

\item{xlab}{x-label for all plots}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{catchasnumbers}{Is catch in numbers instead of biomass? Should be set
automatically if set to NULL. If fleets include a mix of biomass and
numbers, then catch plots should be interpreted carefully.}

\item{catchbars}{Show catch by fleet as barplot instead of stacked polygons?
(default=TRUE)}

\item{addmax}{Add a point on the y-axis for the maximum catch (default=TRUE)}

\item{ymax}{Optional input for ymax value (can be used to add or subtract
white space at the top of the figure)}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Plot catch related quantities from Stock Synthesis output. Plots include
harvest rate, continuous F, landings, and discard fraction.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SS_output]{SS_output()}}
}
\author{
Ian Taylor, Ian Stewart
}
