% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotMCMC_ExtraSelex.R
\name{SSplotMCMC_ExtraSelex}
\alias{SSplotMCMC_ExtraSelex}
\title{Plot uncertainty around chosen selectivity ogive from MCMC.}
\usage{
SSplotMCMC_ExtraSelex(
  post,
  add = FALSE,
  nsexes = 1,
  shift = 0,
  fleetname = "default",
  col = "blue"
)
}
\arguments{
\item{post}{A data frame containing either derived_posteriors.sso or a good
subset of it. This can be an element of the list created by the the
\code{\link{SSgetMCMC}} function.}

\item{add}{TRUE/FALSE option to add results to an existing plot.}

\item{nsexes}{Number of sexes in the model (should match model values but is
only used in the title).}

\item{shift}{Optional adjustment to the x values to avoid overlap of
intervals when overplotting on an existing plot.}

\item{fleetname}{Optional input to make the title better. Default will be
something like "Fleet 1", using the numbering from the model.}

\item{col}{Color for points and lines.}
}
\description{
Plot uncertainty in selectivity from an MCMC output for whichever fleet/year
was chosen in the optional extra "more stddev reporting"
}
\author{
Ian Taylor
}
