% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotMovementRates.R
\name{SSplotMovementRates}
\alias{SSplotMovementRates}
\title{Plot movement rates from model output}
\usage{
SSplotMovementRates(
  replist,
  plot = TRUE,
  print = FALSE,
  subplots = 1:2,
  plotdir = replist[["inputs"]][["dir"]],
  colvec = "default",
  ylim = "default",
  legend = TRUE,
  legendloc = "topleft",
  moveseas = "all",
  min.move.age = 0.5,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{subplots}{which subplots to create.}

\item{plotdir}{Directory where PNG files will be written.}

\item{colvec}{vector of colors for each movement rate in the plot}

\item{ylim}{optional input for y range of the plot. By default plot ranges
from 0 to 10\% above highest movement rate (not including fish staying in an
area).}

\item{legend}{add a legend designating which color goes with which pair of
areas?}

\item{legendloc}{Location of legend. Either a string like "topleft" or a
vector of two numeric values representing the fraction of the maximum in
the x and y dimensions, respectively. See \code{help("legend")} for more info
on the string options.}

\item{moveseas}{choice of season for which movement rates are shown}

\item{min.move.age}{Minimum age of movement (in future will come from Report file)}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Plots estimated movement rates in final year for each area/season with movement as
reported in Report.sso. If movement is time-varying, an additional figure shows
pattern across years (if the MGparm_By_Year_after_adjustments table (report:7)
is available in the Report.sso file)
}
\examples{
\dontrun{
SSplotMovementRates(myreplist)
}

}
\seealso{
\code{\link[=SS_output]{SS_output()}}, \code{\link[=SSplotMovementRates]{SSplotMovementRates()}},
}
\author{
Ian Taylor
}
