% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotProfile.R
\name{SSplotProfile}
\alias{SSplotProfile}
\title{Plot likelihood profile results}
\usage{
SSplotProfile(
  summaryoutput,
  plot = TRUE,
  print = FALSE,
  models = "all",
  profile.string = "steep",
  profile.label = NULL,
  exact = FALSE,
  ylab = "Change in -log-likelihood",
  components = c("TOTAL", "Catch", "Equil_catch", "Survey", "Discard", "Mean_body_wt",
    "Length_comp", "Age_comp", "Size_at_age", "SizeFreq", "Morphcomp", "Tag_comp",
    "Tag_negbin", "Recruitment", "InitEQ_Regime", "Forecast_Recruitment", "Parm_priors",
    "Parm_softbounds", "Parm_devs", "F_Ballpark", "Crash_Pen"),
  component.labels = c("Total", "Catch", "Equilibrium catch", "Index data", "Discard",
    "Mean body weight", "Length data", "Age data", "Size-at-age data",
    "Generalized size data", "Morph composition data", "Tag recapture distribution",
    "Tag recapture total", "Recruitment", "Initital equilibrium recruitment",
    "Forecast recruitment", "Priors", "Soft bounds", "Parameter deviations",
    "F Ballpark", "Crash penalty"),
  minfraction = 0.01,
  sort.by.max.change = TRUE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lty.total = 1,
  lwd = 2,
  lwd.total = 3,
  cex = 1,
  cex.total = 1.5,
  xlim = NULL,
  ymax = NULL,
  xaxs = "r",
  yaxs = "r",
  type = "o",
  legend = TRUE,
  legendloc = "topright",
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  add_cutoff = FALSE,
  cutoff_prob = 0.95,
  add_no_prior_line = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{summaryoutput}{List created by the function \code{\link[=SSsummarize]{SSsummarize()}}.}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{models}{Optional subset of the models described in
\code{summaryoutput}. Either "all" or a vector of numbers indicating
columns in summary tables.}

\item{profile.string}{Character string used to find parameter over which the
profile was conducted. If \code{exact=FALSE}, this can be a substring of
one of the SS3 parameter labels found in the Report.sso file.
For instance, the default input 'steep'
matches the parameter 'SR_BH_steep'. If \code{exact=TRUE}, then
profile.string needs to be an exact match to the parameter label.}

\item{profile.label}{Label for x-axis describing the parameter over which
the profile was conducted. NULL value will be replaced by an informative
label if the parameter label contains one of the follow strings:
"steep", "R0", "NatM", "L_at_Amax", "sigmaR", or "LnQ".}

\item{exact}{Should the \code{profile.string} have to match the parameter
label exactly, or is a substring OK.}

\item{ylab}{Label for y-axis. Default is "Change in -log-likelihood".}

\item{components}{Vector of likelihood components that may be included in
plot. List is further refined by any components that are not present in
model or have little change over range of profile (based on limit
\code{minfraction}). Hopefully this doesn't need to be changed.}

\item{component.labels}{Vector of labels for use in the legend that matches
the vector in \code{components}.}

\item{minfraction}{Minimum change in likelihood (over range considered) as a
fraction of change in total likelihood for a component to be included in the
figure.}

\item{sort.by.max.change}{Switch giving option to sort components in legend
in order of maximum amount of change in likelihood (over range considered).
Default=TRUE.}

\item{col}{Optional vector of colors for each line.}

\item{pch}{Optional vector of plot characters for the points.}

\item{lty}{Line type for the likelihood components.}

\item{lty.total}{Line type for the total likelihood.}

\item{lwd}{Line width for the likelihood components.}

\item{lwd.total}{Line width for the total likelihood.}

\item{cex}{Character expansion for the points representing the likelihood
components.}

\item{cex.total}{Character expansion for the points representing the total
likelihood.}

\item{xlim}{Range for x-axis. Change in likelihood is calculated relative to
values within this range.}

\item{ymax}{Maximum y-value. Default is 10\% greater than largest value
plotted.}

\item{xaxs}{The style of axis interval calculation to be used for the x-axis
(see ?par for more info)}

\item{yaxs}{The style of axis interval calculation to be used for the y-axis
(see ?par for more info).}

\item{type}{Line type (see ?plot for more info).}

\item{legend}{Add a legend?}

\item{legendloc}{Location of legend. Either a string like "topleft" or a
vector of two numeric values representing the fraction of the maximum in
the x and y dimensions, respectively. See \code{help("legend")} for more info
on the string options.}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{plotdir}{Directory where PNG files will be written.}

\item{add_cutoff}{Add dashed line at ~1.92 to indicate 95\% confidence interval
based on common cutoff of half of chi-squared of p=.95 with 1 degree of
freedom: \code{0.5*qchisq(p=cutoff_prob, df=1)}. The probability value
can be adjusted using the \code{cutoff_prob} below.}

\item{cutoff_prob}{Probability associated with \code{add_cutoff} above.}

\item{add_no_prior_line}{Add line showing total likelihood without
the prior (only appears when profiled parameter that includes a prior)}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{\dots}{Additional arguments passed to the \code{plot} command.}
}
\description{
Makes a plot of change in negative-log-likelihood for each likelihood
component that contributes more than some minimum fraction of change in
total.
}
\note{
Someday the function \code{\link[=profile]{profile()}} will be improved and
made to work directly with this plotting function, but they don't yet work
well together. Thus, even if \code{\link[=profile]{profile()}} is used, the output
should be read using \code{\link[=SSgetoutput]{SSgetoutput()}} or by multiple calls to
\code{\link[=SS_output]{SS_output()}}.
}
\seealso{
\code{\link[=SSsummarize]{SSsummarize()}}, \code{\link[=SSgetoutput]{SSgetoutput()}}

Other profile functions:
\code{\link[=PinerPlot]{PinerPlot()}},
\code{\link[=profile]{profile()}}
}
\author{
Ian G. Taylor, Ian J. Stewart
}
\concept{profile functions}
