% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRecdevs.R
\name{SSplotRecdevs}
\alias{SSplotRecdevs}
\title{Plot recruitment deviations}
\usage{
SSplotRecdevs(
  replist,
  subplots = 1:3,
  plot = TRUE,
  print = FALSE,
  add = FALSE,
  uncertainty = TRUE,
  minyr = -Inf,
  maxyr = Inf,
  forecastplot = FALSE,
  col1 = "black",
  col2 = "blue",
  col3 = "green3",
  col4 = "red",
  legendloc = "topleft",
  labels = c("Year", "Asymptotic standard error estimate", "Log recruitment deviation",
    "Bias adjustment fraction, 1 - stddev^2 / sigmaR^2"),
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = replist[["inputs"]][["dir"]],
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{subplots}{vector controlling which subplots to create}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{add}{add to existing plot (not yet implemented)}

\item{uncertainty}{include plots showing uncertainty?}

\item{minyr}{optional input for minimum year to show in plots}

\item{maxyr}{optional input for maximum year to show in plots}

\item{forecastplot}{include points from forecast years?}

\item{col1}{first color used}

\item{col2}{second color used}

\item{col3}{third color used}

\item{col4}{fourth color used}

\item{legendloc}{Location of legend. Either a string like "topleft" or a
vector of two numeric values representing the fraction of the maximum in
the x and y dimensions, respectively. See \code{help("legend")} for more info
on the string options.}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{plotdir}{Directory where PNG files will be written.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Plot recruitment deviations and associated quantities including derived
measures related to bias adjustment.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SS_fitbiasramp]{SS_fitbiasramp()}}
}
\author{
Ian Taylor, Ian Stewart
}
