% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRecdist.R
\name{SSplotRecdist}
\alias{SSplotRecdist}
\title{Plot of recruitment distribution among areas and seasons}
\usage{
SSplotRecdist(
  replist,
  plot = TRUE,
  print = FALSE,
  areanames = NULL,
  seasnames = NULL,
  xlab = "",
  ylab = "",
  main = "distribution of recruitment by area and season",
  period = c("Initial", "Benchmark", "End year"),
  sexes = 1:2,
  plotdir = "default",
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{areanames}{optional vector to replace c("Area1","Area2",...)}

\item{seasnames}{optional vector to replace c("Season1","Season2",...)}

\item{xlab}{optional x-axis label (if the area names aren\'t informative
enough)}

\item{ylab}{optional y-axis label (if the season names aren\'t informative
enough)}

\item{main}{title for plot}

\item{period}{period of recruitment distribution to show among the options
"Initial", "Benchmark", and "End year"}

\item{sexes}{either 1 to only plot female distribution, 2 for males, or 1:2
to make both plots}

\item{plotdir}{Directory where PNG files will be written.}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Image plot shows fraction of recruitment in each combination of area and
season. This is based on the RECRUITMENT_DIST section of the Report.sso
file.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SSplotRecdevs]{SSplotRecdevs()}}
}
\author{
Ian Taylor
}
