% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotSPR.R
\name{SSplotSPR}
\alias{SSplotSPR}
\title{Plot Spawning Potential Ratio (SPR) quantities.}
\usage{
SSplotSPR(
  replist,
  add = FALSE,
  plot = TRUE,
  print = FALSE,
  uncertainty = TRUE,
  subplots = 1:4,
  forecastplot = FALSE,
  col1 = "black",
  col2 = "blue",
  col3 = "green3",
  col4 = "red",
  sprtarg = replist[["sprtarg"]],
  btarg = replist[["btarg"]],
  minbthresh = replist[["minbthresh"]],
  labels = c("Year", "SPR", "1-SPR", "Relative fishing intensity",
    "Fraction of unfished spawning output"),
  pwidth = 6.5,
  pheight = 5,
  pheight_tall = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = replist[["inputs"]][["dir"]],
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{add}{add to existing plot (not yet implemented)}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{uncertainty}{include plots showing uncertainty?}

\item{subplots}{vector controlling which subplots to create
Numbering of subplots is as follows:
\enumerate{
\item timeseries of SPR,
\item timeseries of 1 - SPR,
\item timeseries of SPR ratio (as specified in the starter file), and
\item phase plot of Biomass ratio vs SPR ratio (as specified in the
starter file).
}}

\item{forecastplot}{Include forecast years in plot?}

\item{col1}{first color used}

\item{col2}{second color used}

\item{col3}{third color used}

\item{col4}{fourth color used}

\item{sprtarg}{F/SPR proxy target. "default" chooses based on model output,
where models which have SPR_std_basis = 0 or 1 specified in the starter
file will use the SPR target specified in the forecast file. Models which
have SPR_std_basis = 2 will use SPR at MSY for the SPR target
and models which have the SPR_std_basis = 3 will use
SPR at Btarget for the SPR target in these plots. Zero or negative values of
sprtarg input here will cause no horizontal line to be plotted.}

\item{btarg}{target depletion to be used in plots showing depletion. May be
omitted by setting to NA. "default" chooses based on model output.}

\item{minbthresh}{minimum biomass threshold to be used in plots
showing depletion. May be omitted by setting to NA. "default" chooses
based on model output.}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{pheight_tall}{Height of tall plots printed to png files in units of
\code{punits}, where the tall plots are a subset of the plots which typically
work best in a taller format.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{plotdir}{Directory where PNG files will be written.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Plot time series of SPR, 1-SPR, the chosen SPR ratio and the phase plot.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SS_output]{SS_output()}}
}
\author{
Ian Stewart, Ian Taylor
}
