% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotSummaryF.R
\name{SSplotSummaryF}
\alias{SSplotSummaryF}
\title{Plot the summary F (or harvest rate).}
\usage{
SSplotSummaryF(
  replist,
  yrs = "all",
  Ftgt = NA,
  ylab = "Summary Fishing Mortality",
  plot = TRUE,
  print = FALSE,
  plotdir = "default",
  verbose = TRUE,
  uncertainty = TRUE,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10
)
}
\arguments{
\item{replist}{List created by \code{SS_output}}

\item{yrs}{Which years to include.}

\item{Ftgt}{Target F where horizontal line is shown.}

\item{ylab}{Y-axis label.}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{plotdir}{Directory where PNG files will be written. By default it will
be the directory where the model was run.}

\item{verbose}{Verbose output to R console?}

\item{uncertainty}{Show 95\% uncertainty intervals around point estimates?}

\item{pwidth}{width of plot}

\item{pheight}{height of plot}

\item{punits}{units for PNG file}

\item{res}{resolution for PNG file}

\item{ptsize}{point size for PNG file}
}
\description{
Plots the summary F (or harvest rate) as set up in the starter file Needs a
lot of work to be generalized
}
\seealso{
\code{\link{SSplotTimeseries}}, ~~~
}
\author{
Allan Hicks
}
