% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotTimeseries.R
\name{SSplotTimeseries}
\alias{SSplotTimeseries}
\title{Plot timeseries data}
\usage{
SSplotTimeseries(
  replist,
  subplot,
  add = FALSE,
  areas = "all",
  areacols = NULL,
  areanames = "default",
  forecastplot = TRUE,
  uncertainty = TRUE,
  bioscale = 1,
  minyr = -Inf,
  maxyr = Inf,
  plot = TRUE,
  print = FALSE,
  plotdir = replist[["inputs"]][["dir"]],
  verbose = TRUE,
  btarg = replist[["btarg"]],
  minbthresh = replist[["minbthresh"]],
  xlab = "Year",
  labels = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  mainTitle = FALSE,
  mar = NULL
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{subplot}{number controlling which subplot to create
Numbering of subplots is as follows, where the spawning biomass plots
(7 to 10) are provided first when this function is called by \code{\link[=SS_plots]{SS_plots()}}:
\itemize{
\item 1 Total biomass (t) with forecast
\item 2 Total biomass by area (spatial models only)
\item 3 Total biomass (t) at beginning of spawning season with forecast
\item 4 Summary biomass (t) with forecast
\item 5 Summary biomass (t) by area (spatial models only)
\item 6 Summary biomass (t) at beginning of season 1 with forecast
\item 7 Spawning output with forecast with ~95\% asymptotic intervals
\item 8 Spawning output by area (spatial models only)
\item 9 Fraction of unfished spawning output with forecast with ~95\% asymptotic intervals
\item 10 Fraction of unfished spawning output by area (spatial models only)
\item 11 Age-0 recruits (1,000s) with forecast with ~95\% asymptotic intervals
\item 12 Age-0 recruits by area (spatial models only)
\item 13 Fraction of recruits by area (spatial models only)
\item 14 Age-0 recruits (1,000s) by birth season with forecast
\item 15 Fraction of total Age-0 recruits by birth season with forecast
}}

\item{add}{add to existing plot? (not yet implemented)}

\item{areas}{optional subset of areas to plot for spatial models}

\item{areacols}{Optional vector of colors for each area if model has
multiple areas. NULL value will be replaced by a default set of areas.}

\item{areanames}{names for areas. Default is to use Area1, Area2,...}

\item{forecastplot}{add points from forecast years}

\item{uncertainty}{add intervals around quantities for which uncertainty is
available}

\item{bioscale}{scaling for spawning biomass. Default = 1.
Previously this was set to 0.5 for single-sex models, and 1.0 for all
others, but now single-sex models are assumed to use the -1 option for
Nsexes in the data file so the scaling is done automatically by SS3.}

\item{minyr}{optional input for minimum year to show in plots}

\item{maxyr}{optional input for maximum year to show in plots}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{plotdir}{Directory where PNG files will be written.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}

\item{btarg}{Target depletion to be used in plots showing depletion. May be
omitted by setting to 0. "default" chooses value based on modeloutput.}

\item{minbthresh}{Threshold depletion to be used in plots showing depletion.
May be omitted by setting to 0. "default" assumes 0.25 unless btarg in model
output is 0.25 in which case minbthresh = 0.125 (U.S. west coast flatfish).}

\item{xlab}{x axis label for all plots}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{mainTitle}{Logical indicating if a title should be included at the top
(not yet implemented for all plots).}

\item{mar}{Either NULL to allow the default (which depends on whether the
main title is included or not) or a numerical vector of the form
c(bottom, left, top, right) which gives the number of lines of margin to
be specified on the four sides of the plot, which is passed to \code{par()}.}
}
\description{
Plot timeseries data contained in TIME_SERIES output from Stock Synthesis
report file. Some values have optional uncertainty intervals.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SS_output]{SS_output()}}
}
\author{
Ian Taylor, Ian Stewart
}
