% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotTimeseries.R
\name{SSplotTimeseries}
\alias{SSplotTimeseries}
\title{Plot timeseries data}
\usage{
SSplotTimeseries(
  replist,
  subplot,
  add = FALSE,
  areas = "all",
  areacols = "default",
  areanames = "default",
  forecastplot = TRUE,
  uncertainty = TRUE,
  bioscale = 1,
  minyr = -Inf,
  maxyr = Inf,
  plot = TRUE,
  print = FALSE,
  plotdir = "default",
  verbose = TRUE,
  btarg = "default",
  minbthresh = "default",
  xlab = "Year",
  labels = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{subplot}{number controlling which subplot to create
Numbering of subplots is as follows:
\itemize{
\item 1: Total biomass (mt) with forecast
\item 2: Total biomass by area (spatial models only)
\item 3: Total biomass (mt) at beginning of spawning season with forecast
\item 4: Summary biomass (mt) with forecast
\item 5: Summary biomass (mt) by area (spatial models only)
\item 6: Summary biomass (mt) at beginning of season 1 with forecast
\item 7: Spawning output with forecast with ~95\% asymptotic intervals
\item 8: Spawning output by area (spatial models only)
\item 9: Relative spawning output with forecast with ~95\% asymptotic intervals
\item 10: Relative spawning output by area (spatial models only)
\item 11: Age-0 recruits (1,000s) with forecast with ~95\% asymptotic intervals
\item 12: Age-0 recruits by area (spatial models only)
\item 13: Fraction of recruits by area (spatial models only)
\item 14: Age-0 recruits (1,000s) by birth season with forecast
\item 15: Fraction of total Age-0 recruits by birth season with forecast
}}

\item{add}{add to existing plot? (not yet implemented)}

\item{areas}{optional subset of areas to plot for spatial models}

\item{areacols}{vector of colors by area. Default uses rich.colors by Arni
Magnusson}

\item{areanames}{names for areas. Default is to use Area1, Area2,...}

\item{forecastplot}{add points from forecast years}

\item{uncertainty}{add intervals around quantities for which uncertainty is
available}

\item{bioscale}{scaling for spawning biomass. Default = 1.
Previously this was set to
0.5 for single-sex models, and 1.0 for all others, but now single-sex
models are assumed to use the -1 option for Nsexes in the data file so the
scaling is done automatically by SS.}

\item{minyr}{optional input for minimum year to show in plots}

\item{maxyr}{optional input for maximum year to show in plots}

\item{plot}{plot to active plot device?}

\item{print}{print to PNG files?}

\item{plotdir}{directory where PNG or PDF files will be written. by default
it will be the directory where the model was run.}

\item{verbose}{report progress to R GUI?}

\item{btarg}{Target depletion to be used in plots showing depletion. May be
omitted by setting to 0. "default" chooses value based on modeloutput.}

\item{minbthresh}{Threshold depletion to be used in plots showing depletion.
May be omitted by setting to 0. "default" assumes 0.25 unless btarg in model
output is 0.25 in which case minbthresh = 0.125 (U.S. west coast flatfish).}

\item{xlab}{x axis label for all plots}

\item{labels}{vector of labels for plots (titles and axis labels)}

\item{pwidth}{width of plot}

\item{pheight}{height of plot}

\item{punits}{units for PNG file}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{point size for PNG file}

\item{cex.main}{character expansion for plot titles}
}
\description{
Plot timeseries data contained in TIME_SERIES output from Stock Synthesis
report file. Some values have optional uncertainty intervals.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SS_output]{SS_output()}}
}
\author{
Ian Taylor, Ian Stewart
}
