% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotYield.R
\name{SSplotYield}
\alias{SSplotYield}
\title{Plot yield and surplus production.}
\usage{
SSplotYield(
  replist,
  subplots = 1:5,
  refpoints = c("MSY", "Btgt", "SPR", "Current"),
  add = FALSE,
  plot = TRUE,
  print = FALSE,
  labels = c("Fraction of unfished spawning biomass", "Equilibrium yield (t)",
    "Total biomass (t)", "Surplus production (t)", "Yield per recruit (kg)",
    "Spawning output"),
  col = "blue",
  col2 = "black",
  lty = 1,
  lwd = 2,
  cex.main = 1,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  plotdir = replist[["inputs"]][["dir"]],
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{subplots}{vector controlling which subplots to create
Numbering of subplots is as follows:
\itemize{
\item 1 yield curve
\item 2 yield curve with reference points
\item 3 surplus production vs. total biomass plots (Walters et al. 2008)
\item 4 surplus production vs. spawning biomass plots (Forrest et al. 2023)
\item 5 yield per recruit plot
}}

\item{refpoints}{character vector of which reference points to display in
subplot 2, from the options 'MSY', 'Btgt', and 'SPR'.}

\item{add}{add to existing plot? (not yet implemented)}

\item{plot}{Plot to active plot device?}

\item{print}{Print to PNG files?}

\item{labels}{Vector of labels for plots (titles and axis labels).}

\item{col}{line color for equilibrium plot}

\item{col2}{line color for dynamic surplus production plot}

\item{lty}{line type (only applied to equilibrium yield plot at this time)}

\item{lwd}{Line width for plot elements.}

\item{cex.main}{Character expansion for plot titles.
The default is \code{cex.main=1}.}

\item{pwidth}{Default width of plots printed to files in units of
\code{punits}.}

\item{pheight}{Height of plots printed to png files in units of \code{punits}.
Default is designed to allow two plots per page, with \code{pheight_tall} used
for plots that work best with a taller format and a single plot per page.}

\item{punits}{Units for \code{pwidth} and \code{pheight}. Can be "px"
(pixels), "in" (inches), "cm" (centimeters), or "mm" (millimeters).
The default is \code{punits="in"}.}

\item{res}{Resolution of plots printed to files.
The default is \code{res = 300}.}

\item{ptsize}{Point size for plotted text in plots printed to files (see
\code{help("png")} in R for details).}

\item{plotdir}{Directory where PNG files will be written.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Plot yield and surplus production from Stock Synthesis output. Surplus
production is based on Walters et al. (2008).
}
\references{
Walters, Hilborn, and Christensen, 2008, Surplus production
dynamics in declining and recovering fish populations. \emph{Can. J. Fish. Aquat.
Sci.} 65: 2536-2551. \url{https://doi.org/10.1139/F08-170}.

Forrest, Kronlund, Cleary, and Grinnell. 2023. An
evidence-based approach for selecting a limit reference point for Pacific
herring (\emph{Clupea pallasii}) stocks in British Columbia, Canada. \emph{Can. J. Fish.
Aquat. Sci.} 80: 1071-1083. \url{https://doi.org/10.1139/cjfas-2022-0168}.
}
\seealso{
\code{\link[=SS_plots]{SS_plots()}}, \code{\link[=SS_output]{SS_output()}}
}
\author{
Ian Stewart, Ian Taylor
}
