% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_all.R
\name{table_all}
\alias{table_all}
\title{Run all the tables functions}
\usage{
table_all(
  replist,
  dir = replist[["inputs"]][["dir"]],
  fleetnames = NULL,
  selexyr = NULL,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{fleetnames}{String of fleet names. Default is NULL which will use the
the model fleet names.}

\item{selexyr}{The year to summarize selectivity, the default is the final
model year.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Not necessarily a good idea but helpful for testing
}
\seealso{
Other table functions:
\code{\link[=table_biology]{table_biology()}},
\code{\link[=table_compweight]{table_compweight()}},
\code{\link[=table_config]{table_config()}},
\code{\link[=table_exec_summary]{table_exec_summary()}},
\code{\link[=table_parcounts]{table_parcounts()}},
\code{\link[=table_pars]{table_pars()}},
\code{\link[=table_ts]{table_ts()}}
}
\author{
Ian G. Taylor
}
\concept{table functions}
