% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_biology.R
\name{table_biology}
\alias{table_biology}
\title{Create a table of biology for assessment reporting:
length, weight, \% mature, fecundity, and selectivity}
\usage{
table_biology(
  replist,
  dir = replist[["inputs"]][["dir"]],
  fleetnames = replist[["FleetNames"]],
  selexyr = NULL,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{fleetnames}{String of fleet names. Default is NULL which will use the
the model fleet names.}

\item{selexyr}{The year to summarize selectivity, NULL will use the most
recent year with selectivity reported in the model output.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
Individual .rda files containing a list of table and caption
}
\description{
Takes the object created by \code{\link[=SS_output]{SS_output()}} to create table for reporting
biology at age. This function was formerly known as \code{SStablebiology()}
but that older version is now deprecated.
}
\seealso{
Other table functions:
\code{\link[=table_all]{table_all()}},
\code{\link[=table_compweight]{table_compweight()}},
\code{\link[=table_config]{table_config()}},
\code{\link[=table_exec_summary]{table_exec_summary()}},
\code{\link[=table_parcounts]{table_parcounts()}},
\code{\link[=table_pars]{table_pars()}},
\code{\link[=table_ts]{table_ts()}}
}
\author{
Chantel R. Wetzel
}
\concept{table functions}
