% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_biology.R
\name{table_biology}
\alias{table_biology}
\title{Create a table of biology for assessment reporting:
length, weight, \% mature, fecundity, and selectivity}
\usage{
table_biology(
  replist,
  dir = NULL,
  fleetnames = NULL,
  selexyr = NULL,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{fleetnames}{Optional replacement for fleetnames used in data file.}

\item{selexyr}{The year to summarize selectivity, the default is the final
model year.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
Individual .rda files containing a list of table and caption
}
\description{
Takes the object created by \code{\link[=SS_output]{SS_output()}} to create table for reporting
biology at age. This function was formerly known as \code{SStablebiology()}
but that older version is now deprecated.
}
\seealso{
Other table functions: 
\code{\link{table_compweight}()},
\code{\link{table_config}()},
\code{\link{table_exec_summary}()},
\code{\link{table_parcounts}()},
\code{\link{table_pars}()},
\code{\link{table_ts}()}
}
\author{
Chantel R. Wetzel
}
\concept{table functions}
