% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_compweight.R
\name{table_compweight}
\alias{table_compweight}
\title{Table of length and age comp Francis weights
(doesn't yet include other data types)}
\usage{
table_compweight(
  output,
  caption = paste("Data weightings applied to length and age compositions",
    "according to the `Francis' method. `Obs.' refers to the number of unique",
    "composition vectors included in the likelihood. `N input' and `N adj.'",
    "refer to the sample sizes of those vectors before and after being adjusted",
    "by the the weights."),
  caption_CAAL = "`CAAL' is conditional age-at-length data.",
  caption_extra = "",
  label = "table-compweight-base"
)
}
\arguments{
\item{output}{A list from \link{SS_output}.}

\item{caption}{A character string for the caption.}

\item{caption_extra}{Additional model-specific information pasted
at the end of the caption}

\item{label}{A character string with the label for the table.
No underscores allowed.}
}
\description{
Table of length and age comp Francis weights
(doesn't yet include other data types)
}
\author{
Kelli F. Johnson, Ian G. Taylor
}
