% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_config.R
\name{table_config}
\alias{table_config}
\title{Create a table to summarize the configuration of the SS3 model}
\usage{
table_config(replist, dir = NULL, verbose = TRUE)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
Returns invisibly and optionally saves an .rda files containing
a list of table and caption
}
\description{
Create a table to summarize the configuration of the SS3 model
}
\examples{
\dontrun{
# Load the model output
output <- r4ss::SS_output()
# Create the table
table_config(output)

# compare configuration for multiple models
config1 <- table_config(output1)
config2 <- table_config(output2)
table_compare <- data.frame(
  new_model = config1$table,
  old_model = config2$table$Configuration
)
names(table_compare) <- c("Section", "New model", "Old model")
}
}
\seealso{
Other table functions: 
\code{\link{table_biology}()},
\code{\link{table_compweight}()},
\code{\link{table_exec_summary}()},
\code{\link{table_parcounts}()},
\code{\link{table_pars}()},
\code{\link{table_ts}()}
}
\author{
Kelli F. Johnson, Ian G. Taylor
}
\concept{table functions}
