% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_exec_summary.R
\name{table_exec_summary}
\alias{table_exec_summary}
\title{Create executive summary tables from an SS3 Report.sso file}
\usage{
table_exec_summary(
  replist,
  dir = replist[["inputs"]][["dir"]],
  ci_value = 0.95,
  fleetnames = replist[["FleetNames"]],
  so_units = "biomass (mt)",
  endyr = NULL,
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{ci_value}{To calculate confidence intervals, the desired interval must
be specified. The default is 0.95.}

\item{fleetnames}{String of fleet names. Default is NULL which will use the
the model fleet names.}

\item{so_units}{A single character object specifying the unit of measurement
that spawning output is reported in. The default is "millions of eggs".
This text will be used in the table captions. If fecundity is equal to
weight-at-length, then the units are hard-wired to \code{"mt"} regardless of
what is used within this argument.}

\item{endyr}{Optional input to choose a different ending year for tables,
which could be useful for catch-only updates. The default is \code{NULL}, which
leads to using the ending year defined in Report.sso.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\value{
Individual .rda files containing a list of table and caption
}
\description{
Take the output from \code{SS_output()} and create executive summary .rda files
as required by the current Terms of Reference for U.S. West Coast
groundfish assessments. Additionally, .rda files of historical catches,
time-series, and numbers-at-age are created. A CSV file with captions is
also created. This function is modified from \code{SSexecutivesummary()}
associated with the adoption of the {asar} template.
}
\examples{
\dontrun{
### running function
# read model output
model <- SS_output(...)
# run this function
tables <- table_exec_summary(model)
# load an individual table (list added to workspace includes
# the table and caption)
load(file.path(model[["inputs"]][["dir"]], "tables", "projections.rda"))

### modify projections table for model that falls below target
# get table from list saved above
tab <- tables[["projections"]][["table"]]
# get buffers from forecast file
buffers <- inputs[["fore"]][["Flimitfraction_m"]]
# which rows in the table are missing buffer and ABC values?
NA_rows <- is.na(tab[["Buffer"]])
# fill in Buffer column from forecast file values
tab[["Buffer"]][NA_rows] <- buffers[["fraction"]][match(tab[["Year"]][NA_rows], buffers[["year"]])]
# fill in ABC as product of Buffer and OFL
tab[NA_rows, "ABC (mt)"] <- tab[["Buffer"]][NA_rows] * tab$"OFL (mt)"[NA_rows]
}
}
\seealso{
Other table functions:
\code{\link[=table_all]{table_all()}},
\code{\link[=table_biology]{table_biology()}},
\code{\link[=table_compweight]{table_compweight()}},
\code{\link[=table_config]{table_config()}},
\code{\link[=table_parcounts]{table_parcounts()}},
\code{\link[=table_pars]{table_pars()}},
\code{\link[=table_ts]{table_ts()}}
}
\author{
Chantel R. Wetzel, Kelli F. Johnson, Ian G. Taylor
}
\concept{table functions}
