% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_exec_summary.R
\name{table_exec_summary}
\alias{table_exec_summary}
\title{Create executive summary tables from an SS3 Report.sso file}
\usage{
table_exec_summary(
  replist,
  dir = NULL,
  ci_value = 0.95,
  fleetnames = NULL,
  so_units = "biomass (mt)",
  divide_by_2 = FALSE,
  endyr = NULL,
  verbose = TRUE
)
}
\arguments{
\item{replist}{List created by \code{r4ss::SS_output()}}

\item{dir}{Directory where the .rds files will be written. The default is
the dir location where the Report.sso file is located.}

\item{ci_value}{To calculate confidence intervals, the desired interval must
be specified. The default is 0.95.}

\item{fleetnames}{String of fleetnames. Default is NULL which will use the
the model fleet names.}

\item{so_units}{A single character object specifying the unit of measurement
that spawning output is reported in. The default is "millions of eggs".
This text will be used in the table captions. If fecundity is equal to
weight-at-length, then the units are hard-wired to \code{"mt"} regardless of
what is used within this argument.}

\item{divide_by_2}{A logical allowing the output to be based on single sex
values based on the new sex specification (-1) in SS3 for single sex
models. Default value is \code{FALSE}. \code{TRUE} will lead to dividing values by
2.}

\item{endyr}{Optional input to choose a different ending year for tables,
which could be useful for catch-only updates. The default is \code{NULL}, which
leads to using the ending year defined in Report.sso.}

\item{verbose}{Print messages to the screen.}

\item{add_text}{A single character object, where the default is \code{"model area"}. The text will be added to some of the table captions to indicate
what the results apply to. Besides the default, one could use \code{"base model"}, \code{"sub-area model South of Point Conception."}, etc. Just know
that the text will be appended to \code{"for the"}, and thus, the default text
leads to \code{"for the model area."}. Another thing to note is that a full
stop is not needed but can be used because a full stop is appended to the
end of the caption if it does not already exist.}
}
\value{
Individual .rds tables
}
\description{
Take the output from \code{SS_output()} and create executive summary .rds files
as required by the current Terms of Reference for U.S. West Coast
groundfish assessments. Additionally, .rds files of historical catches,
time-series, and numbers-at-age are created. A CSV file with captions is
also created. This function is modified from \code{SSexecutivesummary()}
associated with the adoption of the {asar} template.
}
\author{
Chantel R. Wetzel, Kelli F. Johnson, Ian G. Taylor
}
