% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_outputconfig.R
\name{table_outputconfig}
\alias{table_outputconfig}
\title{Create a table to summarize the configuration of the SS3 model}
\usage{
table_outputconfig(output, dir = NULL)
}
\arguments{
\item{output}{A list from `r4ss::SS_output``}
}
\description{
Create a table to summarize the configuration of the SS3 model
}
\examples{
\dontrun{
  # Load the model output
  output <- r4ss::SS_output()
  # Create the table
  table_outputconfig(output)
}
}
\author{
Kelli F. Johnson, Ian G. Taylor
}
