% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parcounts.R
\name{table_parcounts}
\alias{table_parcounts}
\title{Summarize the estimated parameters based on an SS3 control file}
\usage{
table_parcounts(output, dir = output$inputs$dir)
}
\arguments{
\item{output}{Model output produced by \code{r4ss::SS_output()}}

\item{dir}{A directory which contains a control file, if different
from output$inputs$dir}

\item{ctlfile}{The filename for the control file}
}
\description{
Summarize the estimated parameters based on an SS3 control file
}
\author{
Ian G. Taylor
}
