% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parcounts.R
\name{table_parcounts}
\alias{table_parcounts}
\title{Summarize the estimated parameters based on an SS3 control file}
\usage{
table_parcounts(
  replist,
  inputs = NULL,
  dir = NULL,
  caption = "Estimated parameters in the model",
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{inputs}{An options list of inputs from the SS3 model created by \code{SS_read()}.
If NULL, then the function will run \code{SS_read(replist$inputs$dir)} to get the inputs.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Requires either
}
\examples{
\dontrun{
# Load the model output
output <- r4ss::SS_output()
# Create the table
table_parcounts <- table_parcounts(output)
# filter for types with at least one estimated parameter
table_parcounts$table |> dplyr::filter(Count > 0)
}
}
\seealso{
Other table functions: 
\code{\link{table_all}()},
\code{\link{table_biology}()},
\code{\link{table_compweight}()},
\code{\link{table_config}()},
\code{\link{table_exec_summary}()},
\code{\link{table_pars}()},
\code{\link{table_ts}()}
}
\author{
Ian G. Taylor
}
\concept{table functions}
