% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pars.R
\name{table_pars}
\alias{table_pars}
\title{Table of parameters}
\usage{
table_pars(
  output,
  rows = NULL,
  caption =
    "Parameter estimates, estimation phase, parameter bounds, estimation status, estimated standard deviation (SD), prior information [distribution(mean, SD)] used in the base model.",
  label = "table-pars-base"
)
}
\arguments{
\item{output}{A list from \code{\link[=SS_output]{SS_output()}}.}

\item{rows}{Which rows to include from the \code{parameters} table created by
\code{\link[=SS_output]{SS_output()}}. NULL will cause all rows to be included.}

\item{caption}{A character string for the caption.}

\item{label}{A character string with the label for the table.
No underscores allowed.}
}
\description{
Table of parameters
}
\seealso{
\code{\link[=table_pars_split]{table_pars_split()}}
}
\author{
Kelli F. Johnson
}
