% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pars.R
\name{table_pars}
\alias{table_pars}
\title{Table of parameters}
\usage{
table_pars(
  replist,
  rows = NULL,
  caption =
    "Parameter estimates, estimation phase, parameter bounds, estimation status, estimated standard deviation (SD), prior information [distribution(mean, SD)] used in the base model."
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{rows}{Which rows to include from the \code{parameters} table created by
\code{\link[=SS_output]{SS_output()}}. NULL will cause all rows to be included.}

\item{caption}{A character string for the caption.}
}
\description{
Table of parameters
}
\seealso{
Other table functions: 
\code{\link{table_compweight}()},
\code{\link{table_exec_summary}()},
\code{\link{table_outputconfig}()},
\code{\link{table_parcounts}()},
\code{\link{table_ts}()}
}
\author{
Kelli F. Johnson
}
\concept{table functions}
