% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ts.R
\name{table_ts}
\alias{table_ts}
\title{Table of estimated time series}
\usage{
table_ts(
  replist,
  dir = NULL,
  caption = "Time series of population estimates for the base model.",
  verbose = TRUE
)
}
\arguments{
\item{replist}{A list object created by \code{\link{SS_output}()}.}

\item{dir}{Directory where the .rda files will be written. The default value
is NULL where a table folder will be created where the Report.sso file is
located associated with \code{replist}.}

\item{caption}{A character string for the caption.}

\item{verbose}{A logical value specifying if output should be printed
to the screen.}
}
\description{
Adapted from Lingcod 2021:
https://github.com/pfmc-assessments/lingcod/blob/main/R/table_ts.R
}
\examples{
\dontrun{
table_ts(model)
}
}
\seealso{
Other table functions: 
\code{\link{table_compweight}()},
\code{\link{table_exec_summary}()},
\code{\link{table_outputconfig}()},
\code{\link{table_parcounts}()},
\code{\link{table_pars}()}
}
\author{
Kelli F. Johnson, Ian G. Taylor
}
\concept{table functions}
