% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update_r4ss_files}
\alias{update_r4ss_files}
\title{Updates r4ss files to newest versions on web.}
\usage{
update_r4ss_files(local = NULL, save = FALSE, revision = "newest",
  GitHub = TRUE, override = FALSE)
}
\arguments{
\item{local}{A local directory from which to source the files instead of
getting them from the web.}

\item{save}{If TRUE, then copy files from web to local directory, then
source from this same local directory}

\item{revision}{Either "newest" (the default), or an optional revision
number of the files to source. These numbers are found within the list of
changes to the r4ss code at \url{http://code.google.com/p/r4ss/source/list}.
If you're using an out-of-date version of SS, or some recent update to the
code isn't working for your model, an older revision may help. Otherwise, we
recommend using the newest revision.}

\item{GitHub}{Get files from GitHub (as opposed to Google Code)}

\item{override}{Override the message about how you should get code from github?}
}
\description{
Sources files containing R functions r4ss package from the google code
repository. These may often be newer than those available form CRAN mirrors.
It is probably wise to run this function every time you load the r4ss
library.
}
\examples{
\dontrun{
update_r4ss_files()
# getting file names from http://r4ss.googlecode.com/svn/trunk/
# most recent change: Today (6 hours ago)
# 64 files found in http://r4ss.googlecode.com/svn/trunk/
#   sourcing IOTCmove.R
#   sourcing RebuildPlot.R
#   sourcing SSFishGraph.R
#   sourcing SS_changepars.R
#   sourcing SS_fitbiasramp.R
#   sourcing SS_makedatlist.R
#   ...
#   sourcing stackpoly.R
#   sourcing update_r4ss_files.R
# update complete.

# copy files from web to local directory and then source them
update_r4ss_files(local='c:/SS/R/r4ss_files/',save=T)

# source files from a local directory (i.e. if no network available)
update_r4ss_files(local='c:/SS/R/r4ss_files/',save=F)

# update the updater function to get the new options:
source("http://r4ss.googlecode.com/svn/trunk/update_r4ss_files.R")

# get version 523 (for latest version, no "revision" input is needed)
update_r4ss_files(revision=523)
}
}
\author{
Ian Taylor
}
\keyword{file}

