% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_r4ss_files.R
\name{update_r4ss_files}
\alias{update_r4ss_files}
\title{Old function to updates r4ss files, deprecated in favor of devtools method}
\usage{
update_r4ss_files(...)
}
\arguments{
\item{...}{Any arguments that you would have passed to the function.}
}
\description{
Left in place only as a pointer to the new installation approach.
}
\author{
Ian Taylor
}

